crypto_bigint

Module int

Stack-allocated big signed integers.

Modules

add 🔒

Int addition operations.

bit_and 🔒

Int bitwise AND operations.

bit_not 🔒

Int bitwise NOT operations.

bit_or 🔒

Int bitwise OR operations.

bit_xor 🔒

Int bitwise XOR operations.

cmp 🔒

Int comparisons.

ct 🔒

Constant-time support: impls of Ct* traits and constant-time const fn operations.

div 🔒

Int division operations.

div_unsigned 🔒

Operations related to dividing an Int by a Uint.

encoding 🔒

Const-friendly decoding/encoding operations for Int.

from 🔒

From-like conversions for Int.

gcd 🔒

This module implements (a constant variant of) the Optimized Extended Binary GCD algorithm, which is described by Pornin in “Optimized Binary GCD for Modular Inversion”. Ref: https://eprint.iacr.org/2020/972.pdf

invert_mod 🔒

Operations related to computing the inverse of an Int modulo a Uint.

mod_symbol 🔒

Support for computing modular symbols.

mul 🔒

Int multiplication operations.

mul_unsigned 🔒

neg 🔒

Int negation-related operations.

rand 🔒

Random number generator support

resize 🔒

shl 🔒

Int bitwise left shift operations.

shr 🔒

Int bitwise right shift operations.

sign 🔒

sqrt 🔒

Int square root operations.

sub 🔒

Int subtraction operations.

types 🔒

Selection of Int types. todo: replace with macro implementation once serde is set up.

Structs

Int

Stack-allocated big signed integer. See Uint for unsigned integers.