crypto_bigint

Module limb

Big integers are represented as an array of smaller CPU word-size integers called “limbs”.

Modules

add 🔒

Limb addition

bit_and 🔒

Limb bit and operations.

bit_not 🔒

Limb bit not operations.

bit_or 🔒

Limb bit or operations.

bit_xor 🔒

Limb bit xor operations.

bits 🔒

cmp 🔒

Limb comparisons

ct 🔒

Constant-time support: impls of Ct* traits and constant-time const fn operations.

div 🔒

Limb division

encoding 🔒

Limb encoding

from 🔒

From-like conversions for Limb.

gcd 🔒

GCD support for Limb.

invert_mod 🔒

mul 🔒

Limb multiplication

neg 🔒

Limb negation

rand 🔒

Random number generator support

shl 🔒

Limb left bitshift

shr 🔒

Limb right bitshift

sqrt 🔒

Square root implementation for Limb.

sub 🔒

Limb subtraction

Structs

Limb

Big integers are represented as an array/vector of smaller CPU word-size integers called “limbs”.

Functions

nlimbs

Calculate the number of limbs required to represent the given number of bits.