crypto_bigint::limb

Struct Limb

#[repr(transparent)]
pub struct Limb(pub Word);

Big integers are represented as an array/vector of smaller CPU word-size integers called “limbs”.

The Limb type uses a 32-bit or 64-bit saturated representation, depending on the target. All bits of an inner Word are used to represent larger big integer types.

Tuple Fields
0: Word
Implementations

impl Limb

const fn bounded_div2k_mod_q( self, k: u32, k_upper_bound: u32, one_half_mod_q: Self, ) -> (Self, Self)

Compute self / 2^t mod q, returning the result, as well as the minimal factor f such that 2^t divides self + q·f.

Here, q := 2·one_half_mod_q + 1 is assumed odd and t := min(k, k_upper_bound).

Executes in variable time w.r.t. k_upper_bound only; executes in constant time w.r.t k.

impl Limb

pub const fn adc(self, rhs: Limb, carry: Limb) -> (Limb, Limb)

👎Deprecated since 0.7.0:
please use carrying_add instead

Computes self + rhs + carry, returning the result along with the new carry.

pub const fn carrying_add(self, rhs: Limb, carry: Limb) -> (Limb, Limb)

Computes self + rhs + carry, returning the result along with the new carry.

pub const fn overflowing_add(self, rhs: Limb) -> (Limb, Limb)

Computes self + rhs, returning the result along with the carry.

pub const fn saturating_add(&self, rhs: Self) -> Self

Perform saturating addition.

pub const fn wrapping_add(&self, rhs: Self) -> Self

Perform wrapping addition, discarding overflow.

impl Limb

pub const fn bitand(self, rhs: Self) -> Self

Calculates a & b.

impl Limb

pub const fn not(self) -> Self

Calculates !a.

impl Limb

pub const fn bitor(self, rhs: Self) -> Self

Calculates a | b.

impl Limb

pub const fn bitxor(self, rhs: Self) -> Self

Calculates a ^ b.

impl Limb

pub const fn bit(self, index: u32) -> Choice

Get the value of the bit at position index, as a truthy or falsy Choice. Returns the falsy value for indices out of range.

pub const fn bit_vartime(self, index: u32) -> bool

Returns true if the bit at position index is set, false for an unset bit or for indices out of range.

Remarks

This operation is variable time with respect to index only.

fn set_bit(self, index: u32, bit_value: Choice) -> Self

Sets the bit at index to 0 or 1 depending on the value of bit_value.

fn set_bit_vartime(self, index: u32, bit_value: bool) -> Self

Sets the bit at index to 0 or 1 depending on the value of bit_value, in variable-time with respect to index.

pub const fn bits(self) -> u32

Calculate the number of bits needed to represent this number.

pub const fn leading_zeros(self) -> u32

Calculate the number of leading zeros in the binary representation of this number.

pub const fn trailing_zeros(self) -> u32

Calculate the number of trailing zeros in the binary representation of this number.

pub const fn trailing_ones(self) -> u32

Calculate the number of trailing ones the binary representation of this number.

pub const fn restrict_bits(self, len: u32) -> Self

Clear bits at or above the given bit position.

impl Limb

pub fn is_odd(self) -> Choice

Is this limb an odd number?

pub fn cmp_vartime(self, other: Self) -> Ordering

Perform a comparison of the inner value in variable-time.

Note that the PartialOrd and Ord impls wrap constant-time comparisons using the subtle crate.

pub const fn eq_vartime(self, other: Self) -> bool

Performs an equality check in variable-time.

pub(crate) const fn is_nonzero(self) -> Choice

Returns the truthy value if self != 0 and the falsy value otherwise.

impl Limb

pub(crate) const fn select(a: Self, b: Self, c: Choice) -> Self

Return b if c is truthy, otherwise return a.

pub(crate) const fn ct_conditional_swap(a: &mut Self, b: &mut Self, c: Choice)

Swap the values of a and b if c is truthy, otherwise do nothing.

impl Limb

pub const fn div_rem(self, rhs: NonZero<Self>) -> (Limb, Limb)

Computes self / rhs, returning the quotient and remainder.

pub const fn div_rem_with_reciprocal(self, recip: &Reciprocal) -> (Limb, Limb)

Computes self / rhs where recip is a Reciprocal created from a non-zero Limb rhs. Returns the quotient and remainder.

pub const fn checked_div(self, rhs: Self) -> CtOption<Limb>

Computes the checked division self / rhs, returning the quotient if the divisor is non-zero, and CtOption::none() otherwise.

pub const fn checked_rem(self, rhs: Self) -> CtOption<Limb>

Computes the checked division self / rhs, returning the remainder if the divisor is non-zero, and CtOption::none() otherwise.

pub const fn div_exact(&self, rhs: NonZero<Limb>) -> CtOption<Self>

Exactly divides self by rhs, returning CtOption::none() if self is not divisible by rhs.

pub const fn div_exact_vartime(&self, rhs: NonZero<Limb>) -> CtOption<Self>

Exactly divides self by rhs, returning CtOption::none() if self is not divisible by rhs.

This is variable-time only with respect to rhs.

When used with a fixed rhs, this function is constant-time with respect to self.

impl Limb

pub const fn from_be_bytes(bytes: [u8; 8]) -> Self

Decode from big endian bytes.

pub const fn from_le_bytes(bytes: [u8; 8]) -> Self

Decode from little endian bytes.

pub const fn from_be_slice(bytes: &[u8]) -> Self

Decode limb from a big endian byte slice, which may be shorter than Limb::BYTES.

Panics

If the slice larger than Limb::BYTES.

pub const fn from_le_slice(bytes: &[u8]) -> Self

Decode limb from a little endian byte slice, which may be shorter than Limb::BYTES.

Panics

If the slice larger than Limb::BYTES.

pub const fn from_be_slice_truncated(bytes: &[u8], bits_precision: u32) -> Self

Decode limb from the provided big endian bytes, zero padding if necessary, and truncating to the least significant bytes in the event the given amount of data exceeds bits_precision.

Panics

If bits_precision > Self::BITS.

pub const fn from_le_slice_truncated(bytes: &[u8], bits_precision: u32) -> Self

Decode limb from the provided little endian bytes, zero padding if necessary, and truncating to the least significant bytes in the event the given amount of data exceeds bits_precision.

Panics

If bits_precision > Self::BITS.

pub const fn to_be_bytes(&self) -> [u8; 8]

Encode as big endian bytes.

pub const fn to_le_bytes(&self) -> [u8; 8]

Encode as little endian bytes.

pub(crate) const fn mask_to_precision(&mut self, bits_precision: u32)

Mask to the given number of bits precision.

impl Limb

pub const fn from_u8(n: u8) -> Self

Create a Limb from a u8 integer (const-friendly)

pub const fn from_u16(n: u16) -> Self

Create a Limb from a u16 integer (const-friendly)

pub const fn from_u32(n: u32) -> Self

Create a Limb from a u32 integer (const-friendly)

pub const fn from_u64(n: u64) -> Self

Create a Limb from a u64 integer (const-friendly)

impl Limb

pub const fn mac(self, b: Limb, c: Limb, carry: Limb) -> (Limb, Limb)

👎Deprecated since 0.7.0:
please use carrying_mul_add instead (ordering of arguments changes)

Computes self + (b * c) + carry, returning the result along with the new carry.

pub const fn carrying_mul_add( self, rhs: Limb, addend: Limb, carry: Limb, ) -> (Limb, Limb)

Computes (self * rhs) + addend + carry, returning the result along with the new carry.

pub const fn saturating_mul(self, rhs: Self) -> Self

Perform saturating multiplication.

pub const fn wrapping_mul(self, rhs: Self) -> Self

Perform wrapping multiplication, discarding overflow.

pub(crate) const fn widening_mul(self, rhs: Self) -> (Self, Self)

Compute “wide” multiplication, with a product twice the size of the input.

pub(crate) const fn widening_square(self) -> (Self, Self)

Compute “wide” squaring, with a product twice the size of the input.

pub(crate) const fn rem_wide(lo_hi: (Self, Self), p: NonZero<Self>) -> Self

Compute “wide” squaring, with a product twice the size of the input.

impl Limb

pub const fn wrapping_neg(self) -> Self

Perform wrapping negation.

impl Limb

pub const fn shl(self, shift: u32) -> Self

Computes self << shift.

Panics

if shift overflows Limb::BITS.

pub const fn overflowing_shl(self, shift: u32) -> CtOption<Self>

Computes self << shift, returning CtOption::none() if the shift exceeds the capacity.

pub const fn overflowing_shl_vartime(self, shift: u32) -> Option<Self>

Computes self << shift, returning None if the shift exceeds the capacity.

This method is variable time in shift only.

pub const fn unbounded_shl(self, shift: u32) -> Self

Computes self << shift, returning Limb::ZERO if the shift exceeds the capacity.

pub const fn unbounded_shl_vartime(self, shift: u32) -> Self

Computes self << shift, returning Limb::ZERO if the shift exceeds the capacity.

This method is variable time in shift only.

pub const fn wrapping_shl(self, shift: u32) -> Self

Computes self << shift in a panic-free manner, masking off bits of shift which would cause the shift to exceed the type’s width.

impl Limb

pub const fn shr(self, shift: u32) -> Self

Computes self >> shift.

Panics

if shift overflows Limb::BITS.

pub(crate) const fn shr1(self) -> (Self, Self)

Computes self >> 1 and return the result and the carry (0 or 1 << HI_BIT).

pub const fn overflowing_shr(self, shift: u32) -> CtOption<Self>

Computes self >> shift, returning CtOption::none() if the shift exceeds the capacity.

pub const fn overflowing_shr_vartime(self, shift: u32) -> Option<Self>

Computes self >> shift, returning None if the shift exceeds the capacity.

This method is variable time in shift only.

pub const fn unbounded_shr(self, shift: u32) -> Self

Computes self >> shift, returning Limb::ZERO if the shift exceeds the capacity.

pub const fn unbounded_shr_vartime(self, shift: u32) -> Self

Computes self >> shift, returning Limb::ZERO if the shift exceeds the capacity.

This method is variable time in shift only.

pub const fn wrapping_shr(self, shift: u32) -> Self

Computes self >> shift in a panic-free manner, masking off bits of shift which would cause the shift to exceed the type’s width.

impl Limb

pub const fn sbb(self, rhs: Limb, borrow: Limb) -> (Limb, Limb)

👎Deprecated since 0.7.0:
please use borrowing_sub instead

Computes self - (rhs + borrow), returning the result along with the new borrow.

pub const fn borrowing_sub(self, rhs: Limb, borrow: Limb) -> (Limb, Limb)

Computes self - (rhs + borrow), returning the result along with the new borrow.

pub const fn saturating_sub(&self, rhs: Self) -> Self

Perform saturating subtraction.

pub const fn wrapping_sub(&self, rhs: Self) -> Self

Perform wrapping subtraction, discarding underflow and wrapping around the boundary of the type.

impl Limb

pub const ZERO: Self

The value 0.

pub const ONE: Self

The value 1.

pub const MAX: Self

Maximum value this Limb can express.

pub(crate) const HI_BIT: u32

Highest bit in a Limb.

pub const BITS: u32 = 64

Size of the inner integer in bits.

pub const BYTES: usize = 8

Size of the inner integer in bytes.

pub const LOG2_BITS: u32

floor(log2(Self::BITS)).

pub const fn is_zero(&self) -> Choice

Is this limb equal to Limb::ZERO?

pub const fn to_nz(self) -> CtOption<NonZero<Self>>

Convert to a NonZero<Limb>.

Returns some if the original value is non-zero, and none otherwise.

pub(crate) const fn to_nz_or_one(self) -> (NonZero<Self>, Choice)

Convert to a NonZero<Limb>, defaulting to Self::ONE.

Returns a pair consisting of a NonZero<Limb>, and a Choice indicating whether the original value was non-zero (and preserved).

pub const fn lsb_to_choice(self) -> Choice

Convert the least significant bit of this Limb to a Choice.

pub const fn array_as_words<const N: usize>(array: &[Self; N]) -> &[Word; N]

Convert a shared reference to an array of Limbs into a shared reference to their inner Words for each limb.

pub const fn array_as_mut_words<const N: usize>( array: &mut [Self; N], ) -> &mut [Word; N]

Convert a mutable reference to an array of Limbs into a mutable reference to their inner Words for each limb.

pub const fn slice_as_words(slice: &[Self]) -> &[Word]

Convert a shared reference to an array of Limbs into a shared reference to their inner Words for each limb.

pub const fn slice_as_mut_words(slice: &mut [Self]) -> &mut [Word]

Convert a mutable reference to an array of Limbs into a mutable reference to their inner Words for each limb.