crypto_bigint
Macro impl_modulus

macro_rules! impl_modulus {
    ($name:ident, $uint_type:ty, $value:expr) => { ... };
}

Implements a modulus with the given name, type, and value, in that specific order. Please use crypto_bigint::traits::Encoding to make this work. For example, impl_modulus!(MyModulus, U256, "73eda753299d7d483339d80809a1d80553bda402fffe5bfeffffffff00000001"); implements a 256-bit modulus named MyModulus. The modulus must be odd, or this will panic.