crypto_bigint::modular::bingcd

Module matrix

Structs

DividedMatrix 🔒

A matrix whose elements still need to be divided by 2^k.

DividedPatternMatrix 🔒

Variation on PatternMatrix, where the contents of the matrix need to be divided by 2^k. The internal state represents the matrix

IntMatrix 🔒

2x2-Matrix with integers for elements.

PatternMatrix 🔒

2x2-Matrix where either the diagonal or off-diagonal elements are negative.

Traits

Unit

Type Aliases

DividedIntMatrix 🔒

ExtraLimbInt 🔒

[Int] with an extra limb.

Vector 🔒