crypto_bigint::modular

Module boxed_monty_form

Implements heap-allocated BoxedMontyForms, supporting modular arithmetic with a modulus set at runtime.

Modules

add 🔒

Additions between boxed Montgoemery form integers.

ct 🔒

Constant-time support: impls of Ct* traits.

from 🔒

From-like conversions for BoxedMontyForm and BoxedMontyParams.

invert 🔒

Multiplicative inverses of boxed integers in Montgomery form.

lincomb 🔒

Linear combinations of integers in Montgomery form with a modulus set at runtime.

mul 🔒

Multiplication between boxed integers in Montgomery form (i.e. Montgomery multiplication).

neg 🔒

Negations of boxed integers in Montgomery form.

pow 🔒

Modular exponentiation support for BoxedMontyForm.

sub 🔒

Subtractions between boxed integers in Montgomery form.

Structs

BoxedMontyForm

An integer in Montgomery form represented using heap-allocated limbs.

Functions

convert_to_montgomery 🔒

Convert the given integer into the Montgomery domain.