Struct crypto_bigint::modular::constant_mod::Residue

pub struct Residue<MOD, const LIMBS: usize>
where
    MOD: ResidueParams<LIMBS>,
{ /* private fields */ }

A residue mod MOD, represented using LIMBS limbs. The modulus of this residue is constant, so it cannot be set at runtime. Internally, the value is stored in Montgomery form (multiplied by MOD::R) until it is retrieved.

Implementations

impl<MOD: ResidueParams<LIMBS>, const LIMBS: usize> Residue<MOD, LIMBS>

pub const fn add(&self, rhs: &Residue<MOD, LIMBS>) -> Self

Adds rhs.

impl<MOD: ResidueParams<LIMBS>, const LIMBS: usize> Residue<MOD, LIMBS>

pub const fn invert(&self) -> (Self, CtChoice)

Computes the residue self^-1 representing the multiplicative inverse of self. I.e. self * self^-1 = 1. If the number was invertible, the second element of the tuple is the truthy value, otherwise it is the falsy value (in which case the first element’s value is unspecified).

impl<MOD: ResidueParams<LIMBS>, const LIMBS: usize> Residue<MOD, LIMBS>

pub const fn mul(&self, rhs: &Self) -> Self

Multiplies by rhs.

pub const fn square(&self) -> Self

Computes the (reduced) square of a residue.

impl<MOD: ResidueParams<LIMBS>, const LIMBS: usize> Residue<MOD, LIMBS>

pub const fn neg(&self) -> Self

Negates the number.

impl<MOD: ResidueParams<LIMBS>, const LIMBS: usize> Residue<MOD, LIMBS>

pub const fn pow<const RHS_LIMBS: usize>( &self, exponent: &Uint<RHS_LIMBS>, ) -> Residue<MOD, LIMBS>

Raises to the exponent power.

pub const fn pow_bounded_exp<const RHS_LIMBS: usize>( &self, exponent: &Uint<RHS_LIMBS>, exponent_bits: usize, ) -> Residue<MOD, LIMBS>

Raises to the exponent power, with exponent_bits representing the number of (least significant) bits to take into account for the exponent.

NOTE: exponent_bits may be leaked in the time pattern.

impl<MOD: ResidueParams<LIMBS>, const LIMBS: usize> Residue<MOD, LIMBS>

pub const fn sub(&self, rhs: &Self) -> Self

Subtracts rhs.

impl<MOD: ResidueParams<LIMBS>, const LIMBS: usize> Residue<MOD, LIMBS>

pub const ZERO: Self = _

The representation of 0 mod MOD.

pub const ONE: Self = _

The representation of 1 mod MOD.

pub const fn new(integer: &Uint<LIMBS>) -> Self

Instantiates a new Residue that represents this integer mod MOD. If the modulus represented by MOD is not odd, this function will panic; use new_checked if you want to be able to detect an invalid modulus.

pub fn new_checked(integer: &Uint<LIMBS>) -> CtOption<Self>

Instantiates a new Residue that represents this integer mod MOD if the modulus is odd. Returns a CtOption that is None if the provided modulus is not odd; this is a safer version of new, which can panic.

pub const fn retrieve(&self) -> Uint<LIMBS>

Retrieves the integer currently encoded in this Residue, guaranteed to be reduced.

pub const fn as_montgomery(&self) -> &Uint<LIMBS>

Access the Residue value in Montgomery form.

pub fn as_montgomery_mut(&mut self) -> &mut Uint<LIMBS>

Mutably access the Residue value in Montgomery form.

pub const fn from_montgomery(integer: Uint<LIMBS>) -> Self

Create a Residue from a value in Montgomery form.

pub const fn to_montgomery(&self) -> Uint<LIMBS>

Extract the value from the Residue in Montgomery form.

pub fn div_by_2(&self) -> Self

Performs the modular division by 2, that is for given x returns y such that y * 2 = x mod p. This means:

if x is even, returns x / 2,

if x is odd, returns (x + p) / 2 (since the modulus p in Montgomery form is always odd, this divides entirely).