crypto_bigint::modular::monty_params

Struct MontyParams

pub struct MontyParams<U: Unsigned> {
    pub(super) modulus: Odd<U>,
    pub(super) one: U,
    pub(super) r2: U,
    pub(super) mod_inv: U64,
    pub(super) mod_leading_zeros: u32,
}

Parameters to efficiently go to/from the Montgomery form for an odd modulus provided at runtime.

This version is generic over the underlying unsigned integer type.

Fields
modulus: Odd<U>

The constant modulus.

one: U

1 in Montgomery form (a.k.a. R).

r2: U

R^2 mod modulus, used to move into Montgomery form.

mod_inv: U64

The lowest limbs of MODULUS^-1 mod 2**64.

This value is used in Montgomery reduction and modular inversion.

mod_leading_zeros: u32

Leading zeros in the modulus, used to choose optimized algorithms.

Implementations

impl<const LIMBS: usize> MontyParams<Uint<LIMBS>>

const fn inverter(&self) -> SafeGcdInverter<LIMBS>

Create a modular inverter for the modulus of these params.

impl<U: Unsigned> MontyParams<U>

pub const fn modulus(&self) -> &Odd<U>

Returns the modulus which was used to initialize these parameters.

pub const fn one(&self) -> &U

1 in Montgomery form (a.k.a. R).

pub const fn r2(&self) -> &U

R^2 mod modulus, used to move into Montgomery form.

pub const fn mod_inv(&self) -> &U64

Returns the lowest limbs of MODULUS^-1 mod 2**64.

This value is used in Montgomery reduction and modular inversion.

pub const fn mod_neg_inv(&self) -> Limb

Returns wrapping negation of first limb of mod_inv.

pub const fn mod_leading_zeros(&self) -> u32

Returns leading zeros in the modulus, used to choose optimized algorithms.

pub(crate) fn debug_struct(&self, debug: DebugStruct<'_, '_>) -> Result

Core implementation of the debug impl which lets us customize it for various types/type aliases.

impl<const LIMBS: usize> MontyParams<Uint<LIMBS>>

pub const fn new(modulus: Odd<Uint<LIMBS>>) -> Self

Instantiates a new set of MontyParams representing the given odd modulus.

impl<const LIMBS: usize> MontyParams<Uint<LIMBS>>

pub const fn new_vartime(modulus: Odd<Uint<LIMBS>>) -> Self

Instantiates a new set of MontyParams representing the given odd modulus.