crypto_bigint::modular

Module pow

Constants

BITS_PER_WINDOW 🔒

WINDOW 🔒

WINDOW_COUNT 🔒

WINDOW_MASK 🔒

Functions

compute_powers 🔒

multi_exponentiate_montgomery_form_array

multi_exponentiate_montgomery_form_internal 🔒

multi_exponentiate_montgomery_form_slice

Performs modular multi-exponentiation using Montgomery’s ladder. exponent_bits represents the number of bits to take into account for the exponent.

pow_init 🔒

pow_montgomery_form

Performs modular exponentiation using Montgomery’s ladder. exponent_bits represents the number of bits to take into account for the exponent.

pow_montgomery_form_amm

Performs modular exponentiation using “Almost Montgomery Multiplication”.