crypto_bigint::modular::safegcd

Module boxed

Implementation of Bernstein-Yang modular inversion and GCD algorithm (a.k.a. safegcd) as described in: https://eprint.iacr.org/2019/266.

See parent module for more information.

Structs

BoxedSafeGcdInverter 🔒

Modular multiplicative inverter based on the Bernstein-Yang method.

SignedBoxedInt 🔒

A Uint which carries a separate sign in order to maintain the same range.

Functions

gcd

Calculate the greatest common denominator of f and g.

gcd_nz

Calculate the greatest common denominator of nonzero f, and g.

gcd_odd

Calculate the greatest common denominator of odd f, and g.

invert_odd_mod

invert_odd_mod_precomp 🔒

Calculate the multiplicative inverse of a modulo m.

update_de 🔒

update_fg 🔒