crypto_bigint::modular

Struct FixedMontyForm

pub struct FixedMontyForm<const LIMBS: usize> { /* private fields */ }

An integer in Montgomery form represented using LIMBS limbs. The odd modulus is set at runtime.

Implementations

impl<const LIMBS: usize> FixedMontyForm<LIMBS>

pub const fn add(&self, rhs: &Self) -> Self

Adds rhs.

pub const fn double(&self) -> Self

Double self.

impl<const LIMBS: usize> FixedMontyForm<LIMBS>

pub const fn inv(&self) -> CtOption<Self>

👎Deprecated since 0.7.0:
please use invert instead

Computes self^-1 representing the multiplicative inverse of self. i.e. self * self^-1 = 1.

If the number was invertible, the second element of the tuple is the truthy value, otherwise it is the falsy value (in which case the first element’s value is unspecified).

pub const fn invert(&self) -> CtOption<Self>

Computes self^-1 representing the multiplicative inverse of self. i.e. self * self^-1 = 1.

If the number was invertible, the second element of the tuple is the truthy value, otherwise it is the falsy value (in which case the first element’s value is unspecified).

pub const fn inv_vartime(&self) -> CtOption<Self>

👎Deprecated since 0.7.0:
please use invert_vartime instead

Computes self^-1 representing the multiplicative inverse of self. i.e. self * self^-1 = 1.

If the number was invertible, the second element of the tuple is the truthy value, otherwise it is the falsy value (in which case the first element’s value is unspecified).

This version is variable-time with respect to the value of self, but constant-time with respect to self’s params.

pub const fn invert_vartime(&self) -> CtOption<Self>

Computes self^-1 representing the multiplicative inverse of self. i.e. self * self^-1 = 1.

If the number was invertible, the second element of the tuple is the truthy value, otherwise it is the falsy value (in which case the first element’s value is unspecified).

This version is variable-time with respect to the value of self, but constant-time with respect to self’s params.

impl<const LIMBS: usize> FixedMontyForm<LIMBS>

pub const fn lincomb_vartime(products: &[(&Self, &Self)]) -> Self

Calculate the sum of products of pairs (a, b) in products.

This method is variable time only with the value of the modulus. For a modulus with leading zeros, this method is more efficient than a naive sum of products.

This method will panic if products is empty. All terms must be associated with equivalent MontyParams.

Panics

if products is empty.

impl<const LIMBS: usize> FixedMontyForm<LIMBS>

pub const fn jacobi_symbol(&self) -> JacobiSymbol

Compute the Jacobi symbol (self|modulus).

For a prime modulus, this corresponds to the Legendre symbol and indicates whether self is quadratic residue.

pub const fn jacobi_symbol_vartime(&self) -> JacobiSymbol

Compute the Jacobi symbol (self|modulus).

For a prime modulus, this corresponds to the Legendre symbol and indicates whether self is quadratic residue.

This method is variable-time with respect to the value of self.

impl<const LIMBS: usize> FixedMontyForm<LIMBS>

pub const fn mul(&self, rhs: &Self) -> Self

Multiplies by rhs.

pub const fn square(&self) -> Self

Computes the (reduced) square.

pub const fn square_repeat_vartime(&self, n: u32) -> Self

Computes the reduced product self^2n.

This method is variable time in n only.

impl<const LIMBS: usize> FixedMontyForm<LIMBS>

pub const fn neg(&self) -> Self

Negates the number.

impl<const LIMBS: usize> FixedMontyForm<LIMBS>

pub const fn pow<const RHS_LIMBS: usize>( &self, exponent: &Uint<RHS_LIMBS>, ) -> Self

Raises to the exponent power.

pub const fn pow_bounded_exp<const RHS_LIMBS: usize>( &self, exponent: &Uint<RHS_LIMBS>, exponent_bits: u32, ) -> Self

Raises to the exponent power, with exponent_bits representing the number of (least significant) bits to take into account for the exponent.

NOTE: exponent_bits may be leaked in the time pattern.

pub const fn pow_vartime<const RHS_LIMBS: usize>( &self, exponent: &Uint<RHS_LIMBS>, ) -> Self

Raises to the exponent power.

This method is variable time in exponent.

pub fn pow_amm(&self, exponent: &Uint<LIMBS>) -> Self

Raises to the exponent power using Almost Montgomery Multiplication (AMM).

pub fn pow_amm_bounded_exp( &self, exponent: &Uint<LIMBS>, exponent_bits: u32, ) -> Self

Raises to the exponent power using Almost Montgomery Multiplication (AMM) with exponent_bits representing the number of (least significant) bits to take into account for the exponent.

NOTE: exponent_bits may be leaked in the time pattern.

impl<const LIMBS: usize> FixedMontyForm<LIMBS>

pub const fn sub(&self, rhs: &Self) -> Self

Subtracts rhs.

impl<const LIMBS: usize> FixedMontyForm<LIMBS>

pub const fn new( integer: &Uint<LIMBS>, params: &FixedMontyParams<LIMBS>, ) -> Self

Instantiates a new MontyForm that represents this integer mod MOD.

pub const fn retrieve(&self) -> Uint<LIMBS>

Retrieves the integer currently encoded in this MontyForm, guaranteed to be reduced.

pub const fn zero(params: &FixedMontyParams<LIMBS>) -> Self

Instantiates a new MontyForm that represents zero.

pub const fn one(params: &FixedMontyParams<LIMBS>) -> Self

Instantiates a new MontyForm that represents 1.

pub const fn params(&self) -> &FixedMontyParams<LIMBS>

Returns the parameter struct used to initialize this object.

pub const fn as_montgomery(&self) -> &Uint<LIMBS>

Access the MontyForm value in Montgomery form.

pub fn as_montgomery_mut(&mut self) -> &mut Uint<LIMBS>

Mutably access the MontyForm value in Montgomery form.

pub const fn from_montgomery( integer: Uint<LIMBS>, params: &FixedMontyParams<LIMBS>, ) -> Self

Create a MontyForm from a value in Montgomery form.

pub const fn to_montgomery(&self) -> Uint<LIMBS>

Extract the value from the MontyForm in Montgomery form.

pub const fn div_by_2(&self) -> Self

Performs division by 2, that is returns x such that x + x = self.