crypto_bigint::modular

Trait ConstMontyParams

pub trait ConstMontyParams<const LIMBS: usize>:
    Copy
    + Debug
    + Default
    + Eq
    + Send
    + Sync
    + 'static {
    const LIMBS: usize;
    const PARAMS: FixedMontyParams<LIMBS>;
}

Trait representing a modulus and its associated constants for converting in and out of Montgomery form.

To define a type which impls this trait, use the const_monty_params! macro.

Required Associated Constants

const LIMBS: usize

Number of limbs required to encode the Montgomery form

const PARAMS: FixedMontyParams<LIMBS>

Montgomery parameters constant.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.