crypto_bigint::odd

Struct Odd

#[repr(transparent)]
pub struct Odd<T: ?Sized>(T);

Wrapper type for odd integers.

These are frequently used in cryptography, e.g. as a modulus.

Tuple Fields
0: T
Implementations

impl<const LIMBS: usize> Odd<Uint<LIMBS>>

const fn half_mod(q: &Self) -> Self

Compute 1/2 mod q.

const fn mul_add_div2k( &self, b: Limb, addend: &Uint<LIMBS>, k: u32, ) -> Uint<LIMBS>

Compute ((self * b) + addend) / 2^k

impl<const LIMBS: usize> Odd<Uint<LIMBS>>

pub(crate) const MINIMAL_BINGCD_ITERATIONS: u32

The minimal number of binary GCD iterations required to guarantee successful completion.

pub(crate) const fn classic_bingcd(&self, rhs: &Uint<LIMBS>) -> Self

Computes gcd(self, rhs), leveraging (a constant time implementation of) the classic Binary GCD algorithm.

Note: this algorithm is efficient for Uints with relatively few LIMBS.

Ref: Pornin, Optimized Binary GCD for Modular Inversion, Algorithm 1. https://eprint.iacr.org/2020/972.pdf

pub(crate) const fn classic_bingcd_(&self, rhs: &Uint<LIMBS>) -> (Self, Word)

Computes gcd(self, rhs), leveraging (a constant time implementation of) the classic Binary GCD algorithm.

Note: this algorithm is efficient for Uints with relatively few LIMBS.

Ref: Pornin, Optimized Binary GCD for Modular Inversion, Algorithm 1. https://eprint.iacr.org/2020/972.pdf

This method returns a pair consisting of the GCD and the sign of the Jacobi symbol, 0 for positive and 1 for negative.

pub(crate) const fn classic_bingcd_vartime(&self, rhs: &Uint<LIMBS>) -> Self

Variable time equivalent of Self::classic_bingcd.

pub(crate) const fn classic_bingcd_vartime_( &self, rhs: &Uint<LIMBS>, ) -> (Self, Word)

Variable time equivalent of Self::classic_bingcd_.

pub(crate) const fn optimized_bingcd(&self, rhs: &Uint<LIMBS>) -> Self

Computes gcd(self, rhs), leveraging the optimized Binary GCD algorithm.

Note: this algorithm becomes more efficient than the classical algorithm for Uints with relatively many LIMBS. A best-effort threshold is presented in Self::bingcd.

Note: the full algorithm has an additional parameter; this function selects the best-effort value for this parameter. You might be able to further tune your performance by calling the Self::optimized_bingcd_ function directly.

Ref: Pornin, Optimized Binary GCD for Modular Inversion, Algorithm 2. https://eprint.iacr.org/2020/972.pdf

pub(crate) const fn optimized_bingcd_<const K: u32, const LIMBS_K: usize, const LIMBS_2K: usize>( &self, rhs: &Uint<LIMBS>, batch_max: u32, ) -> (Self, Word)

Computes gcd(self, rhs), leveraging the optimized Binary GCD algorithm.

Ref: Pornin, Optimized Binary GCD for Modular Inversion, Algorithm 2. https://eprint.iacr.org/2020/972.pdf

In summary, the optimized algorithm does not operate on self and rhs directly, but instead of condensed summaries that fit in few registers. Based on these summaries, an update matrix is constructed by which self and rhs are updated in larger steps.

This function is generic over the following three values:

K: the number of bits used when summarizing self and rhs for the inner loop. The K top bits and K least significant bits are selected. It is recommended to keep K close to a (multiple of) the number of bits that fit in a single register.

LIMBS_K: should be chosen as the minimum number s.t. Uint::<LIMBS>::BITS ≥ K,

LIMBS_2K: should be chosen as the minimum number s.t. Uint::<LIMBS>::BITS ≥ 2K.

This method returns a pair consisting of the GCD and the sign of the Jacobi symbol, 0 for positive and 1 for negative.

pub(crate) const fn optimized_bingcd_vartime(&self, rhs: &Uint<LIMBS>) -> Self

Variable time equivalent of Self::optimized_bingcd.

pub(crate) const fn optimized_bingcd_vartime_<const K: u32, const LIMBS_K: usize, const LIMBS_2K: usize>( &self, rhs: &Uint<LIMBS>, batch_max: u32, ) -> (Self, Word)

Variable time equivalent of Self::optimized_bingcd_.

impl<const LIMBS: usize> Odd<Uint<LIMBS>>

const MIN_BINXGCD_ITERATIONS: u32

The minimal number of binary GCD iterations required to guarantee successful completion.

pub(crate) const fn binxgcd_nz( &self, rhs: &NonZeroUint<LIMBS>, ) -> RawXgcdOutput<LIMBS, PatternMatrix<LIMBS>>

Given (self, rhs), computes (g, x, y) s.t. self * x + rhs * y = g = gcd(self, rhs), leveraging the Binary Extended GCD algorithm.

pub(crate) const fn binxgcd_odd( &self, rhs: &Self, ) -> RawXgcdOutput<LIMBS, PatternMatrix<LIMBS>>

Execute the classic Extended GCD algorithm.

Given (self, rhs), computes (g, x, y) s.t. self * x + rhs * y = g = gcd(self, rhs).

pub(crate) const fn classic_binxgcd( &self, rhs: &Self, ) -> RawXgcdOutput<LIMBS, DividedPatternMatrix<LIMBS>>

Execute the classic Binary Extended GCD algorithm.

Given (self, rhs), computes (g, x, y) s.t. self * x + rhs * y = g = gcd(self, rhs).

Ref: Pornin, Optimized Binary GCD for Modular Inversion, Algorithm 1. https://eprint.iacr.org/2020/972.pdf.

pub(crate) const fn optimized_binxgcd( &self, rhs: &Self, ) -> RawXgcdOutput<LIMBS, DividedPatternMatrix<LIMBS>>

Given (self, rhs), computes (g, x, y) s.t. self * x + rhs * y = g = gcd(self, rhs), leveraging the Binary Extended GCD algorithm.

Warning: self and rhs must be contained in an U128 or larger.

Note: this algorithm becomes more efficient than the classical algorithm for Uints with relatively many LIMBS. A best-effort threshold is presented in [Self::binxgcd_].

Note: the full algorithm has an additional parameter; this function selects the best-effort value for this parameter. You might be able to further tune your performance by calling the Self::optimized_bingcd_ function directly.

Ref: Pornin, Optimized Binary GCD for Modular Inversion, Algorithm 2. https://eprint.iacr.org/2020/972.pdf.

pub(crate) const fn optimized_binxgcd_<const K: u32, const LIMBS_K: usize, const LIMBS_2K: usize>( &self, rhs: &Self, ) -> RawXgcdOutput<LIMBS, DividedPatternMatrix<LIMBS>>

Given (self, rhs), computes (g, x, y), s.t. self * x + rhs * y = g = gcd(self, rhs), leveraging the optimized Binary Extended GCD algorithm.

Ref: Pornin, Optimized Binary GCD for Modular Inversion, Algorithm 2. https://eprint.iacr.org/2020/972.pdf

In summary, the optimized algorithm does not operate on self and rhs directly, but instead of condensed summaries that fit in few registers. Based on these summaries, an update matrix is constructed by which self and rhs are updated in larger steps.

This function is generic over the following three values:

K: the number of bits used when summarizing self and rhs for the inner loop. The K+1 top bits and K-1 least significant bits are selected. It is recommended to keep K close to a (multiple of) the number of bits that fit in a single register.

LIMBS_K: should be chosen as the minimum number s.t. Uint::<LIMBS>::BITS ≥ K,

LIMBS_2K: should be chosen as the minimum number s.t. Uint::<LIMBS>::BITS ≥ 2K.

pub(super) const fn partial_binxgcd<const UPDATE_LIMBS: usize>( &self, rhs: &Uint<LIMBS>, iterations: u32, halt_at_zero: Choice, ) -> (Self, Uint<LIMBS>, DividedPatternMatrix<UPDATE_LIMBS>, Word)

Executes the optimized Binary GCD inner loop.

Ref: Pornin, Optimized Binary GCD for Modular Inversion, Algorithm 2. https://eprint.iacr.org/2020/972.pdf.

The function outputs the reduced values (a, b) for the input values (self, rhs) as well as the matrix that yields the former two when multiplied with the latter two.

Note: this implementation deviates slightly from the paper, in that it can be instructed to “run in place” (i.e., execute iterations that do nothing) once a becomes zero. This is done by passing a truthy halt_at_zero.

The function executes in time variable in iterations.

impl<const LIMBS: usize> Odd<Int<LIMBS>>

pub const fn gcd_unsigned(&self, rhs: &Uint<LIMBS>) -> Odd<Uint<LIMBS>>

Compute the greatest common divisor of self and rhs.

pub const fn gcd_unsigned_vartime(&self, rhs: &Uint<LIMBS>) -> OddUint<LIMBS>

Compute the greatest common divisor of self and rhs.

Executes in variable time w.r.t. all input parameters.

pub const fn xgcd( &self, rhs: &NonZero<Int<LIMBS>>, ) -> XgcdOutput<LIMBS, OddUint<LIMBS>>

Execute the Extended GCD algorithm.

Given (self, rhs), computes (g, x, y) s.t. self * x + rhs * y = g = gcd(self, rhs).

impl Odd<Limb>

pub(crate) const fn multiplicative_inverse(self) -> Limb

Returns the multiplicative inverse of the argument modulo 2^N, where 2^N is the capacity of a Limb.

impl<T> Odd<T>

pub fn new(n: T) -> CtOption<Self>
where T: Integer,

Create a new odd integer.

pub(crate) const fn new_unchecked(n: T) -> Odd<T>

Create a new Odd<T> without first checking that the contained value is odd.

Use with care! This method bypasses invariant checks.

Warning: Panics

We don’t explicitly flag this function as unsafe because it doesn’t have a memory safety impact, however functions called with Odd arguments assume this value is odd and may panic if given an even value.

pub fn get(self) -> T

Returns the inner value.

pub const fn get_copy(self) -> T
where T: Copy,

Returns a copy of the inner value for Copy types.

This allows the function to be const fn, since Copy is implicitly !Drop, which avoids problems around const fn destructors.

pub fn into_nz(self) -> NonZero<T>

All odd integers are definitionally non-zero, so we can convert into the equivalent NonZero type.

impl<T: ?Sized> Odd<T>

pub const fn as_ref(&self) -> &T

Provides access to the contents of Odd in a const context.

pub const fn as_nz_ref(&self) -> &NonZero<T>

All odd integers are definitionally non-zero, so we can also obtain a reference to the equivalent NonZero type.

pub(crate) const fn new_ref_unchecked(refval: &T) -> &Odd<T>

Cast a reference to Odd without first checking that the referenced value is odd.

Use with care! This method bypasses invariant checks.

Warning: Panics

We don’t explicitly flag this function as unsafe because it doesn’t have a memory safety impact, however functions called with Odd arguments assume this value is odd and may panic if given an even value.

impl<T> Odd<T>
where T: Bounded + ?Sized,

pub const BITS: u32 = T::BITS

Total size of the represented integer in bits.

pub const BYTES: usize = T::BYTES

Total size of the represented integer in bytes.

impl<const LIMBS: usize> Odd<Uint<LIMBS>>

pub const fn from_be_hex(hex: &str) -> Self

Create a new OddUint from the provided big endian hex string.

Panics

if the hex is malformed or not zero-padded accordingly for the size.

if the value is even.

pub const fn from_le_hex(hex: &str) -> Self

Create a new Odd<Uint<LIMBS>> from the provided little endian hex string.

Panics

if the hex is malformed or not zero-padded accordingly for the size.

if the value is even.

pub const fn as_uint_ref(&self) -> &OddUintRef

Borrow this OddUint as a &OddUintRef.

pub const fn resize<const T: usize>(&self) -> Odd<Uint<T>>

Construct an Odd<Uint<T>> from the unsigned integer value, truncating the upper bits if the value is too large to be represented.

impl<const LIMBS: usize> Odd<Int<LIMBS>>

pub const fn abs_sign(&self) -> (Odd<Uint<LIMBS>>, Choice)

The sign and magnitude of this Odd<Int<{LIMBS}>>.

pub const fn abs(&self) -> Odd<Uint<LIMBS>>

The magnitude of this Odd<Int<{LIMBS}>>.

impl Odd<UintRef>

pub const fn to_uint_resize<const T: usize>(&self) -> Odd<Uint<T>>

Construct an Odd<Uint<T>> from the unsigned integer value, truncating the upper bits if the value is too large to be represented.

impl Odd<BoxedUint>

pub const fn as_uint_ref(&self) -> &OddUintRef

Borrow this OddBoxedUint as a &OddUintRef.

pub fn random<R: TryRng + ?Sized>(rng: &mut R, bit_length: u32) -> Self

Generate a random Odd<Uint<T>>.

impl<const LIMBS: usize> Odd<Uint<LIMBS>>

pub const fn gcd_unsigned(&self, rhs: &Uint<LIMBS>) -> Self

Compute the greatest common divisor of self and rhs.

pub const fn gcd_unsigned_vartime(&self, rhs: &Uint<LIMBS>) -> Self

Compute the greatest common divisor of self and rhs.

Executes in variable time w.r.t. all input parameters.

pub const fn xgcd(&self, rhs: &Self) -> XgcdOutput<LIMBS, OddUint<LIMBS>>

Execute the Extended GCD algorithm.

Given (self, rhs), computes (g, x, y) s.t. self * x + rhs * y = g = gcd(self, rhs).

impl<const LIMBS: usize> Odd<Uint<LIMBS>>

pub(crate) const fn invert_mod_precision(&self) -> Uint<LIMBS>

Compute a full-width quadratic inversion, self^-1 mod 2^Self::BITS.

pub(crate) const fn invert_mod2k_vartime(&self, k: u32) -> Uint<LIMBS>

Compute a quadratic inversion, self^-1 mod 2^k where k <= Self::BITS.

This method is variable-time in k only.

impl Odd<UintRef>

pub(crate) const fn invert_mod_limb(&self) -> Limb

Returns the multiplicative inverse of the argument modulo 2^N, where 2^N is the capacity of a Limb.

pub const fn invert_mod_u64(&self) -> u64

Returns the multiplicative inverse of the argument modulo 2^64.

impl Odd<BoxedUint>

pub(crate) fn invert_mod_precision(&self) -> BoxedUint

Compute a full-width quadratic inversion, self^-1 mod 2^bits_precision().

pub(crate) fn invert_mod2k_vartime(&self, k: u32) -> BoxedUint

Compute a quadratic inversion, self^-1 mod 2^k where k <= bits_precision().

This method is variable-time in k only.