crypto_bigint::odd

Type Alias OddUintRef

pub type OddUintRef = Odd<UintRef>;

Odd unsigned integer reference.

Aliased Type

#[repr(transparent)]
pub struct OddUintRef(UintRef);

Tuple Fields
0: UintRef
Implementations

impl OddUintRef

pub const fn to_uint_resize<const T: usize>(&self) -> Odd<Uint<T>>

Construct an Odd<Uint<T>> from the unsigned integer value, truncating the upper bits if the value is too large to be represented.