crypto_bigint::prelude

Trait BitOps

pub trait BitOps {

    // Required methods
    fn bits_precision(&self) -> u32;
    fn bytes_precision(&self) -> usize;
    fn bit(&self, index: u32) -> Choice;
    fn set_bit(&mut self, index: u32, bit_value: Choice);
    fn trailing_zeros(&self) -> u32;
    fn trailing_ones(&self) -> u32;
    fn leading_zeros(&self) -> u32;
    fn bit_vartime(&self, index: u32) -> bool;
    fn bits_vartime(&self) -> u32;
    fn set_bit_vartime(&mut self, index: u32, bit_value: bool);
    fn trailing_zeros_vartime(&self) -> u32;
    fn trailing_ones_vartime(&self) -> u32;

    // Provided methods
    fn log2_bits(&self) -> u32 { ... }
    fn bits(&self) -> u32 { ... }
    fn leading_zeros_vartime(&self) -> u32 { ... }

}

Bit counting and bit operations.

Required Methods

fn bits_precision(&self) -> u32

Precision of this integer in bits.

fn bytes_precision(&self) -> usize

Precision of this integer in bytes.

fn bit(&self, index: u32) -> Choice

Get the value of the bit at position index, as a truthy or falsy Choice. Returns the falsy value for indices out of range.

fn set_bit(&mut self, index: u32, bit_value: Choice)

Sets the bit at index to 0 or 1 depending on the value of bit_value.

fn trailing_zeros(&self) -> u32

Calculate the number of trailing zeros in the binary representation of this number.

fn trailing_ones(&self) -> u32

Calculate the number of trailing ones in the binary representation of this number.

fn leading_zeros(&self) -> u32

Calculate the number of leading zeros in the binary representation of this number.

fn bit_vartime(&self, index: u32) -> bool

Returns true if the bit at position index is set, false otherwise.

Remarks

This operation is variable time with respect to index only.

fn bits_vartime(&self) -> u32

Calculate the number of bits required to represent a given number in variable-time with respect to self.

fn set_bit_vartime(&mut self, index: u32, bit_value: bool)

Sets the bit at index to 0 or 1 depending on the value of bit_value, variable time in self.

fn trailing_zeros_vartime(&self) -> u32

Calculate the number of trailing zeros in the binary representation of this number in variable-time with respect to self.

fn trailing_ones_vartime(&self) -> u32

Calculate the number of trailing ones in the binary representation of this number, variable time in self.

Provided Methods

fn log2_bits(&self) -> u32

floor(log2(self.bits_precision())).

fn bits(&self) -> u32

Calculate the number of bits required to represent a given number.

fn leading_zeros_vartime(&self) -> u32

Calculate the number of leading zeros in the binary representation of this number.