Trait crypto_bigint::prelude::Split

pub trait Split: SplitMixed<Self::Output, Self::Output> {
    type Output;

    // Provided method
    fn split(&self) -> (Self::Output, Self::Output) { ... }
}

Split a number in half, returning the most significant half followed by the least significant.

Required Associated Types

type Output

Split output: high/low components of the value.

Provided Methods

fn split(&self) -> (Self::Output, Self::Output)

Split this number in half, returning its high and low components respectively.

Object Safety

This trait is not object safe.