crypto_bigint::prelude

Trait Unsigned

pub trait Unsigned:
    AddMod<Output = Self>
    + BitOps
    + Div<NonZero<Self>, Output = Self>
    + for<'a> Div<&'a NonZero<Self>, Output = Self>
    + DivRemLimb
    + FloorSquareRoot<Output = Self>
    + From<u8>
    + From<u16>
    + From<u32>
    + From<u64>
    + From<Limb>
    + Integer
    + MulMod<Output = Self>
    + NegMod<Output = Self>
    + RemLimb
    + SquareMod<Output = Self>
    + SubMod<Output = Self> {
    type Monty: Monty<Integer = Self>;

    // Required method
    fn from_limb_like(limb: Limb, other: &Self) -> Self;
}

Unsigned Integers.

Required Associated Types

type Monty: Monty<Integer = Self>

The corresponding Montgomery representation, optimized for the performance of modular operations at the price of a conversion overhead.

Required Methods

fn from_limb_like(limb: Limb, other: &Self) -> Self

Returns an integer with the first limb set to limb, and the same precision as other.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.