crypto_bigint::prelude

Trait Xgcd

pub trait Xgcd<Rhs = Self>: Sized {
    type Output;

    // Required methods
    fn xgcd(&self, rhs: &Rhs) -> Self::Output;
    fn xgcd_vartime(&self, rhs: &Rhs) -> Self::Output;
}

Compute the extended greatest common divisor of two integers.

Required Associated Types

type Output

Output type.

Required Methods

fn xgcd(&self, rhs: &Rhs) -> Self::Output

Compute the extended greatest common divisor of self and rhs.

fn xgcd_vartime(&self, rhs: &Rhs) -> Self::Output

Compute the extended greatest common divisor of self and rhs in variable time.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.