crypto_bigint

Module primitives

Functions

borrowing_sub 🔒

Computes lhs - (rhs + borrow), returning the result along with the new borrow.

carrying_add 🔒

Computes lhs + rhs + carry, returning the result along with the new carry (0, 1, or 2).

carrying_mul_add 🔒

Computes (lhs * rhs) + addend + carry, returning the result along with the new carry.

overflowing_add 🔒

Computes lhs + rhs, returning the result along with the carry (0 or 1).

u32_bits 🔒

Compute the number of bits needed to represent n.

u32_max 🔒

const fn equivalent of u32::max(a, b).

u32_min 🔒

const equivalent of u32::min(a, b).

u32_rem 🔒

Remainder calculation, constant time for a given divisor d. Based on “Faster Remainder by Direct Computation: Applications to Compilers and Software Libraries” by Daniel Lemire, Owen Kaser, and Nathan Kurz., Fig. 1.

u64_invert_odd 🔒

Returns the multiplicative inverse of the argument modulo 2^64. The implementation is based on Hurchalla’s method for computing the multiplicative inverse modulo a power of two, and is essentially an optimized Newton iteration.

usize_lt 🔒

Return a Choice representing whether a < b.

widening_mul 🔒

Computes lhs * rhs, returning the low and the high words of the result.