crypto_bigint

Struct BoxedUint

pub struct BoxedUint { /* private fields */ }

Fixed-precision heap-allocated big unsigned integer.

Alternative to the stack-allocated Uint but with a fixed precision chosen at runtime instead of compile time.

Unlike many other heap-allocated big integer libraries, this type is not arbitrary precision and will wrap at its fixed-precision rather than automatically growing.

Implementations

impl BoxedUint

pub fn adc(&self, rhs: &Self, carry: Limb) -> (Self, Limb)

👎Deprecated since 0.7.0: please use carrying_add instead

Computes self + rhs + carry, returning the result along with the new carry.

pub fn carrying_add(&self, rhs: &Self, carry: Limb) -> (Self, Limb)

Computes self + rhs + carry, returning the result along with the new carry.

pub fn adc_assign(&mut self, rhs: impl AsRef<[Limb]>, carry: Limb) -> Limb

👎Deprecated since 0.7.0: please use carrying_add_assign instead

Computes self + rhs + carry in-place, returning the new carry.

Panics if rhs has a larger precision than self.

pub fn carrying_add_assign( &mut self, rhs: impl AsRef<[Limb]>, carry: Limb, ) -> Limb

Computes self + rhs + carry in-place, returning the new carry.

Panics if rhs has a larger precision than self.

pub fn concatenating_add(&self, rhs: &Self) -> Self

Computes self + rhs, returning a result which is concatenated with the overflow limb which would be returned if carrying_add were called with the same operands.

pub fn overflowing_add(&self, rhs: &Self) -> (Self, Choice)

Computes self + rhs, returning a tuple of the sum along with a Choice which indicates whether an overflow occurred.

If an overflow occurred, then the wrapped value is returned.

pub fn wrapping_add(&self, rhs: &Self) -> Self

Perform wrapping addition, discarding overflow.

impl BoxedUint

pub fn add_mod(&self, rhs: &Self, p: &NonZero<Self>) -> Self

Computes self + rhs mod p.

Assumes self + rhs as unbounded integer is < 2p.

pub fn add_mod_assign(&mut self, rhs: &Self, p: &NonZero<Self>)

Computes self + rhs mod p and writes the result in self.

Assumes self + rhs as unbounded integer is < 2p.

pub fn double_mod(&self, p: &NonZero<Self>) -> Self

Computes self + self mod p.

Assumes self as unbounded integer is < p.

impl BoxedUint

pub fn bitand(&self, rhs: &Self) -> Self

Computes bitwise a & b.

pub fn bitand_limb(&self, rhs: Limb) -> Self

Perform bitwise AND between self and the given Limb, performing the AND operation on every limb of self.

pub fn wrapping_and(&self, rhs: &Self) -> Self

Perform wrapping bitwise AND.

There’s no way wrapping could ever happen. This function exists so that all operations are accounted for in the wrapping operations

pub fn checked_and(&self, rhs: &Self) -> CtOption<Self>

Perform checked bitwise AND, returning a CtOption which is_some always

impl BoxedUint

pub fn not(&self) -> Self

Computes bitwise !a.

impl BoxedUint

pub fn bitor(&self, rhs: &Self) -> Self

Computes bitwise a & b.

pub fn wrapping_or(&self, rhs: &Self) -> Self

Perform wrapping bitwise OR.

There’s no way wrapping could ever happen. This function exists so that all operations are accounted for in the wrapping operations

pub fn checked_or(&self, rhs: &Self) -> CtOption<Self>

Perform checked bitwise OR, returning a CtOption which is_some always

impl BoxedUint

pub fn bitxor(&self, rhs: &Self) -> Self

Computes bitwise a ^ b.

pub fn wrapping_xor(&self, rhs: &Self) -> Self

Perform wrapping bitwise `XOR``.

There’s no way wrapping could ever happen. This function exists so that all operations are accounted for in the wrapping operations

pub fn checked_xor(&self, rhs: &Self) -> CtOption<Self>

Perform checked bitwise XOR, returning a CtOption which is_some always

impl BoxedUint

pub fn bit(&self, index: u32) -> Choice

Get the value of the bit at position index, as a truthy or falsy Choice. Returns the falsy value for indices out of range.

pub const fn bit_vartime(&self, index: u32) -> bool

Returns true if the bit at position index is set, false otherwise.

Remarks

This operation is variable time with respect to index only.

pub fn bits(&self) -> u32

Calculate the number of bits needed to represent this number, i.e. the index of the highest set bit.

Use BoxedUint::bits_precision to get the total capacity of this integer.

pub fn bits_vartime(&self) -> u32

Calculate the number of bits needed to represent this number in variable-time with respect to self.

pub const fn leading_zeros(&self) -> u32

Calculate the number of leading zeros in the binary representation of this number.

pub fn bits_precision(&self) -> u32

Get the precision of this BoxedUint in bits.

pub fn trailing_zeros(&self) -> u32

Calculate the number of trailing zeros in the binary representation of this number.

pub fn trailing_ones(&self) -> u32

Calculate the number of trailing ones in the binary representation of this number.

pub fn trailing_zeros_vartime(&self) -> u32

Calculate the number of trailing zeros in the binary representation of this number in variable-time with respect to self.

pub fn trailing_ones_vartime(&self) -> u32

Calculate the number of trailing ones in the binary representation of this number, variable time in self.

impl BoxedUint

pub fn cmp_vartime(&self, rhs: &Self) -> Ordering

Returns the Ordering between self and rhs in variable time.

impl BoxedUint

pub fn div_rem_limb_with_reciprocal( &self, reciprocal: &Reciprocal, ) -> (Self, Limb)

Computes self / rhs using a pre-made reciprocal, returns the quotient (q) and remainder (r).

pub fn div_rem_limb(&self, rhs: NonZero<Limb>) -> (Self, Limb)

Computes self / rhs, returns the quotient (q) and remainder (r).

pub fn rem_limb_with_reciprocal(&self, reciprocal: &Reciprocal) -> Limb

Computes self % rhs using a pre-made reciprocal.

pub fn rem_limb(&self, rhs: NonZero<Limb>) -> Limb

Computes self % rhs.

pub fn div_rem(&self, rhs: &NonZero<Self>) -> (Self, Self)

Computes self / rhs, returns the quotient, remainder.

pub fn rem(&self, rhs: &NonZero<Self>) -> Self

Computes self % rhs, returns the remainder.

pub fn div_rem_vartime(&self, rhs: &NonZero<Self>) -> (Self, Self)

Computes self / rhs, returns the quotient and remainder.

Variable-time with respect to rhs

pub fn rem_vartime(&self, rhs: &NonZero<Self>) -> Self

Computes self % rhs, returns the remainder.

Variable-time with respect to rhs.

pub fn wrapping_div(&self, rhs: &NonZero<Self>) -> Self

Wrapped division is just normal division i.e. self / rhs There’s no way wrapping could ever happen.

This function exists, so that all operations are accounted for in the wrapping operations.

Panics if rhs == 0.

pub fn wrapping_div_vartime(&self, rhs: &NonZero<Self>) -> Self

Wrapped division is just normal division i.e. self / rhs

There’s no way wrapping could ever happen. This function exists, so that all operations are accounted for in the wrapping operations

pub fn checked_div(&self, rhs: &Self) -> CtOption<Self>

Perform checked division, returning a CtOption which is_some only if the rhs != 0

impl BoxedUint

pub fn from_be_slice( bytes: &[u8], bits_precision: u32, ) -> Result<Self, DecodeError>

Create a new BoxedUint from the provided big endian bytes.

The bits_precision argument represents the precision of the resulting integer, which is fixed as this type is not arbitrary-precision. The new BoxedUint will be created with bits_precision rounded up to a multiple of Limb::BITS.

If the length of bytes is larger than bits_precision (rounded up to a multiple of 8) this function will return DecodeError::InputSize. If the size of the decoded integer is larger than bits_precision, this function will return DecodeError::Precision.

pub fn from_be_slice_vartime(bytes: &[u8]) -> Self

Create a new BoxedUint from the provided big endian bytes, automatically selecting its precision based on the size of the input.

This method is variable-time with respect to all subsequent operations since it chooses the limb count based on the input size, and is therefore only suitable for public inputs.

When working with secret values, use BoxedUint::from_be_slice.

pub fn from_le_slice( bytes: &[u8], bits_precision: u32, ) -> Result<Self, DecodeError>

Create a new BoxedUint from the provided little endian bytes.

The bits_precision argument represents the precision of the resulting integer, which is fixed as this type is not arbitrary-precision. The new BoxedUint will be created with bits_precision rounded up to a multiple of Limb::BITS.

If the length of bytes is larger than bits_precision (rounded up to a multiple of 8) this function will return DecodeError::InputSize. If the size of the decoded integer is larger than bits_precision, this function will return DecodeError::Precision.

pub fn from_le_slice_vartime(bytes: &[u8]) -> Self

Create a new BoxedUint from the provided little endian bytes, automatically selecting its precision based on the size of the input.

This method is variable-time with respect to all subsequent operations since it chooses the limb count based on the input size, and is therefore only suitable for public inputs.

When working with secret values, use BoxedUint::from_le_slice.

pub fn to_be_bytes(&self) -> Box<[u8]>

Serialize this BoxedUint as big-endian.

pub fn to_be_bytes_trimmed_vartime(&self) -> Box<[u8]>

Serialize this BoxedUint as big-endian without leading zeroes.

pub fn to_le_bytes(&self) -> Box<[u8]>

Serialize this BoxedUint as little-endian.

pub fn to_le_bytes_trimmed_vartime(&self) -> Box<[u8]>

Serialize this BoxedUint as little-endian without trailing zeroes.

pub fn from_be_hex(hex: &str, bits_precision: u32) -> CtOption<Self>

Create a new BoxedUint from the provided big endian hex string.

pub fn from_str_radix_vartime( src: &str, radix: u32, ) -> Result<Self, DecodeError>

Create a new BoxedUint from a big-endian string in a given base.

The string may begin with a + character, and may use underscore characters to separate digits.

If the input value contains non-digit characters or digits outside of the range 0..radix this function will return DecodeError::InvalidDigit. Panics if radix is not in the range from 2 to 36.

pub fn from_str_radix_with_precision_vartime( src: &str, radix: u32, bits_precision: u32, ) -> Result<Self, DecodeError>

Create a new BoxedUint from a big-endian string in a given base, with a given precision.

The string may begin with a + character, and may use underscore characters to separate digits.

The bits_precision argument represents the precision of the resulting integer, which is fixed as this type is not arbitrary-precision. The new BoxedUint will be created with bits_precision rounded up to a multiple of Limb::BITS.

If the input value contains non-digit characters or digits outside of the range 0..radix this function will return DecodeError::InvalidDigit. If the length of bytes is larger than bits_precision (rounded up to a multiple of 8) this function will return DecodeError::InputSize. If the size of the decoded integer is larger than bits_precision, this function will return DecodeError::Precision. Panics if radix is not in the range from 2 to 36.

pub fn to_string_radix_vartime(&self, radix: u32) -> String

Format a BoxedUint as a string in a given base.

Panics if radix is not in the range from 2 to 36.

impl BoxedUint

pub fn inv_odd_mod(&self, modulus: &Odd<Self>) -> CtOption<Self>

👎Deprecated since 0.7.0: please use invert_odd_mod instead

Computes the multiplicative inverse of self mod modulus, where modulus is odd.

pub fn invert_odd_mod(&self, modulus: &Odd<Self>) -> CtOption<Self>

Computes the multiplicative inverse of self mod modulus, where modulus is odd.

pub fn invert_odd_mod_vartime(&self, modulus: &Odd<Self>) -> CtOption<Self>

Computes the multiplicative inverse of self mod modulus, where modulus is odd.

pub fn inv_mod2k_vartime(&self, k: u32) -> (Self, Choice)

👎Deprecated since 0.7.0: please use invert_mod2k_vartime instead

Computes 1/self mod 2^k. This method is constant-time w.r.t. self but not k.

If the inverse does not exist (k > 0 and self is even, or k > bits_precision()), returns Choice::FALSE as the second element of the tuple, otherwise returns Choice::TRUE.

pub fn invert_mod2k_vartime(&self, k: u32) -> (Self, Choice)

Computes 1/self mod 2^k. This method is constant-time w.r.t. self but not k.

If the inverse does not exist (k > 0 and self is even, or k > bits_precision()), returns Choice::FALSE as the second element of the tuple, otherwise returns Choice::TRUE.

pub fn inv_mod2k(&self, k: u32) -> (Self, Choice)

👎Deprecated since 0.7.0: please use invert_mod2k instead

Computes 1/self mod 2^k.

If the inverse does not exist (k > 0 and self is even, or k > bits_precision()), returns Choice::FALSE as the second element of the tuple, otherwise returns Choice::TRUE.

pub fn invert_mod2k(&self, k: u32) -> (Self, Choice)

Computes 1/self mod 2^k.

If the inverse does not exist (k > 0 and self is even, or k > bits_precision()), returns Choice::FALSE as the second element of the tuple, otherwise returns Choice::TRUE.

pub fn inv_mod(&self, modulus: &Self) -> CtOption<Self>

👎Deprecated since 0.7.0: please use invert_mod instead

Computes the multiplicative inverse of self mod modulus

self and modulus must have the same number of limbs, or the function will panic

TODO: maybe some better documentation is needed

pub fn invert_mod(&self, modulus: &NonZero<Self>) -> CtOption<Self>

Computes the multiplicative inverse of self mod modulus

self and modulus must have the same number of limbs, or the function will panic

TODO: maybe some better documentation is needed

impl BoxedUint

pub fn mul(&self, rhs: &Self) -> Self

Multiply self by rhs.

Returns a widened output with a limb count equal to the sums of the input limb counts.

pub fn wrapping_mul(&self, rhs: &Self) -> Self

Perform wrapping multiplication, wrapping to the width of self.

pub fn checked_mul(&self, rhs: &Self) -> CtOption<Self>

Multiply self by rhs, wrapping to the width of self. Returns CtOption::None if the result overflowed the precision of self.

pub fn saturating_mul(&self, rhs: &BoxedUint) -> Self

Perform saturating multiplication, returning MAX on overflow.

pub fn square(&self) -> Self

Multiply self by itself.

pub fn wrapping_square(&self) -> Self

Multiply self by itself, wrapping to the width of self.

pub fn checked_square(&self) -> CtOption<Self>

Multiply self by itself, wrapping to the width of self. Returns CtOption::None if the result overflowed the precision of self.

pub fn saturating_square(&self) -> Self

Perform saturating squaring, returning MAX on overflow.

impl BoxedUint

pub fn mul_mod(&self, rhs: &BoxedUint, p: &NonZero<BoxedUint>) -> BoxedUint

Computes self * rhs mod p for non-zero p.

pub fn mul_mod_special(&self, rhs: &Self, c: Limb) -> Self

Computes self * rhs mod p for the special modulus p = MAX+1-c where c is small enough to fit in a single Limb.

For the modulus reduction, this function implements Algorithm 14.47 from the “Handbook of Applied Cryptography”, by A. Menezes, P. van Oorschot, and S. Vanstone, CRC Press, 1996.

pub fn square_mod(&self, p: &NonZero<BoxedUint>) -> Self

Computes self * self mod p.

pub fn square_mod_vartime(&self, p: &NonZero<BoxedUint>) -> Self

Computes self * self mod p in variable time with respect to p.

impl BoxedUint

pub fn wrapping_neg(&self) -> Self

Perform wrapping negation.

impl BoxedUint

pub fn neg_mod(&self, p: &NonZero<Self>) -> Self

Computes -a mod p. Assumes self is in [0, p).

pub fn neg_mod_special(&self, c: Limb) -> Self

Computes -a mod p for the special modulus p = MAX+1-c where c is small enough to fit in a single Limb.

impl BoxedUint

pub fn pow_mod(&self, rhs: &BoxedUint, p: &Odd<BoxedUint>) -> BoxedUint

Computes self ^ rhs mod p for odd p.

impl BoxedUint

pub fn shl(&self, shift: u32) -> BoxedUint

Computes self << shift.

Panics if shift >= Self::BITS.

pub fn shl_assign(&mut self, shift: u32)

Computes self <<= shift.

Panics if shift >= Self::BITS.

pub fn overflowing_shl(&self, shift: u32) -> (Self, Choice)

Computes self << shift.

Returns a zero and a truthy Choice if shift >= self.bits_precision(), or the result and a falsy Choice otherwise.

pub fn overflowing_shl_vartime(&self, shift: u32) -> (Self, Choice)

Computes self << shift in variable-time.

Returns a zero and a truthy Choice if shift >= self.bits_precision(), or the result and a falsy Choice otherwise.

pub fn overflowing_shl_assign(&mut self, shift: u32) -> Choice

Computes self <<= shift.

Returns a truthy Choice if shift >= self.bits_precision() or a falsy Choice otherwise.

pub fn overflowing_shl_assign_vartime(&mut self, shift: u32) -> Choice

Computes self <<= shift in variable-time.

Returns a truthy Choice if shift >= self.bits_precision() or a falsy Choice otherwise.

pub fn wrapping_shl(&self, shift: u32) -> Self

Computes self << shift in a panic-free manner, masking off bits of shift which would cause the shift to exceed the type’s width.

pub fn wrapping_shl_assign(&mut self, shift: u32)

Computes self <<= shift in a panic-free manner, masking off bits of shift which would cause the shift to exceed the type’s width.

pub fn wrapping_shl_vartime(&self, shift: u32) -> Self

Computes self << shift in variable-time in a panic-free manner, masking off bits of shift which would cause the shift to exceed the type’s width.

pub fn wrapping_shl_assign_vartime(&mut self, shift: u32)

Computes self <<= shift in variable-time in a panic-free manner, producing zero in the case of overflow.

pub fn shl_vartime(&self, shift: u32) -> Option<Self>

Computes self << shift. Returns None if shift >= self.bits_precision().

NOTE: this operation is variable time with respect to shift ONLY.

When used with a fixed shift, this function is constant-time with respect to self.

impl BoxedUint

pub fn shr(&self, shift: u32) -> BoxedUint

Computes self >> shift.

Panics if shift >= Self::BITS.

pub fn shr_assign(&mut self, shift: u32)

Computes self >>= shift.

Panics if shift >= Self::BITS.

pub fn overflowing_shr(&self, shift: u32) -> (Self, Choice)

Computes self >> shift.

Returns a zero and a truthy Choice if shift >= self.bits_precision(), or the result and a falsy Choice otherwise.

pub fn overflowing_shr_vartime(&self, shift: u32) -> (Self, Choice)

Computes self >> shift in variable-time.

Returns a zero and a truthy Choice if shift >= self.bits_precision(), or the result and a falsy Choice otherwise.

pub fn overflowing_shr_assign(&mut self, shift: u32) -> Choice

Computes self >>= shift.

Returns a truthy Choice if shift >= self.bits_precision() or a falsy Choice otherwise.

pub fn overflowing_shr_assign_vartime(&mut self, shift: u32) -> Choice

Computes self >>= shift in variable-time.

Returns a truthy Choice if shift >= self.bits_precision() or a falsy Choice otherwise.

pub fn wrapping_shr(&self, shift: u32) -> Self

Computes self >> shift in a panic-free manner, masking off bits of shift which would cause the shift to exceed the type’s width.

pub fn wrapping_shr_assign(&mut self, shift: u32)

Computes self >>= shift in a panic-free manner, masking off bits of shift which would cause the shift to exceed the type’s width.

pub fn wrapping_shr_vartime(&self, shift: u32) -> Self

Computes self >> shift in variable-time in a panic-free manner, masking off bits of shift which would cause the shift to exceed the type’s width.

pub fn wrapping_shr_assign_vartime(&mut self, shift: u32)

Computes self >>= shift in variable-time in a panic-free manner, producing zero in the case of overflow.

pub fn shr_vartime(&self, shift: u32) -> Option<Self>

Computes self >> shift. Returns None if shift >= self.bits_precision().

NOTE: this operation is variable time with respect to shift ONLY.

When used with a fixed shift, this function is constant-time with respect to self.

impl BoxedUint

pub fn sqrt(&self) -> Self

👎Deprecated since 0.7.0: please use floor_sqrt instead

Computes floor(√(self)) in constant time.

Callers can check if self is a square by squaring the result.

pub fn floor_sqrt(&self) -> Self

Computes √(self) in constant time.

Callers can check if self is a square by squaring the result.

pub fn sqrt_vartime(&self) -> Self

👎Deprecated since 0.7.0: please use floor_sqrt_vartime instead

Computes floor(√(self)).

Callers can check if self is a square by squaring the result.

Variable time with respect to self.

pub fn floor_sqrt_vartime(&self) -> Self

Computes √(self).

Callers can check if self is a square by squaring the result.

Variable time with respect to self.

pub fn wrapping_sqrt(&self) -> Self

Wrapped sqrt is just floor(√(self)). There’s no way wrapping could ever happen. This function exists so that all operations are accounted for in the wrapping operations.

pub fn wrapping_sqrt_vartime(&self) -> Self

Wrapped sqrt is just floor(√(self)). There’s no way wrapping could ever happen. This function exists so that all operations are accounted for in the wrapping operations.

Variable time with respect to self.

pub fn checked_sqrt(&self) -> CtOption<Self>

Perform checked sqrt, returning a CtOption which is_some only if the square root is exact.

pub fn checked_sqrt_vartime(&self) -> Option<Self>

Perform checked sqrt, returning an Option which is_some only if the square root is exact.

Variable time with respect to self.

impl BoxedUint

pub fn sbb(&self, rhs: &Self, borrow: Limb) -> (Self, Limb)

👎Deprecated since 0.7.0: please use borrowing_sub instead

Computes self - (rhs + borrow), returning the result along with the new borrow.

pub fn borrowing_sub(&self, rhs: &Self, borrow: Limb) -> (Self, Limb)

Computes self - (rhs + borrow), returning the result along with the new borrow.

pub fn sbb_assign(&mut self, rhs: impl AsRef<[Limb]>, borrow: Limb) -> Limb

👎Deprecated since 0.7.0: please use borrowing_sub_assign instead

Computes a - (b + borrow) in-place, returning the new borrow.

Panics if rhs has a larger precision than self.

pub fn borrowing_sub_assign( &mut self, rhs: impl AsRef<[Limb]>, borrow: Limb, ) -> Limb

Computes a - (b + borrow) in-place, returning the new borrow.

Panics if rhs has a larger precision than self.

pub fn wrapping_sub(&self, rhs: &Self) -> Self

Perform wrapping subtraction, discarding overflow.

impl BoxedUint

pub fn sub_mod(&self, rhs: &Self, p: &NonZero<Self>) -> Self

Computes self - rhs mod p.

Assumes self - rhs as unbounded signed integer is in [-p, p).

pub fn sub_mod_special(&self, rhs: &Self, c: Limb) -> Self

Computes self - rhs mod p for the special modulus p = MAX+1-c where c is small enough to fit in a single Limb.

Assumes self - rhs as unbounded signed integer is in [-p, p).

impl BoxedUint

pub fn zero() -> Self

Get the value 0 represented as succinctly as possible.

pub fn zero_with_precision(at_least_bits_precision: u32) -> Self

Get the value 0 with the given number of bits of precision.

at_least_bits_precision is rounded up to a multiple of Limb::BITS.

pub fn one() -> Self

Get the value 1, represented as succinctly as possible.

pub fn one_with_precision(at_least_bits_precision: u32) -> Self

Get the value 1 with the given number of bits of precision.

at_least_bits_precision is rounded up to a multiple of Limb::BITS.

pub fn is_zero(&self) -> Choice

Is this BoxedUint equal to zero?

pub fn is_nonzero(&self) -> Choice

Is this BoxedUint NOT equal to zero?

pub fn is_one(&self) -> Choice

Is this BoxedUint equal to one?

pub fn max(at_least_bits_precision: u32) -> Self

Get the maximum value for a BoxedUint created with at_least_bits_precision precision bits requested.

That is, returns the value 2^self.bits_precision() - 1.

pub fn from_words(words: impl IntoIterator<Item = Word>) -> Self

Create a BoxedUint from an array of Words (i.e. word-sized unsigned integers).

pub fn from_words_with_precision( words: impl IntoIterator<Item = Word>, at_least_bits_precision: u32, ) -> Self

Create a BoxedUint from an array of Words (i.e. word-sized unsigned integers), specifying the precision of the result. Any words above the given precision will be dropped.

pub fn to_words(&self) -> Box<[Word]>

Create a boxed slice of Words (i.e. word-sized unsigned integers) from a BoxedUint.

pub fn as_words(&self) -> &[Word]

Borrow the inner limbs as a slice of Words.

pub fn as_mut_words(&mut self) -> &mut [Word]

Borrow the inner limbs as a mutable slice of Words.

pub fn as_words_mut(&mut self) -> &mut [Word]

👎Deprecated since 0.7.0: please use as_mut_words instead

Borrow the inner limbs as a mutable slice of Words.

pub fn as_limbs(&self) -> &[Limb]

Borrow the limbs of this BoxedUint.

pub fn as_mut_limbs(&mut self) -> &mut [Limb]

Borrow the limbs of this BoxedUint mutably.

pub fn as_limbs_mut(&mut self) -> &mut [Limb]

👎Deprecated since 0.7.0: please use as_mut_limbs instead

Borrow the limbs of this BoxedUint mutably.

pub fn to_limbs(&self) -> Box<[Limb]>

Convert this BoxedUint into its inner limbs.

pub fn into_limbs(self) -> Box<[Limb]>

Convert this BoxedUint into its inner limbs.

pub fn nlimbs(&self) -> usize

Get the number of limbs in this BoxedUint.

pub fn to_nz(&self) -> CtOption<NonZero<Self>>

Convert to a NonZero<BoxedUint>.

Returns some if the original value is non-zero, and false otherwise.

pub fn to_odd(&self) -> CtOption<Odd<Self>>

Convert to an Odd<BoxedUint>.

Returns some if the original value is odd, and false otherwise.

pub fn into_nz(self) -> CtOption<NonZero<Self>>

Convert to a NonZero<BoxedUint>.

Returns some if the original value is non-zero, and false otherwise.

pub fn into_odd(self) -> CtOption<Odd<Self>>

Convert to an Odd<BoxedUint>.

Returns some if the original value is odd, and false otherwise.

pub fn widen(&self, at_least_bits_precision: u32) -> BoxedUint

👎Deprecated since 0.7.0: please use resize instead

Widen this type’s precision to the given number of bits.

Panics if at_least_bits_precision is smaller than the current precision.

pub fn shorten(&self, at_least_bits_precision: u32) -> BoxedUint

👎Deprecated since 0.7.0: please use resize instead

Shortens this type’s precision to the given number of bits.

Panics if at_least_bits_precision is larger than the current precision.