crypto_bigint

Struct Limb

#[repr(transparent)]
pub struct Limb(pub Word);

Big integers are represented as an array/vector of smaller CPU word-size integers called “limbs”.

The Limb type uses a 32-bit or 64-bit saturated representation, depending on the target. All bits of an inner Word are used to represent larger big integer types.

Tuple Fields
0: Word
Implementations

impl Limb

pub const fn adc(self, rhs: Limb, carry: Limb) -> (Limb, Limb)

👎Deprecated since 0.7.0: please use carrying_add instead

Computes self + rhs + carry, returning the result along with the new carry.

pub const fn carrying_add(self, rhs: Limb, carry: Limb) -> (Limb, Limb)

Computes self + rhs + carry, returning the result along with the new carry.

pub const fn overflowing_add(self, rhs: Limb) -> (Limb, Limb)

Computes self + rhs, returning the result along with the carry.

pub const fn saturating_add(&self, rhs: Self) -> Self

Perform saturating addition.

pub const fn wrapping_add(&self, rhs: Self) -> Self

Perform wrapping addition, discarding overflow.

impl Limb

pub const fn bitand(self, rhs: Self) -> Self

Calculates a & b.

impl Limb

pub const fn not(self) -> Self

Calculates !a.

impl Limb

pub const fn bitor(self, rhs: Self) -> Self

Calculates a | b.

impl Limb

pub const fn bitxor(self, rhs: Self) -> Self

Calculates a ^ b.

impl Limb

pub const fn bits(self) -> u32

Calculate the number of bits needed to represent this number.

pub const fn leading_zeros(self) -> u32

Calculate the number of leading zeros in the binary representation of this number.

pub const fn trailing_zeros(self) -> u32

Calculate the number of trailing zeros in the binary representation of this number.

pub const fn trailing_ones(self) -> u32

Calculate the number of trailing ones the binary representation of this number.

pub const fn restrict_bits(self, len: u32) -> Self

Clear bits at or above the given bit position.

impl Limb

pub fn is_odd(&self) -> Choice

Is this limb an odd number?

pub fn cmp_vartime(&self, other: &Self) -> Ordering

Perform a comparison of the inner value in variable-time.

Note that the PartialOrd and Ord impls wrap constant-time comparisons using the subtle crate.

pub const fn eq_vartime(&self, other: &Self) -> bool

Performs an equality check in variable-time.

impl Limb

pub const fn from_u8(n: u8) -> Self

Create a Limb from a u8 integer (const-friendly)

pub const fn from_u16(n: u16) -> Self

Create a Limb from a u16 integer (const-friendly)

pub const fn from_u32(n: u32) -> Self

Create a Limb from a u32 integer (const-friendly)

pub const fn from_u64(n: u64) -> Self

Create a Limb from a u64 integer (const-friendly)

impl Limb

pub const fn mac(self, b: Limb, c: Limb, carry: Limb) -> (Limb, Limb)

👎Deprecated since 0.7.0: please use carrying_mul_add instead (ordering of arguments changes)

Computes self + (b * c) + carry, returning the result along with the new carry.

pub const fn carrying_mul_add( self, rhs: Limb, addend: Limb, carry: Limb, ) -> (Limb, Limb)

Computes (self * rhs) + addend + carry, returning the result along with the new carry.

pub const fn saturating_mul(&self, rhs: Self) -> Self

Perform saturating multiplication.

pub const fn wrapping_mul(&self, rhs: Self) -> Self

Perform wrapping multiplication, discarding overflow.

impl Limb

pub const fn wrapping_neg(self) -> Self

Perform wrapping negation.

impl Limb

pub const fn shl(self, shift: u32) -> Self

Computes self << shift. Panics if shift overflows Limb::BITS.

impl Limb

pub const fn shr(self, shift: u32) -> Self

Computes self >> shift. Panics if shift overflows Limb::BITS.

impl Limb

pub const fn sbb(self, rhs: Limb, borrow: Limb) -> (Limb, Limb)

👎Deprecated since 0.7.0: please use borrowing_sub instead

Computes self - (rhs + borrow), returning the result along with the new borrow.

pub const fn borrowing_sub(self, rhs: Limb, borrow: Limb) -> (Limb, Limb)

Computes self - (rhs + borrow), returning the result along with the new borrow.

pub const fn saturating_sub(&self, rhs: Self) -> Self

Perform saturating subtraction.

pub const fn wrapping_sub(&self, rhs: Self) -> Self

Perform wrapping subtraction, discarding underflow and wrapping around the boundary of the type.

impl Limb

pub const ZERO: Self

The value 0.

pub const ONE: Self

The value 1.

pub const MAX: Self

Maximum value this Limb can express.

pub const BITS: u32 = 64

Size of the inner integer in bits.

pub const BYTES: usize = 8

Size of the inner integer in bytes.

pub const LOG2_BITS: u32

floor(log2(Self::BITS)).

pub const fn is_zero(&self) -> Choice

Is this limb equal to Limb::ZERO?

pub const fn to_nz(self) -> CtOption<NonZero<Self>>

Convert to a NonZero<Limb>.

Returns some if the original value is non-zero, and false otherwise.

pub const fn lsb_to_choice(self) -> Choice

Convert the least significant bit of this Limb to a Choice.