crypto_bigint

Struct NonZero

pub struct NonZero<T: ?Sized>(/* private fields */);

Wrapper type for non-zero integers.

Implementations

impl<const LIMBS: usize> NonZero<Int<LIMBS>>

pub const fn gcd_unsigned(&self, rhs: &Uint<LIMBS>) -> NonZero<Uint<LIMBS>>

Compute the greatest common divisor of self and rhs.

pub const fn gcd_unsigned_vartime( &self, rhs: &Uint<LIMBS>, ) -> NonZeroUint<LIMBS>

Compute the greatest common divisor of self and rhs.

Executes in variable time w.r.t. all input parameters.

pub const fn xgcd(&self, rhs: &Self) -> XgcdOutput<LIMBS, NonZeroUint<LIMBS>>

Execute the Extended GCD algorithm.

Given (self, rhs), computes (g, x, y) s.t. self * x + rhs * y = g = gcd(self, rhs).

impl<T> NonZero<T>

pub fn new(n: T) -> CtOption<Self>
where T: Zero + One + CtSelect,

Create a new non-zero integer.

pub fn get(self) -> T

Returns the inner value.

pub const fn get_copy(self) -> T
where T: Copy,

Returns a copy of the inner value for Copy types.

This allows the function to be const fn, since Copy is implicitly !Drop, which avoids problems around const fn destructors.

impl<T: ?Sized> NonZero<T>

pub const fn as_ref(&self) -> &T

Provides access to the contents of NonZero in a const context.

impl<T> NonZero<T>
where T: Bounded + ?Sized,

pub const BITS: u32 = T::BITS

Total size of the represented integer in bits.

pub const BYTES: usize = T::BYTES

Total size of the represented integer in bytes.

impl<T> NonZero<T>
where T: Constants,

pub const ONE: Self

The value 1.

pub const MAX: Self

Maximum value this integer can express.

impl<T> NonZero<T>
where T: Zero + One + CtSelect + Encoding,

pub fn from_be_bytes(bytes: T::Repr) -> CtOption<Self>

Decode from big endian bytes.

pub fn from_le_bytes(bytes: T::Repr) -> CtOption<Self>

Decode from little endian bytes.

impl NonZero<Limb>

pub const fn new_unwrap(n: Limb) -> Self

Creates a new non-zero limb in a const context. Panics if the value is zero.

In future versions of Rust it should be possible to replace this with NonZero::new(…).unwrap()

pub const fn from_u8(n: NonZeroU8) -> Self

Create a NonZero<Limb> from a NonZeroU8 (const-friendly)

pub const fn from_u16(n: NonZeroU16) -> Self

Create a NonZero<Limb> from a NonZeroU16 (const-friendly)

pub const fn from_u32(n: NonZeroU32) -> Self

Create a NonZero<Limb> from a NonZeroU32 (const-friendly)

pub const fn from_u64(n: NonZeroU64) -> Self

Create a NonZero<Limb> from a NonZeroU64 (const-friendly)

impl<const LIMBS: usize> NonZero<Uint<LIMBS>>

pub const fn new_unwrap(n: Uint<LIMBS>) -> Self

Creates a new non-zero integer in a const context.

In future versions of Rust it should be possible to replace this with NonZero::new(…).unwrap()

Panics

if the value is zero.

pub const fn from_be_hex(hex: &str) -> Self

Create a new NonZero<Uint> from the provided big endian hex string.

Panics

if the hex is zero, malformed, or not zero-padded accordingly for the size.

pub const fn from_le_hex(hex: &str) -> Self

Create a new NonZero<Uint> from the provided little endian hex string.

Panics

if the hex is zero, malformed, or not zero-padded accordingly for the size.

pub const fn from_u8(n: NonZeroU8) -> Self

Create a NonZeroUint from a NonZeroU8 (const-friendly)

pub const fn from_u16(n: NonZeroU16) -> Self

Create a NonZeroUint from a NonZeroU16 (const-friendly)

pub const fn from_u32(n: NonZeroU32) -> Self

Create a NonZeroUint from a NonZeroU32 (const-friendly)

pub const fn from_u64(n: NonZeroU64) -> Self

Create a NonZeroUint from a NonZeroU64 (const-friendly)

pub const fn from_u128(n: NonZeroU128) -> Self

Create a NonZeroUint from a NonZeroU128 (const-friendly)

impl<const LIMBS: usize> NonZero<Int<LIMBS>>

pub const fn new_unwrap(n: Int<LIMBS>) -> Self

Creates a new non-zero integer in a const context. Panics if the value is zero.

In future versions of Rust it should be possible to replace this with NonZero::new(…).unwrap()

pub const fn abs_sign(&self) -> (NonZero<Uint<LIMBS>>, Choice)

The sign and magnitude of this NonZeroInt.

pub const fn abs(&self) -> NonZero<Uint<LIMBS>>

The magnitude of this NonZeroInt.

impl<const LIMBS: usize> NonZero<Uint<LIMBS>>

pub const fn gcd_unsigned(&self, rhs: &Uint<LIMBS>) -> Self

Compute the greatest common divisor of self and rhs.

pub const fn gcd_unsigned_vartime(&self, rhs: &Uint<LIMBS>) -> Self

Compute the greatest common divisor of self and rhs.

Executes in variable time w.r.t. all input parameters.

pub const fn xgcd(&self, rhs: &Self) -> XgcdOutput<LIMBS, NonZeroUint<LIMBS>>

Execute the Extended GCD algorithm.

Given (self, rhs), computes (g, x, y) s.t. self * x + rhs * y = g = gcd(self, rhs).

impl<const LIMBS: usize> NonZero<Uint<LIMBS>>

pub const fn floor_sqrt(&self) -> Self

Computes floor(√(self)) in constant time.

Callers can check if self is a square by squaring the result.

pub const fn floor_sqrt_vartime(&self) -> Self

Computes floor(√(self)).

Callers can check if self is a square by squaring the result.

Variable time with respect to self.

pub fn checked_sqrt(&self) -> CtOption<Self>

Perform checked sqrt, returning a CtOption which is_some only if the square root is exact.

pub fn checked_sqrt_vartime(&self) -> Option<Self>

Perform checked sqrt, returning an Option which is_some only if the square root is exact.