Trait crypto_bigint::Concat

pub trait Concat<Rhs = Self> {
    type Output;

    // Required method
    fn concat(&self, rhs: &Self) -> Self::Output;
}

Concatenate two numbers into a “wide” twice-width value, using the rhs value as the least significant value.

Required Associated Types

type Output

Concatenated output: twice the width of Self.

Required Methods

fn concat(&self, rhs: &Self) -> Self::Output

Concate the two values, with self as most significant and rhs as the least significant.