crypto_bigint

Trait Split

pub trait Split<const LO: usize> {
    type Output: Integer;
}

Define the result of splitting a number into two parts, with the first part having the width LO.

Required Associated Types

type Output: Integer

High limbs of output: having the width of Self minus LO.