Trait crypto_bigint::SplitMixed

pub trait SplitMixed<Hi, Lo> {
    // Required method
    fn split_mixed(&self) -> (Hi, Lo);
}

Split a number into parts, returning the most significant part followed by the least significant.

Required Methods

fn split_mixed(&self) -> (Hi, Lo)

Split this number into parts, returning its high and low components respectively.