crypto_bigint::traits

Trait CtSelect

pub trait CtSelect: Sized {
    // Required method
    fn ct_select(&self, other: &Self, choice: Choice) -> Self;

    // Provided method
    fn ct_swap(&mut self, other: &mut Self, choice: Choice) { ... }
}

Constant-time selection: choose between two values based on a given Choice.

This crate provides built-in implementations for the following types:

i8, i16, i32, i64, i128, isize

u8, u16, u32, u64, u128, usize

NonZeroI8, NonZeroI16, NonZeroI32, NonZeroI64, NonZeroI128, NonZeroI128

NonZeroU8, NonZeroU16, NonZeroU32, NonZeroU64, NonZeroU128,, NonZeroUsize

cmp::Ordering

Choice

[T; N] where T impls CtSelectArray, which the previously mentioned types all do, as well as any type which impls Clone + CtAssignSlice + CtSelect.

Required Methods

fn ct_select(&self, other: &Self, choice: Choice) -> Self

Select between self and other based on choice, returning a copy of the value.

Returns

self if choice is Choice::FALSE.

other if choice is Choice::TRUE.

Provided Methods

fn ct_swap(&mut self, other: &mut Self, choice: Choice)

Conditionally swap self and other if choice is Choice::TRUE.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.