crypto_bigint

Type Alias OddUint

pub type OddUint<const LIMBS: usize> = Odd<Uint<LIMBS>>;

Non-zero unsigned integer.

Aliased Type

pub struct OddUint<const LIMBS: usize>(/* private fields */);

Implementations

impl<const LIMBS: usize> OddUint<LIMBS>

pub const fn gcd_unsigned(&self, rhs: &Uint<LIMBS>) -> Self

Compute the greatest common divisor of self and rhs.

pub const fn gcd_unsigned_vartime(&self, rhs: &Uint<LIMBS>) -> Self

Compute the greatest common divisor of self and rhs.

Executes in variable time w.r.t. all input parameters.

pub const fn xgcd(&self, rhs: &Self) -> XgcdOutput<LIMBS, OddUint<LIMBS>>

Execute the Extended GCD algorithm.

Given (self, rhs), computes (g, x, y) s.t. self * x + rhs * y = g = gcd(self, rhs).