crypto_bigint

Type Alias OddUintRef

pub type OddUintRef = Odd<UintRef>;

Odd unsigned integer reference.

Aliased Type

pub struct OddUintRef(/* private fields */);

Implementations

impl OddUintRef

pub const fn to_uint_resize<const T: usize>(&self) -> Odd<Uint<T>>

Construct an Odd<Uint<T>> from the unsigned integer value, truncating the upper bits if the value is too large to be represented.