Type Alias crypto_bigint::U128

pub type U128 = UInt<{ _ }>;

128-bit unsigned big integer

Aliased Type

struct U128 { /* private fields */ }

Implementations

impl<const LIMBS: usize> UInt<LIMBS>

pub const fn adc(&self, rhs: &Self, carry: Limb) -> (Self, Limb)

Computes a + b + carry, returning the result along with the new carry.

pub const fn wrapping_add(&self, rhs: &Self) -> Self

Perform wrapping addition, discarding overflow.

pub fn checked_add(&self, rhs: &Self) -> CtOption<Self>

Perform checked addition, returning a CtOption which is_some only if the operation did not overflow.

impl<const LIMBS: usize> UInt<LIMBS>

pub const fn add_mod(&self, rhs: &UInt<LIMBS>, p: &UInt<LIMBS>) -> UInt<LIMBS>

Computes self + rhs mod p in constant time.

Assumes self and rhs are < p.

impl<const LIMBS: usize> UInt<LIMBS>

pub fn is_zero(&self) -> Choice

Determine if this UInt is equal to zero.

Returns

If zero, return Choice(1). Otherwise, return Choice(0).

pub fn is_odd(&self) -> Choice

Is this UInt an odd number?

impl<const LIMBS: usize> UInt<LIMBS>

pub const fn from_be_slice(bytes: &[u8]) -> Self

Create a new UInt from the provided big endian bytes.

pub const fn from_be_hex(hex: &str) -> Self

Create a new UInt from the provided big endian hex string.

pub const fn from_le_slice(bytes: &[u8]) -> Self

Create a new UInt from the provided little endian bytes.

pub const fn from_le_hex(hex: &str) -> Self

Create a new UInt from the provided little endian hex string.

impl<const LIMBS: usize> UInt<LIMBS>

pub const fn from_u8(n: u8) -> Self

Create a UInt from a u8 (const-friendly)

pub const fn from_u16(n: u16) -> Self

Create a UInt from a u16 (const-friendly)

pub const fn from_u32(n: u32) -> Self

Create a UInt from a u32 (const-friendly)

pub const fn from_u64(n: u64) -> Self

Create a UInt from a u64 (const-friendly)

pub const fn from_u128(n: u128) -> Self

Create a UInt from a u128 (const-friendly)

pub const fn from_uint_array(arr: [Inner; LIMBS]) -> Self

Create a UInt from an array of the limb::Inner unsigned integer type.

pub const fn to_uint_array(self) -> [Inner; LIMBS]

Create an array of limb::Inner unsigned integer type from a UInt.

impl<const LIMBS: usize> UInt<LIMBS>

pub const fn mul_wide(&self, rhs: &Self) -> (Self, Self)

Compute “wide” multiplication, with a product twice the size of the input.

pub const fn wrapping_mul(&self, rhs: &Self) -> Self

Perform wrapping multiplication, discarding overflow.

pub fn checked_mul(&self, rhs: &Self) -> CtOption<Self>

Perform checked multiplication, returning a CtOption which is_some only if the operation did not overflow.

pub fn square(&self) -> <Self as Concat>::Outputwhere Self: Concat,

Square self, returning a “wide” result.

impl<const LIMBS: usize> UInt<LIMBS>

pub const fn neg_mod(&self, p: &Self) -> Self

Computes -a mod p in constant time.

impl<const LIMBS: usize> UInt<LIMBS>

pub const fn shr_vartime(&self, shift: usize) -> Self

Computes self >> n.

NOTE: this operation is variable time with respect to n ONLY.

When used with a fixed n, this function is constant-time with respect to self.

impl<const LIMBS: usize> UInt<LIMBS>

pub const fn sbb(&self, rhs: &Self, borrow: Limb) -> (Self, Limb)

Computes a - (b + borrow), returning the result along with the new borrow.

pub const fn wrapping_sub(&self, rhs: &Self) -> Self

Perform wrapping subtraction, discarding underflow and wrapping around the boundary of the type.

pub fn checked_sub(&self, rhs: &Self) -> CtOption<Self>

Perform checked subtraction, returning a CtOption which is_some only if the operation did not overflow.

impl<const LIMBS: usize> UInt<LIMBS>

pub const fn sub_mod(&self, rhs: &UInt<LIMBS>, p: &UInt<LIMBS>) -> UInt<LIMBS>

Computes self - rhs mod p in constant time.

Assumes self and rhs are < p.

impl<const LIMBS: usize> UInt<LIMBS>

pub fn random(rng: impl CryptoRng + RngCore) -> Self

Generate a cryptographically secure random UInt.

pub fn random_mod(rng: impl CryptoRng + RngCore, modulus: &Self) -> Self

Generate a cryptographically secure random UInt which is less than a given modulus.

This function uses rejection sampling, a method which produces an unbiased distribution of in-range values provided the underlying CryptoRng is unbiased, but runs in variable-time.

The variable-time nature of the algorithm should not pose a security issue so long as the underlying random number generator is truly a CryptoRng, where previous outputs are unrelated to subsequent outputs and do not reveal information about the RNG’s internal state.

impl<const LIMBS: usize> UInt<LIMBS>

pub const ZERO: Self = _

The value 0.

pub const ONE: Self = _

The value 1.

pub const MAX: Self = _

Maximum value this UInt can express.

pub const fn new(limbs: [Limb; LIMBS]) -> Self

Const-friendly UInt constructor.

pub const fn limbs(&self) -> &[Limb; LIMBS]

Borrow the limbs of this UInt.

pub const fn into_limbs(self) -> [Limb; LIMBS]

Convert this UInt into its inner limbs.