crypto_bigint

Module uint

Stack-allocated big unsigned integers.

Modules

add 🔒

Uint addition operations.

add_mod 🔒

Uint modular addition operations.

array 🔒

hybrid-array integration with Uint.

bit_and 🔒

Uint bitwise AND operations.

bit_not 🔒

Uint bitwise NOT operations.

bit_or 🔒

Uint bitwise OR operations.

bit_xor 🔒

Uint bitwise XOR operations.

bits 🔒

boxed 🔒

Heap-allocated big unsigned integers.

cmp 🔒

Uint comparison operations.

concat 🔒

ct 🔒

Constant-time support: impls of Ct* traits and constant-time const fn operations.

div 🔒

Uint division operations.

div_limb 🔒

Implementation of constant-time division via reciprocal precomputation, as described in “Improved Division by Invariant Integers” by Niels Möller and Torbjorn Granlund (DOI: 10.1109/TC.2010.143, https://gmplib.org/~tege/division-paper.pdf).

encoding 🔒

Const-friendly decoding/encoding operations for Uint.

from 🔒

From-like conversions for Uint.

gcd 🔒

This module implements Binary (Extended) GCD for Uint.

invert_mod 🔒

lcm 🔒

This module implements Least common multiple (LCM) for Uint.

macros 🔒

Macros used to define trait impls on aliases of Uint.

mod_symbol 🔒

Support for computing modular symbols.

mul 🔒

Uint multiplication operations.

mul_mod 🔒

Uint modular multiplication operations.

mul_signed 🔒

neg 🔒

neg_mod 🔒

Uint modular negation operations.

pow 🔒

Uint exponentiation operations.

rand 🔒

Random number generator support

ref_type 🔒

Unsigned integer reference type.

resize 🔒

Support for converting between Uint instances with different sizes.

root 🔒

Support for nth root calculation for Uint.

shl 🔒

Uint bitwise left shift operations.

shr 🔒

Uint bitwise right shift operations.

split 🔒

sqrt 🔒

Uint square root operations.

sub 🔒

Uint subtraction operations.

sub_mod 🔒

Uint modular subtraction operations.

Structs

Uint

Stack-allocated big unsigned integer.

Type Aliases

U64

64-bit unsigned big integer.

U128

128-bit unsigned big integer.

U192

192-bit unsigned big integer.

U256

256-bit unsigned big integer.

U320

320-bit unsigned big integer.

U384

384-bit unsigned big integer.

U448

448-bit unsigned big integer.

U512

512-bit unsigned big integer.

U576

576-bit unsigned big integer.

U640

640-bit unsigned big integer.

U704

704-bit unsigned big integer.

U768

768-bit unsigned big integer.

U832

832-bit unsigned big integer.

U896

896-bit unsigned big integer.

U960

960-bit unsigned big integer.

U1024

1024-bit unsigned big integer.

U1280

1280-bit unsigned big integer.

U1536

1536-bit unsigned big integer.

U1792

1792-bit unsigned big integer.

U2048

2048-bit unsigned big integer.

U3072

3072-bit unsigned big integer.

U3584

3584-bit unsigned big integer.

U4096

4096-bit unsigned big integer.

U4224

4224-bit unsigned big integer.

U4352

4352-bit unsigned big integer.

U6144

6144-bit unsigned big integer.

U8192

8192-bit unsigned big integer.

U16384

16384-bit unsigned big integer.

U32768

32768-bit unsigned big integer.