crypto_bigint

Module word

Word represents the core integer type we use as the core of Limb, and is typically the same size as a pointer on a particular CPU.

Functions

choice_from_eq 🔒

Returns the truthy value if x == y, and the falsy value otherwise.

choice_from_gt 🔒

Returns the truthy value if x > y, and the falsy value otherwise.

choice_from_le 🔒

Returns the truthy value if x <= y and the falsy value otherwise.

choice_from_lsb 🔒

Returns the truthy value if value == 1, and the falsy value if value == 0.

choice_from_lt 🔒

Returns the truthy value if x < y, and the falsy value otherwise.

choice_from_mask 🔒

Returns the truthy value if value == Word::MAX, and the falsy value if value == 0.

choice_from_msb 🔒

Returns the truthy value if the most significant bit of value is 1, and the falsy value if it equals 0.

choice_from_nz 🔒

Returns the truthy value if value != 0, and the falsy value otherwise.

choice_from_wide_eq 🔒

Returns the truthy value if x == y, and the falsy value otherwise.

choice_from_wide_le 🔒

Returns the truthy value if x <= y and the falsy value otherwise.

choice_from_wide_nz 🔒

Returns the truthy value if value != 0, and the falsy value otherwise.

choice_to_mask 🔒

Create a Word-sized bitmask.

choice_to_wide_mask 🔒

Create a WideWord-sized bitmask.

join 🔒

select 🔒

Return b if self is truthy, otherwise return a.

select_wide 🔒

Return b if self is truthy, otherwise return a.

split_wide 🔒

Type Aliases

WideWord

Wide integer type: double the width of Word.

Word

Unsigned integer type that the Limb newtype wraps.