Struct crypto_bigint::limb::Limb

#[repr(transparent)]
pub struct Limb(pub Inner);

Big integers are represented as an array of smaller CPU word-size integers called “limbs”.

Tuple Fields
0: Inner
Implementations

impl Limb

pub const fn adc(self, rhs: Limb, carry: Limb) -> (Limb, Limb)

Computes self + rhs + carry, returning the result along with the new carry.

pub const fn wrapping_add(&self, rhs: Self) -> Self

Perform wrapping addition, discarding overflow.

pub fn checked_add(&self, rhs: Self) -> CtOption<Self>

Perform checked addition, returning a CtOption which is_some only if the operation did not overflow.

impl Limb

pub const fn bitand(self, rhs: Self) -> Self

Calculates a & b.

impl Limb

pub const fn bitor(self, rhs: Self) -> Self

Calculates a | b.

impl Limb

pub fn is_zero(&self) -> Choice

Is this limb equal to zero?

pub fn is_odd(&self) -> Choice

Is this limb an odd number?

pub fn cmp_vartime(&self, other: &Self) -> Ordering

Perform a comparison of the inner value in variable-time.

Note that the PartialOrd and Ord impls wrap constant-time comparisons using the subtle crate.

pub const fn eq_vartime(&self, other: &Self) -> bool

Performs an equality check in variable-time.

impl Limb

pub const fn from_u8(n: u8) -> Self

Create a Limb from a u8 integer (const-friendly)

pub const fn from_u16(n: u16) -> Self

Create a Limb from a u16 integer (const-friendly)

pub const fn from_u32(n: u32) -> Self

Create a Limb from a u32 integer (const-friendly)

pub const fn from_u64(n: u64) -> Self

Create a Limb from a u64 integer (const-friendly)

impl Limb

pub const fn mac(self, b: Limb, c: Limb, carry: Limb) -> (Limb, Limb)

Computes self + (b * c) + carry, returning the result along with the new carry.

pub const fn wrapping_mul(&self, rhs: Self) -> Self

Perform wrapping multiplication, discarding overflow.

pub fn checked_mul(&self, rhs: Self) -> CtOption<Self>

Perform checked multiplication, returning a CtOption which is_some only if the operation did not overflow.

impl Limb

pub const fn sbb(self, rhs: Limb, borrow: Limb) -> (Limb, Limb)

Computes self - (rhs + borrow), returning the result along with the new borrow.

pub const fn wrapping_sub(&self, rhs: Self) -> Self

Perform wrapping subtraction, discarding underflow and wrapping around the boundary of the type.

pub fn checked_sub(&self, rhs: Self) -> CtOption<Self>

Perform checked subtraction, returning a CtOption which is_some only if the operation did not overflow.

impl Limb

pub fn random(rng: impl CryptoRng + RngCore) -> Self

Generate a random limb

impl Limb

pub const ZERO: Self

The value 0.

pub const ONE: Self

The value 1.

pub const MAX: Self

Maximum value this Limb can express.