crypto_bigint::modular

Struct BoxedMontyForm

pub struct BoxedMontyForm { /* private fields */ }

An integer in Montgomery form represented using heap-allocated limbs.

Implementations

impl BoxedMontyForm

pub fn add(&self, rhs: &Self) -> Self

Adds rhs.

pub fn double(&self) -> Self

Double self.

impl BoxedMontyForm

pub fn invert(&self) -> CtOption<Self>

Computes self^-1 representing the multiplicative inverse of self, i.e. self * self^-1 = 1.

pub fn invert_vartime(&self) -> CtOption<Self>

Computes self^-1 representing the multiplicative inverse of self, i.e. self * self^-1 = 1.

This version is variable-time with respect to the self of self, but constant-time with respect to self’s params.

impl BoxedMontyForm

pub fn lincomb_vartime(products: &[(&Self, &Self)]) -> Self

Calculate the sum of products of pairs (a, b) in products.

This method is variable time only with the value of the modulus. For a modulus with leading zeros, this method is more efficient than a naive sum of products.

This method will panic if products is empty. All terms must be associated with equivalent MontyParams.

impl BoxedMontyForm

pub fn mul(&self, rhs: &Self) -> Self

Multiplies by rhs.

pub fn square(&self) -> Self

Computes the (reduced) square.

impl BoxedMontyForm

pub fn neg(&self) -> Self

Negates the number.

impl BoxedMontyForm

pub fn pow(&self, exponent: &BoxedUint) -> Self

Raises to the exponent power.

pub fn pow_bounded_exp(&self, exponent: &BoxedUint, exponent_bits: u32) -> Self

Raises to the exponent power, with exponent_bits representing the number of (least significant) bits to take into account for the exponent.

NOTE: exponent_bits may be leaked in the time pattern.

impl BoxedMontyForm

pub fn sub(&self, rhs: &Self) -> Self

Subtracts rhs.

impl BoxedMontyForm

pub fn new(integer: BoxedUint, params: BoxedMontyParams) -> Self

Instantiates a new BoxedMontyForm that represents an integer modulo the provided params.

pub fn bits_precision(&self) -> u32

Bits of precision in the modulus.

pub fn retrieve(&self) -> BoxedUint

Retrieves the integer currently encoded in this BoxedMontyForm, guaranteed to be reduced.

pub fn zero(params: BoxedMontyParams) -> Self

Instantiates a new ConstMontyForm that represents zero.

pub fn one(params: BoxedMontyParams) -> Self

Instantiates a new ConstMontyForm that represents 1.

pub fn is_zero(&self) -> Choice

Determine if this value is equal to zero.

Returns

If zero, returns Choice(1). Otherwise, returns Choice(0).

pub fn is_nonzero(&self) -> Choice

Determine if this value is not equal to zero.

Returns

If zero, returns Choice(0). Otherwise, returns Choice(1).

pub fn params(&self) -> &BoxedMontyParams

Returns the parameter struct used to initialize this object.

pub fn as_montgomery(&self) -> &BoxedUint

Access the BoxedMontyForm value in Montgomery form.

pub fn from_montgomery(integer: BoxedUint, params: BoxedMontyParams) -> Self

Create a BoxedMontyForm from a value in Montgomery form.

pub fn to_montgomery(&self) -> BoxedUint

Extract the value from the BoxedMontyForm in Montgomery form.

pub fn div_by_2(&self) -> Self

Performs division by 2, that is returns x such that x + x = self.

pub fn div_by_2_assign(&mut self)

Performs division by 2 inplace, that is finds x such that x + x = self and writes it into self.