crypto_bigint::modular

Struct ConstMontyForm

pub struct ConstMontyForm<MOD: ConstMontyParams<LIMBS>, const LIMBS: usize> { /* private fields */ }

An integer in Montgomery form modulo MOD, represented using LIMBS limbs. The modulus is constant, so it cannot be set at runtime.

Internally, the value is stored in Montgomery form (multiplied by MOD::PARAMS.one) until it is retrieved.

Implementations

impl<MOD: ConstMontyParams<LIMBS>, const LIMBS: usize> ConstMontyForm<MOD, LIMBS>

pub const fn add(&self, rhs: &ConstMontyForm<MOD, LIMBS>) -> Self

Adds rhs.

pub const fn double(&self) -> Self

Double self.

impl<MOD: ConstMontyParams<LIMBS>, const LIMBS: usize> ConstMontyForm<MOD, LIMBS>

pub const fn inv(&self) -> CtOption<Self>

👎Deprecated since 0.7.0:
please use invert instead

Computes self^-1 representing the multiplicative inverse of self, i.e. self * self^-1 = 1.

If the number was invertible, the second element of the tuple is the truthy value, otherwise it is the falsy value (in which case the first element’s value is unspecified).

pub const fn invert(&self) -> CtOption<Self>

Computes self^-1 representing the multiplicative inverse of self, i.e. self * self^-1 = 1.

If the number was invertible, the second element of the tuple is the truthy value, otherwise it is the falsy value (in which case the first element’s value is unspecified).

pub const fn inv_vartime(&self) -> CtOption<Self>

👎Deprecated since 0.7.0:
please use invert_vartime instead

Computes self^-1 representing the multiplicative inverse of self, i.e. self * self^-1 = 1.

If the number was invertible, the second element of the tuple is the truthy value, otherwise it is the falsy value (in which case the first element’s value is unspecified).

This version is variable-time with respect to the value of self, but constant-time with respect to MOD.

pub const fn invert_vartime(&self) -> CtOption<Self>

Computes self^-1 representing the multiplicative inverse of self, i.e. self * self^-1 = 1.

If the number was invertible, the second element of the tuple is the truthy value, otherwise it is the falsy value (in which case the first element’s value is unspecified).

This version is variable-time with respect to the value of self, but constant-time with respect to MOD.

impl<MOD: ConstMontyParams<LIMBS>, const LIMBS: usize> ConstMontyForm<MOD, LIMBS>

pub const fn lincomb(products: &[(Self, Self)]) -> Self

Calculate the sum of products of pairs (a, b) in products.

impl<MOD: ConstMontyParams<LIMBS>, const LIMBS: usize> ConstMontyForm<MOD, LIMBS>

pub const fn jacobi_symbol(&self) -> JacobiSymbol

Compute the Jacobi symbol (self|modulus).

For a prime modulus, this corresponds to the Legendre symbol and indicates whether self is quadratic residue.

pub const fn jacobi_symbol_vartime(&self) -> JacobiSymbol

Compute the Jacobi symbol (self|modulus).

For a prime modulus, this corresponds to the Legendre symbol and indicates whether self is quadratic residue.

This method is variable-time with respect to the value of self.

impl<MOD: ConstMontyParams<LIMBS>, const LIMBS: usize> ConstMontyForm<MOD, LIMBS>

pub const fn mul(&self, rhs: &Self) -> Self

Multiplies by rhs.

pub const fn square(&self) -> Self

Computes the (reduced) square.

pub const fn square_repeat_vartime(&self, n: u32) -> Self

Computes the reduced product self^2n.

This method is variable time in n only.

impl<MOD: ConstMontyParams<LIMBS>, const LIMBS: usize> ConstMontyForm<MOD, LIMBS>

pub const fn neg(&self) -> Self

Negates the number.

impl<MOD: ConstMontyParams<LIMBS>, const LIMBS: usize> ConstMontyForm<MOD, LIMBS>

pub const fn pow<const RHS_LIMBS: usize>( &self, exponent: &Uint<RHS_LIMBS>, ) -> Self

Raises to the exponent power.

pub const fn pow_bounded_exp<const RHS_LIMBS: usize>( &self, exponent: &Uint<RHS_LIMBS>, exponent_bits: u32, ) -> Self

Raises to the exponent power, with exponent_bits representing the number of (least significant) bits to take into account for the exponent.

NOTE: exponent_bits may be leaked in the time pattern.

pub const fn pow_vartime<const RHS_LIMBS: usize>( &self, exponent: &Uint<RHS_LIMBS>, ) -> Self

Raises to the exponent power.

This method is variable time in exponent.

pub fn pow_amm(&self, exponent: &Uint<LIMBS>) -> Self

Raises to the exponent power using Almost Montgomery Multiplication (AMM).

pub fn pow_amm_bounded_exp( &self, exponent: &Uint<LIMBS>, exponent_bits: u32, ) -> Self

Raises to the exponent power using Almost Montgomery Multiplication (AMM) with exponent_bits representing the number of (least significant) bits to take into account for the exponent.

NOTE: exponent_bits may be leaked in the time pattern.

impl<const LIMBS: usize, MOD> ConstMontyForm<MOD, LIMBS>
where MOD: ConstPrimeMontyParams<LIMBS>,

pub const fn sqrt(&self) -> CtOption<Self>

Compute the modular square root for self, if it exists.

impl<MOD: ConstMontyParams<LIMBS>, const LIMBS: usize> ConstMontyForm<MOD, LIMBS>

pub const fn sub(&self, rhs: &Self) -> Self

Subtracts rhs.

impl<MOD: ConstMontyParams<LIMBS>, const LIMBS: usize> ConstMontyForm<MOD, LIMBS>

pub const ZERO: Self

The representation of 0 mod MOD.

pub const ONE: Self

The representation of 1 mod MOD.

pub const MODULUS: Odd<Uint<LIMBS>>

Modulus as an unsigned integer.

pub const fn new(integer: &Uint<LIMBS>) -> Self

Instantiates a new ConstMontyForm that represents this integer mod MOD.

pub const fn retrieve(&self) -> Uint<LIMBS>

Retrieves the integer currently encoded in this ConstMontyForm, guaranteed to be reduced.

pub const fn as_montgomery(&self) -> &Uint<LIMBS>

Access the ConstMontyForm value in Montgomery form.

pub fn as_montgomery_mut(&mut self) -> &mut Uint<LIMBS>

Mutably access the ConstMontyForm value in Montgomery form.

pub const fn from_montgomery(integer: Uint<LIMBS>) -> Self

Create a ConstMontyForm from a value in Montgomery form.

pub const fn to_montgomery(&self) -> Uint<LIMBS>

Extract the value from the ConstMontyForm in Montgomery form.

pub const fn div_by_2(&self) -> Self

Performs division by 2, that is returns x such that x + x = self.