crypto_bigint::modular

Trait ConstPrimeMontyParams

pub trait ConstPrimeMontyParams<const LIMBS: usize>: ConstMontyParams<LIMBS> {
    const PRIME_PARAMS: PrimeParams<LIMBS>;
}

Trait representing a prime modulus and its associated constants for converting in and out of Montgomery form.

To define a type which impls this trait, use the const_prime_monty_params! macro.

Required Associated Constants

const PRIME_PARAMS: PrimeParams<LIMBS>

Prime parameters constant.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.