crypto_bigint

Module prelude

Import prelude for this crate: includes important traits.

Enums

DecodeError

Possible errors in variable-time integer decoding methods.

RandomBitsError

Possible errors of the methods in RandomBits trait.

Traits

AddMod

Compute self + rhs mod p.

BitOps

Bit counting and bit operations.

Bounded

Integers whose representation takes a bounded amount of space.

CheckedAdd

Checked addition.

CheckedDiv

Checked division.

CheckedMul

Checked multiplication.

CheckedSub

Checked subtraction.

Concat

Concatenate two numbers into a “wide” double-width value, using the hi value as the most significant portion of the resulting value.

ConcatMixed

Concatenate two numbers into a “wide” combined-width value, using the hi value as the most significant value.

ConcatenatingMul

Widening multiply: returns a value with a number of limbs equal to the sum of the inputs.

ConstOne

Defines an associated constant representing the multiplicative identity element for Self.

ConstZero

Defines an associated constant representing the additive identity element for Self.

ConstantTimeSelect

Trait for types which are conditionally selectable in constant time.

Constants

Trait for associating constant values with a type.

DivRemLimb

Support for optimized division by a single limb.

DivVartime

Division in variable time.

Encoding

Encoding support.

FixedInteger

Fixed-width Integers.

Gcd

Compute the greatest common divisor of two integers.

Integer

Integer trait: represents common functionality of integer types provided by this crate.

InvModDeprecated

Compute 1 / self mod p.

Invert

Constant-time inversion.

InvertMod

Compute 1 / self mod p.

Monty

A representation of an integer optimized for the performance of modular operations.

MontyMultiplier

Prepared Montgomery multiplier for tight loops.

MulMod

Compute self * rhs mod p.

MultiExponentiate

Performs modular multi-exponentiation using Montgomery’s ladder.

MultiExponentiateBoundedExp

Performs modular multi-exponentiation using Montgomery’s ladder. exponent_bits represents the number of bits to take into account for the exponent.

NegMod

Compute -self mod p.

One

One values: multiplicative identity element for Self.

Pow

Constant-time exponentiation.

PowBoundedExp

Constant-time exponentiation with exponent of a bounded bit size.

Random

Random number generation support.

RandomBits

Random bits generation support.

RandomMod

Modular random number generation support.

Reduce

Modular reduction from a larger value T.

RemLimb

Support for optimized division by a single limb.

RemMixed

Support for calculating the remainder of two differently sized integers.

Resize

Methods for resizing the allocated storage.

ShlVartime

Left shifts, variable time in shift.

ShrVartime

Right shifts, variable time in shift.

Signed

Signed Integers.

Split

Split a number in half, returning the least significant half followed by the most significant.

SplitMixed

Split a number into parts, returning the least significant part followed by the most significant.

Square

Support for optimized squaring

SquareAssign

Support for optimized squaring in-place

SquareMod

Compute self * self mod p.

SquareRoot

Support for calucaling square roots.

SubMod

Compute self - rhs mod p.

Unsigned

Unsigned Integers.

WideningMulDeprecated

Widening multiply: returns a value with a number of limbs equal to the sum of the inputs.

WrappingAdd

Performs addition that wraps around on overflow.

WrappingMul

Performs multiplication that wraps around on overflow.

WrappingNeg

Performs a negation that does not panic.

WrappingShl

Performs a left shift that does not panic.

WrappingShr

Performs a right shift that does not panic.

WrappingSub

Performs subtraction that wraps around on overflow.

Xgcd

Compute the extended greatest common divisor of two integers.

Zero

Zero values: additive identity element for Self.