Module crypto_bigint::prelude

Import prelude for this crate: includes important traits.

Traits

AddMod

Compute self + rhs mod p.

ArrayDecoding

Support for decoding a GenericArray as a big integer.

ArrayEncoding

Support for encoding a big integer as a GenericArray.

CheckedAdd

Checked addition.

CheckedMul

Checked multiplication.

CheckedSub

Checked substraction.

Concat

Concatenate two numbers into a “wide” twice-width value, using the rhs value as the least significant value.

Encoding

Encoding support.

Integer

Integer type.

MulMod

Compute self * rhs mod p.

NegMod

Compute -self mod p.

Random

Random number generation support.

RandomMod

Modular random number generation support.

Split

Split a number in half, returning the most significant half followed by the least significant.

SubMod

Compute self - rhs mod p.

Zero

Zero values.