crypto_bigint::prelude
Trait Concat

pub trait Concat: ConcatMixed<Self, MixedOutput = Self::Output> {
    type Output;

    // Provided method
    fn concat(&self, lo: &Self) -> Self::Output { ... }
}

Concatenate two numbers into a “wide” double-width value, using the lo value as the least significant value.

Required Associated Types

type Output

Concatenated output: twice the width of Self.

Provided Methods

fn concat(&self, lo: &Self) -> Self::Output

Concatenate the two halves, with self as most significant and lo as the least significant.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.