crypto_bigint::prelude

Trait RandomMod

pub trait RandomMod: Sized + Zero {
    // Required method
    fn try_random_mod<R: TryRngCore + ?Sized>(
        rng: &mut R,
        modulus: &NonZero<Self>,
    ) -> Result<Self, R::Error>;

    // Provided method
    fn random_mod<R: RngCore + ?Sized>(
        rng: &mut R,
        modulus: &NonZero<Self>,
    ) -> Self { ... }
}

Modular random number generation support.

Required Methods

fn try_random_mod<R: TryRngCore + ?Sized>( rng: &mut R, modulus: &NonZero<Self>, ) -> Result<Self, R::Error>

Generate a random number which is less than a given modulus.

This uses rejection sampling.

As a result, it runs in variable time that depends in part on modulus. If the generator rng is cryptographically secure (for example, it implements CryptoRng), then this is guaranteed not to leak anything about the output value aside from it being less than modulus.

Provided Methods

fn random_mod<R: RngCore + ?Sized>(rng: &mut R, modulus: &NonZero<Self>) -> Self

Generate a random number which is less than a given modulus.

This uses rejection sampling.

As a result, it runs in variable time that depends in part on modulus. If the generator rng is cryptographically secure (for example, it implements CryptoRng), then this is guaranteed not to leak anything about the output value aside from it being less than modulus.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.