Struct crypto_bigint::Limb

#[repr(transparent)]
pub struct Limb(pub LimbUInt);

Big integers are represented as an array of smaller CPU word-size integers called “limbs”.

Tuple Fields
0: LimbUInt
Implementations

impl Limb

pub const fn adc(self, rhs: Limb, carry: Limb) -> (Limb, Limb)

Computes self + rhs + carry, returning the result along with the new carry.

pub const fn saturating_add(&self, rhs: Self) -> Self

Perform saturating addition.

pub const fn wrapping_add(&self, rhs: Self) -> Self

Perform wrapping addition, discarding overflow.

impl Limb

pub const fn bitand(self, rhs: Self) -> Self

Calculates a & b.

impl Limb

pub const fn not(self) -> Self

Calculates !a.

impl Limb

pub const fn bitor(self, rhs: Self) -> Self

Calculates a | b.

impl Limb

pub const fn bitxor(self, rhs: Self) -> Self

Calculates a ^ b.

impl Limb

pub const fn bits(self) -> usize

Calculate the number of bits needed to represent this number.

impl Limb

pub fn is_odd(&self) -> Choice

Is this limb an odd number?

pub fn cmp_vartime(&self, other: &Self) -> Ordering

Perform a comparison of the inner value in variable-time.

Note that the PartialOrd and Ord impls wrap constant-time comparisons using the subtle crate.

pub const fn eq_vartime(&self, other: &Self) -> bool

Performs an equality check in variable-time.

impl Limb

pub const fn from_u8(n: u8) -> Self

Create a Limb from a u8 integer (const-friendly)

pub const fn from_u16(n: u16) -> Self

Create a Limb from a u16 integer (const-friendly)

pub const fn from_u32(n: u32) -> Self

Create a Limb from a u32 integer (const-friendly)

pub const fn from_u64(n: u64) -> Self

Create a Limb from a u64 integer (const-friendly)

impl Limb

pub const fn mac(self, b: Limb, c: Limb, carry: Limb) -> (Limb, Limb)

Computes self + (b * c) + carry, returning the result along with the new carry.

pub const fn saturating_mul(&self, rhs: Self) -> Self

Perform saturating multiplication.

pub const fn wrapping_mul(&self, rhs: Self) -> Self

Perform wrapping multiplication, discarding overflow.

impl Limb

pub const fn sbb(self, rhs: Limb, borrow: Limb) -> (Limb, Limb)

Computes self - (rhs + borrow), returning the result along with the new borrow.

pub const fn saturating_sub(&self, rhs: Self) -> Self

Perform saturating subtraction.

pub const fn wrapping_sub(&self, rhs: Self) -> Self

Perform wrapping subtraction, discarding underflow and wrapping around the boundary of the type.

impl Limb

pub const ZERO: Self = _

The value 0.

pub const ONE: Self = _

The value 1.

pub const MAX: Self = _

Maximum value this Limb can express.

pub const BIT_SIZE: usize = 64usize

Size of the inner integer in bits.

pub const BYTE_SIZE: usize = 8usize

Size of the inner integer in bytes.