crypto_bigint

Struct Odd

pub struct Odd<T: ?Sized>(/* private fields */);

Wrapper type for odd integers.

These are frequently used in cryptography, e.g. as a modulus.

Implementations

impl<const LIMBS: usize> Odd<Int<LIMBS>>

pub const fn gcd_uint(&self, rhs: &Uint<LIMBS>) -> Odd<Uint<LIMBS>>

Compute the greatest common divisor of self and rhs.

pub const fn gcd_uint_vartime(&self, rhs: &Uint<LIMBS>) -> OddUint<LIMBS>

Compute the greatest common divisor of self and rhs.

Executes in variable time w.r.t. all input parameters.

pub const fn xgcd( &self, rhs: &NonZero<Int<LIMBS>>, ) -> XgcdOutput<LIMBS, OddUint<LIMBS>>

Execute the Extended GCD algorithm.

Given (self, rhs), computes (g, x, y) s.t. self * x + rhs * y = g = gcd(self, rhs).

impl<T> Odd<T>

pub fn new(n: T) -> CtOption<Self>
where T: Integer,

Create a new odd integer.

pub fn get(self) -> T

Returns the inner value.

impl<T: ?Sized> Odd<T>

pub const fn as_ref(&self) -> &T

Provides access to the contents of Odd in a const context.

pub const fn as_nz_ref(&self) -> &NonZero<T>

All odd integers are definitionally non-zero, so we can also obtain a reference to NonZero.

impl<T> Odd<T>
where T: Bounded + ?Sized,

pub const BITS: u32 = T::BITS

Total size of the represented integer in bits.

pub const BYTES: usize = T::BYTES

Total size of the represented integer in bytes.

impl<const LIMBS: usize> Odd<Uint<LIMBS>>

pub const fn from_be_hex(hex: &str) -> Self

Create a new Odd<Uint<LIMBS>> from the provided big endian hex string.

Panics if the hex is malformed or not zero-padded accordingly for the size, or if the value is even.

pub const fn from_le_hex(hex: &str) -> Self

Create a new Odd<Uint<LIMBS>> from the provided little endian hex string.

Panics if the hex is malformed or not zero-padded accordingly for the size, or if the value is even.

pub const fn resize<const T: usize>(&self) -> Odd<Uint<T>>

Construct an Odd<Uint<T>> from the unsigned integer value, truncating the upper bits if the value is too large to be represented.

impl<const LIMBS: usize> Odd<Int<LIMBS>>

pub const fn abs_sign(&self) -> (Odd<Uint<LIMBS>>, ConstChoice)

The sign and magnitude of this Odd<Int<{LIMBS}>>.

pub const fn abs(&self) -> Odd<Uint<LIMBS>>

The magnitude of this Odd<Int<{LIMBS}>>.

impl Odd<UintRef>

pub const fn to_uint_resize<const T: usize>(&self) -> Odd<Uint<T>>

Construct an Odd<Uint<T>> from the unsigned integer value, truncating the upper bits if the value is too large to be represented.

impl Odd<BoxedUint>

pub fn random<R: TryRngCore + ?Sized>(rng: &mut R, bit_length: u32) -> Self

Generate a random Odd<Uint<T>>.

impl<const LIMBS: usize> Odd<Uint<LIMBS>>

pub const fn gcd_uint(&self, rhs: &Uint<LIMBS>) -> Self

Compute the greatest common divisor of self and rhs.

pub const fn gcd_uint_vartime(&self, rhs: &Uint<LIMBS>) -> Self

Compute the greatest common divisor of self and rhs.

Executes in variable time w.r.t. all input parameters.

pub const fn xgcd(&self, rhs: &Self) -> XgcdOutput<LIMBS, OddUint<LIMBS>>

Execute the Extended GCD algorithm.

Given (self, rhs), computes (g, x, y) s.t. self * x + rhs * y = g = gcd(self, rhs).

impl Odd<UintRef>

pub const fn invert_mod_u64(&self) -> u64

Returns the multiplicative inverse of the argument modulo 2^64. The implementation is based on the Hurchalla’s method for computing the multiplicative inverse modulo a power of two.

For better understanding the implementation, the following paper is recommended: J. Hurchalla, “An Improved Integer Multiplicative Inverse (modulo 2^w)”, https://arxiv.org/pdf/2204.04342.pdf

Variable time with respect to the number of words in value, however that number will be fixed for a given integer size.