Struct crypto_bigint::UInt

pub struct UInt<const LIMBS: usize> { /* fields omitted */ }

Big unsigned integer.

Generic over the given number of LIMBS

Encoding support

This type supports many different types of encodings, either via the Encoding trait or various const fn decoding and encoding functions that can be used with UInt constants.

Optional crate features for encoding (off-by-default):

generic-array: enables ArrayEncoding trait which can be used to UInt as GenericArray<u8, N> and a ArrayDecoding trait which can be used to GenericArray<u8, N> as UInt.

rlp: support for Recursive Length Prefix (RLP) encoding.

Implementations

impl<const LIMBS: usize> UInt<LIMBS>

pub const fn adc(&self, rhs: &Self, carry: Limb) -> (Self, Limb)

Computes a + b + carry, returning the result along with the new carry.

pub const fn saturating_add(&self, rhs: &Self) -> Self

Perform saturating addition, returning MAX on overflow.

pub const fn wrapping_add(&self, rhs: &Self) -> Self

Perform wrapping addition, discarding overflow.

impl<const LIMBS: usize> UInt<LIMBS>

pub const fn add_mod(&self, rhs: &UInt<LIMBS>, p: &UInt<LIMBS>) -> UInt<LIMBS>

Computes self + rhs mod p in constant time.

Assumes self and rhs are < p.

impl<const LIMBS: usize> UInt<LIMBS>

pub const fn bitand(&self, rhs: &Self) -> Self

Computes bitwise a & b.

pub const fn wrapping_and(&self, rhs: &Self) -> Self

Perform wrapping bitwise AND.

There’s no way wrapping could ever happen. This function exists so that all operations are accounted for in the wrapping operations

pub fn checked_and(&self, rhs: &Self) -> CtOption<Self>

Perform checked bitwise AND, returning a CtOption which is_some always

impl<const LIMBS: usize> UInt<LIMBS>

pub const fn not(&self) -> Self

Computes bitwise !a.

impl<const LIMBS: usize> UInt<LIMBS>

pub const fn bitor(&self, rhs: &Self) -> Self

Computes bitwise a & b.

pub const fn wrapping_or(&self, rhs: &Self) -> Self

Perform wrapping bitwise OR.

There’s no way wrapping could ever happen. This function exists so that all operations are accounted for in the wrapping operations

pub fn checked_or(&self, rhs: &Self) -> CtOption<Self>

Perform checked bitwise OR, returning a CtOption which is_some always

impl<const LIMBS: usize> UInt<LIMBS>

pub const fn bitxor(&self, rhs: &Self) -> Self

Computes bitwise a ^ b.

pub const fn wrapping_xor(&self, rhs: &Self) -> Self

Perform wrapping bitwise `XOR``.

There’s no way wrapping could ever happen. This function exists so that all operations are accounted for in the wrapping operations

pub fn checked_xor(&self, rhs: &Self) -> CtOption<Self>

Perform checked bitwise XOR, returning a CtOption which is_some always

impl<const LIMBS: usize> UInt<LIMBS>

pub const fn bits(self) -> usize

Calculate the number of bits needed to represent this number.

impl<const LIMBS: usize> UInt<LIMBS>

pub fn div_rem(&self, rhs: &Self) -> CtOption<(Self, Self)>

Computes self / rhs, returns the quotient, remainder if rhs != 0

pub fn reduce(&self, rhs: &Self) -> CtOption<Self>

Computes self % rhs, returns the remainder if rhs != 0

pub const fn wrapping_div(&self, rhs: &Self) -> Self

Wrapped division is just normal division i.e. self / rhs There’s no way wrapping could ever happen. This function exists, so that all operations are accounted for in the wrapping operations.

pub fn checked_div(&self, rhs: &Self) -> CtOption<Self>

Perform checked division, returning a CtOption which is_some only if the rhs != 0

pub const fn wrapping_rem(&self, rhs: &Self) -> Self

Wrapped (modular) remainder calculation is just self % rhs. There’s no way wrapping could ever happen. This function exists, so that all operations are accounted for in the wrapping operations.

pub fn checked_rem(&self, rhs: &Self) -> CtOption<Self>

Perform checked reduction, returning a CtOption which is_some only if the rhs != 0

impl<const LIMBS: usize> UInt<LIMBS>

pub const fn from_be_slice(bytes: &[u8]) -> Self

Create a new UInt from the provided big endian bytes.

pub const fn from_be_hex(hex: &str) -> Self

Create a new UInt from the provided big endian hex string.

pub const fn from_le_slice(bytes: &[u8]) -> Self

Create a new UInt from the provided little endian bytes.

pub const fn from_le_hex(hex: &str) -> Self

Create a new UInt from the provided little endian hex string.

impl<const LIMBS: usize> UInt<LIMBS>

pub const fn from_u8(n: u8) -> Self

Create a UInt from a u8 (const-friendly)

pub const fn from_u16(n: u16) -> Self

Create a UInt from a u16 (const-friendly)

pub const fn from_u32(n: u32) -> Self

Create a UInt from a u32 (const-friendly)

pub const fn from_u64(n: u64) -> Self

Create a UInt from a u64 (const-friendly)

pub const fn from_u128(n: u128) -> Self

Create a UInt from a u128 (const-friendly)

pub const fn from_uint_array(arr: [LimbUInt; LIMBS]) -> Self

Create a UInt from an array of the LimbUInt unsigned integer type.

pub const fn to_uint_array(self) -> [LimbUInt; LIMBS]

Create an array of LimbUInt unsigned integers from a UInt.

impl<const LIMBS: usize> UInt<LIMBS>

pub const fn mul_wide(&self, rhs: &Self) -> (Self, Self)

Compute “wide” multiplication, with a product twice the size of the input.

Returns a tuple containing the (lo, hi) components of the product.

Ordering note

Releases of crypto-bigint prior to v0.3 used (hi, lo) ordering instead. This has been changed for better consistency with the rest of the APIs in this crate.

For more info see: https://github.com/RustCrypto/crypto-bigint/issues/4

pub const fn saturating_mul(&self, rhs: &Self) -> Self

Perform saturating multiplication, returning MAX on overflow.

pub const fn wrapping_mul(&self, rhs: &Self) -> Self

Perform wrapping multiplication, discarding overflow.

pub fn square(&self) -> <Self as Concat>::Output where
Self: Concat,

Square self, returning a “wide” result.

impl<const LIMBS: usize> UInt<LIMBS>

pub const fn neg_mod(&self, p: &Self) -> Self

Computes -a mod p in constant time.

impl<const LIMBS: usize> UInt<LIMBS>

pub const fn shl_vartime(&self, n: usize) -> Self

Computes self << shift.

NOTE: this operation is variable time with respect to n ONLY.

When used with a fixed n, this function is constant-time with respect to self.

impl<const LIMBS: usize> UInt<LIMBS>

pub const fn shr_vartime(&self, shift: usize) -> Self

Computes self >> n.

NOTE: this operation is variable time with respect to n ONLY.

When used with a fixed n, this function is constant-time with respect to self.

impl<const LIMBS: usize> UInt<LIMBS>

pub const fn sqrt(&self) -> Self

Computes √(self) Uses Brent & Zimmermann, Modern Computer Arithmetic, v0.5.9, Algorithm 1.13

Callers can check if self is a square by squaring the result

pub const fn wrapping_sqrt(&self) -> Self

Wrapped sqrt is just normal √(self) There’s no way wrapping could ever happen. This function exists, so that all operations are accounted for in the wrapping operations.

pub fn checked_sqrt(&self) -> CtOption<Self>

Perform checked sqrt, returning a CtOption which is_some only if the √(self)² == self

impl<const LIMBS: usize> UInt<LIMBS>

pub const fn sbb(&self, rhs: &Self, borrow: Limb) -> (Self, Limb)

Computes a - (b + borrow), returning the result along with the new borrow.

pub const fn saturating_sub(&self, rhs: &Self) -> Self

Perform saturating subtraction, returning ZERO on underflow.

pub const fn wrapping_sub(&self, rhs: &Self) -> Self

Perform wrapping subtraction, discarding underflow and wrapping around the boundary of the type.

impl<const LIMBS: usize> UInt<LIMBS>

pub const fn sub_mod(&self, rhs: &UInt<LIMBS>, p: &UInt<LIMBS>) -> UInt<LIMBS>

Computes self - rhs mod p in constant time.

Assumes self and rhs are < p.

impl<const LIMBS: usize> UInt<LIMBS>

pub const ZERO: Self

The value 0.

pub const ONE: Self

The value 1.

pub const MAX: Self

Maximum value this UInt can express.

pub const fn new(limbs: [Limb; LIMBS]) -> Self

Const-friendly UInt constructor.

pub const fn limbs(&self) -> &[Limb; LIMBS]

Borrow the limbs of this UInt.

pub const fn into_limbs(self) -> [Limb; LIMBS]

Convert this UInt into its inner limbs.