crypto_bigint

Struct Uint

pub struct Uint<const LIMBS: usize> { /* private fields */ }

Stack-allocated big unsigned integer.

Generic over the given number of LIMBS

Encoding support

This type supports many different types of encodings, either via the Encoding trait or various const fn decoding and encoding functions that can be used with Uint constants.

Optional crate features for encoding (off-by-default):

hybrid-array: enables ArrayEncoding trait which can be used to Uint as Array<u8, N> and a ArrayDecoding trait which can be used to Array<u8, N> as Uint.

rlp: support for Recursive Length Prefix (RLP) encoding.

Implementations

impl<const LIMBS: usize> Uint<LIMBS>

pub const fn adc(&self, rhs: &Self, carry: Limb) -> (Self, Limb)

👎Deprecated since 0.7.0:
please use carrying_add instead

Computes self + rhs + carry, returning the result along with the new carry.

pub const fn carrying_add(&self, rhs: &Self, carry: Limb) -> (Self, Limb)

Computes self + rhs + carry, returning the result along with the new carry.

pub const fn saturating_add(&self, rhs: &Self) -> Self

Perform saturating addition, returning MAX on overflow.

pub const fn wrapping_add(&self, rhs: &Self) -> Self

Perform wrapping addition, discarding overflow.

impl<const LIMBS: usize> Uint<LIMBS>

pub const fn add_mod(&self, rhs: &Self, p: &NonZero<Self>) -> Self

Computes self + rhs mod p.

Assumes self + rhs as unbounded integer is < 2p.

pub const fn add_mod_special(&self, rhs: &Self, c: Limb) -> Self

Computes self + rhs mod p for the special modulus p = MAX+1-c where c is small enough to fit in a single Limb.

Assumes self + rhs as unbounded integer is < 2p.

pub const fn double_mod(&self, p: &NonZero<Self>) -> Self

Computes self + self mod p.

Assumes self as unbounded integer is < p.

impl<const LIMBS: usize> Uint<LIMBS>

pub const fn bitand(&self, rhs: &Self) -> Self

Computes bitwise a & b.

pub const fn bitand_limb(&self, rhs: Limb) -> Self

Perform bitwise AND between self and the given Limb, performing the AND operation on every limb of self.

pub const fn wrapping_and(&self, rhs: &Self) -> Self

Perform wrapping bitwise AND.

There’s no way wrapping could ever happen. This function exists so that all operations are accounted for in the wrapping operations

pub const fn checked_and(&self, rhs: &Self) -> CtOption<Self>

Perform checked bitwise AND, returning a CtOption which is_some always

impl<const LIMBS: usize> Uint<LIMBS>

pub const fn not(&self) -> Self

Computes bitwise !a.

impl<const LIMBS: usize> Uint<LIMBS>

pub const fn bitor(&self, rhs: &Self) -> Self

Computes bitwise a | b.

pub const fn wrapping_or(&self, rhs: &Self) -> Self

Perform wrapping bitwise OR.

There’s no way wrapping could ever happen. This function exists so that all operations are accounted for in the wrapping operations

pub const fn checked_or(&self, rhs: &Self) -> CtOption<Self>

Perform checked bitwise OR, returning a CtOption which is_some always

impl<const LIMBS: usize> Uint<LIMBS>

pub const fn bitxor(&self, rhs: &Self) -> Self

Computes bitwise a ^ b.

pub const fn wrapping_xor(&self, rhs: &Self) -> Self

Perform wrapping bitwise XOR.

There’s no way wrapping could ever happen. This function exists so that all operations are accounted for in the wrapping operations

pub const fn checked_xor(&self, rhs: &Self) -> CtOption<Self>

Perform checked bitwise XOR, returning a CtOption which is_some always

impl<const LIMBS: usize> Uint<LIMBS>

pub const fn bit(&self, index: u32) -> Choice

Get the value of the bit at position index, as a truthy or falsy Choice. Returns the falsy value for indices out of range.

pub const fn bit_vartime(&self, index: u32) -> bool

Returns true if the bit at position index is set, false otherwise.

Remarks

This operation is variable time with respect to index only.

pub const fn bits(&self) -> u32

Calculate the number of bits needed to represent this number.

pub const fn bits_vartime(&self) -> u32

Calculate the number of bits needed to represent this number in variable-time with respect to self.

pub const fn leading_zeros(&self) -> u32

Calculate the number of leading zeros in the binary representation of this number.

pub const fn leading_zeros_vartime(&self) -> u32

Calculate the number of leading zeros in the binary representation of this number in variable-time with respect to self.

pub const fn trailing_zeros(&self) -> u32

Calculate the number of trailing zeros in the binary representation of this number.

pub const fn trailing_zeros_vartime(&self) -> u32

Calculate the number of trailing zeros in the binary representation of this number in variable-time with respect to self.

pub const fn trailing_ones(&self) -> u32

Calculate the number of trailing ones in the binary representation of this number.

pub const fn trailing_ones_vartime(&self) -> u32

Calculate the number of trailing ones in the binary representation of this number, variable time in self.

impl<const LIMBS: usize> Uint<LIMBS>

pub const fn is_nonzero(&self) -> Choice

Returns Choice::TRUE if self != 0 or Choice::FALSE otherwise.

pub const fn is_zero_vartime(&self) -> bool

Determine in variable time whether the self is zero.

pub const fn is_odd(&self) -> Choice

Returns the truthy value if self is odd or the falsy value otherwise.

pub const fn cmp_vartime(&self, rhs: &Self) -> Ordering

Returns the Ordering between self and rhs in variable time.

impl<const LIMBS: usize> Uint<LIMBS>

pub const fn concat<const WIDE_LIMBS: usize>( &self, hi: &Self, ) -> Uint<WIDE_LIMBS>
where Self: Concat<LIMBS, Output = Uint<WIDE_LIMBS>>,

Concatenate the two values, with self as least significant and hi as the most significant, with both values having the same size.

pub const fn concat_mixed<const HI_LIMBS: usize, const WIDE_LIMBS: usize>( &self, hi: &Uint<HI_LIMBS>, ) -> Uint<WIDE_LIMBS>
where Self: Concat<HI_LIMBS, Output = Uint<WIDE_LIMBS>>,

Concatenate the two values, with self as least significant and hi as the most significant.

pub const fn concat_checked<const HI_LIMBS: usize, const WIDE_LIMBS: usize>( &self, hi: &Uint<HI_LIMBS>, ) -> Option<Uint<WIDE_LIMBS>>

Concatenate the two values, with self as least significant and hi as the most significant. If WIDE_LIMBS is not equal to the sum of LIMBS and HI_LIMBS, then None is returned.

pub const fn concat_resize<const HI_LIMBS: usize, const WIDE_LIMBS: usize>( &self, hi: &Uint<HI_LIMBS>, ) -> Uint<WIDE_LIMBS>

Concatenate the two values, with self as least significant and hi as the most significant. The resulting wide integer may be truncated or extended with zeros depending upon whether its size is less than or greater than the sum of LIMBS and HI_LIMBS.

impl<const LIMBS: usize> Uint<LIMBS>

pub const fn div_rem_limb_with_reciprocal( &self, reciprocal: &Reciprocal, ) -> (Self, Limb)

Computes self / rhs using a pre-made reciprocal, returning the quotient and remainder.

pub const fn div_rem_limb(&self, rhs: NonZero<Limb>) -> (Self, Limb)

Computes self / rhs, returning the quotient and remainder.

pub const fn rem_limb_with_reciprocal(&self, reciprocal: &Reciprocal) -> Limb

Computes self % rhs using a pre-made reciprocal.

pub const fn rem_limb(&self, rhs: NonZero<Limb>) -> Limb

Computes self % rhs for a Limb-sized divisor.

pub const fn div_rem<const RHS_LIMBS: usize>( &self, rhs: &NonZero<Uint<RHS_LIMBS>>, ) -> (Self, Uint<RHS_LIMBS>)

Computes self / rhs, returning the quotient and the remainder.

This function is constant-time with respect to both self and rhs.

pub const fn div_rem_vartime<const RHS_LIMBS: usize>( &self, rhs: &NonZero<Uint<RHS_LIMBS>>, ) -> (Self, Uint<RHS_LIMBS>)

Computes self / rhs, returning the quotient and the remainder.

This is variable-time only with respect to rhs.

When used with a fixed rhs, this function is constant-time with respect to self.

pub const fn div_exact<const RHS_LIMBS: usize>( &self, rhs: &NonZero<Uint<RHS_LIMBS>>, ) -> CtOption<Self>

Exactly divides self by rhs, returning CtOption::none() if self is not divisible by rhs.

pub const fn div_exact_vartime<const RHS_LIMBS: usize>( &self, rhs: &NonZero<Uint<RHS_LIMBS>>, ) -> CtOption<Self>

Exactly divides self by rhs, returning CtOption::none() if self is not divisible by rhs.

This is variable-time only with respect to rhs.

When used with a fixed rhs, this function is constant-time with respect to self.

pub const fn rem<const RHS_LIMBS: usize>( &self, rhs: &NonZero<Uint<RHS_LIMBS>>, ) -> Uint<RHS_LIMBS>

Computes self % rhs.

pub const fn rem_vartime<const RHS_LIMBS: usize>( &self, rhs: &NonZero<Uint<RHS_LIMBS>>, ) -> Uint<RHS_LIMBS>

Computes self % rhs in variable-time with respect to rhs.

When used with a fixed rhs, this function is constant-time with respect to self.

pub const fn rem_wide(lower_upper: (Self, Self), rhs: &NonZero<Self>) -> Self

Computes self % rhs for a double-width Uint.

pub const fn rem_wide_vartime( lower_upper: (Self, Self), rhs: &NonZero<Self>, ) -> Self

Computes self % rhs.

This is variable-time only with respect to rhs.

When used with a fixed rhs, this function is constant-time with respect to self.

pub const fn rem2k_vartime(&self, k: u32) -> Self

Computes self % 2^k. Faster than reduce since its a power of 2. Limited to 2^16-1 since Uint doesn’t support higher.

Usage:

use crypto_bigint::{U448, Limb};

let a = U448::from(10_u64);
let k = 3; // 2^3 = 8
let remainder = a.rem2k_vartime(k);

// As 10 % 8 = 2
assert_eq!(remainder, U448::from(2_u64));

pub const fn wrapping_div(&self, rhs: &NonZero<Self>) -> Self

Wrapped division is just normal division i.e. self / rhs.

There’s no way wrapping could ever happen. This function exists, so that all operations are accounted for in the wrapping operations.

pub const fn wrapping_div_vartime<const RHS: usize>( &self, rhs: &NonZero<Uint<RHS>>, ) -> Self

Wrapped division is just normal division i.e. self / rhs.

There’s no way wrapping could ever happen. This function exists, so that all operations are accounted for in the wrapping operations.

pub fn checked_div<const RHS_LIMBS: usize>( &self, rhs: &Uint<RHS_LIMBS>, ) -> CtOption<Self>

Perform checked division, returning a CtOption which is_some only if the rhs != 0.

Usage:

use crypto_bigint::{U448, NonZero};

let a = U448::from(8_u64);
let result = NonZero::new(U448::from(4_u64))
    .map(|b| a.div_rem(&b))
    .expect("Division by zero");

assert_eq!(result.0, U448::from(2_u64));

// Check division by zero
let zero = U448::from(0_u64);
assert!(a.checked_div(&zero).is_none().to_bool(), "should be None for division by zero");

pub const fn wrapping_rem_vartime(&self, rhs: &Self) -> Self

This function exists, so that all operations are accounted for in the wrapping operations.

Panics

if rhs == 0.

Usage:

use crypto_bigint::U448;

let a = U448::from(10_u64);
let b = U448::from(3_u64);
let remainder = a.wrapping_rem_vartime(&b);

assert_eq!(remainder, U448::from(1_u64));

pub fn checked_rem<const RHS_LIMBS: usize>( &self, rhs: &Uint<RHS_LIMBS>, ) -> CtOption<Uint<RHS_LIMBS>>

Perform checked reduction, returning a CtOption which is_some only if the rhs != 0

Usage:

use crypto_bigint::{U448, NonZero};

let a = U448::from(10_u64);
let remainder_option = NonZero::new(U448::from(3_u64))
    .map(|b| a.rem(&b));

assert!(bool::from(remainder_option.is_some()));

// Check reduction by zero
let zero = U448::from(0_u64);

assert!(a.checked_rem(&zero).is_none().to_bool(), "should be None for reduction by zero");

impl<const LIMBS: usize> Uint<LIMBS>

pub const fn from_be_slice(bytes: &[u8]) -> Self

Decode Uint from the provided big endian bytes.

Panics

If the supplied byte slice is not equal to the byte length of this Uint, i.e. Limb::BYTES * LIMBS.

pub const fn from_le_slice(bytes: &[u8]) -> Self

Decode Uint from the provided little endian bytes.

Panics

If the supplied byte slice is not equal to the byte length of this Uint, i.e. Limb::BYTES * LIMBS.

pub fn from_be_slice_truncated(bytes: &[u8], bits_precision: u32) -> Self

Decode Uint from the provided big endian bytes, zero padding if necessary, and truncating to the least significant bytes in the event the given amount of data exceeds bits_precision.

Panics

If bits_precision > Self::BITS.

pub fn from_le_slice_truncated(bytes: &[u8], bits_precision: u32) -> Self

Decode Uint from the provided little endian bytes, zero padding if necessary, and truncating to the least significant bytes in the event the given amount of data exceeds bits_precision.

Panics

If bits_precision > Self::BITS.

pub const fn from_slice(bytes: &[u8], byte_order: ByteOrder) -> Self

Decode Uint from the provided slice, using the supplied ByteOrder to determine the endianness.

Panics

If the supplied byte slice is not equal to the byte length of this Uint, i.e. Limb::BYTES * LIMBS.

pub fn from_slice_truncated( bytes: &[u8], bits_precision: u32, byte_order: ByteOrder, ) -> Self

Decode Uint from the provided bytes, using the supplied ByteOrder to determine the endianness.

Zero pads if necessary, and truncates to the least significant bytes in the event the given amount of data exceeds bits_precision.

Panics

If bits_precision > Self::BITS.

pub const fn from_be_hex(hex: &str) -> Self

Create a new Uint from the provided big endian hex string.

Panics

If the hex is malformed or not zero-padded accordingly for the size.

pub const fn from_le_hex(hex: &str) -> Self

Create a new Uint from the provided little endian hex string.

Panics

If the hex is malformed or not zero-padded accordingly for the size.

pub const fn from_hex(hex: &str, byte_order: ByteOrder) -> Self

Create a new Uint from the provided hex string, using the supplied ByteOrder to determine the endianness.

Panics

if the hex is malformed or not zero-padded accordingly for the size.

pub fn from_str_radix_vartime( src: &str, radix: u32, ) -> Result<Self, DecodeError>

Create a new Uint from a string slice in a given base.

The string may begin with a + character, and may use underscore characters to separate digits.

Errors

Returns DecodeError::InputSize if the size of the decoded integer is larger than this type can represent

Returns DecodeError::InvalidDigit if the input value contains non-digit characters or digits outside of the range 0..radix.

Panics

if radix is not in the range from 2 to 36.

pub const fn to_be_bytes(&self) -> EncodedUint<LIMBS>

Serialize as big endian bytes.

pub const fn to_le_bytes(&self) -> EncodedUint<LIMBS>

Serialize as little endian bytes.

impl<const LIMBS: usize> Uint<LIMBS>

pub const fn from_u8(n: u8) -> Self

Create a Uint from a u8 (const-friendly)

pub const fn from_u16(n: u16) -> Self

Create a Uint from a u16 (const-friendly)

pub const fn from_u32(n: u32) -> Self

Create a Uint from a u32 (const-friendly)

pub const fn from_u64(n: u64) -> Self

Create a Uint from a u64 (const-friendly)

pub const fn from_u128(n: u128) -> Self

Create a Uint from a u128 (const-friendly)

pub const fn from_word(n: Word) -> Self

Create a Uint from a Word (const-friendly)

pub const fn from_wide_word(n: WideWord) -> Self

Create a Uint from a WideWord (const-friendly)

impl<const LIMBS: usize> Uint<LIMBS>

pub const fn gcd(&self, rhs: &Self) -> Self

Compute the greatest common divisor of self and rhs.

pub const fn gcd_vartime(&self, rhs: &Self) -> Self

Compute the greatest common divisor of self and rhs.

Executes in variable time w.r.t. all input parameters.

pub const fn xgcd(&self, rhs: &Self) -> XgcdOutput<LIMBS, Uint<LIMBS>>

Executes the Extended GCD algorithm.

Given (self, rhs), computes (g, x, y), s.t. self * x + rhs * y = g = gcd(self, rhs).

impl<const LIMBS: usize> Uint<LIMBS>

pub const fn inv_mod2k_vartime(&self, k: u32) -> CtOption<Self>

👎Deprecated since 0.7.0:
please use invert_mod2k_vartime instead

Computes 1/self mod 2^k. This method is constant-time w.r.t. self but not k.

If the inverse does not exist (k > 0 and self is even), returns Choice::FALSE as the second element of the tuple, otherwise returns Choice::TRUE.

pub const fn invert_mod2k_vartime(&self, k: u32) -> CtOption<Self>

Computes 1/self mod 2^k. This method is constant-time w.r.t. self but not k.

If the inverse does not exist (k > 0 and self is even, or k > Self::BITS), returns CtOption::none, otherwise returns CtOption::some.

pub const fn inv_mod2k(&self, k: u32) -> CtOption<Self>

👎Deprecated since 0.7.0:
please use invert_mod2k instead

Computes 1/self mod 2^k.

If the inverse does not exist (k > 0 and self is even, k > Self::BITS), returns CtOption::none, otherwise returns CtOption::some.

pub const fn invert_mod2k(&self, k: u32) -> CtOption<Self>

Computes 1/self mod 2^k.

If the inverse does not exist (k > 0 and self is even, or k > Self::BITS), returns CtOption::none, otherwise returns CtOption::some.

pub const fn inv_odd_mod(&self, modulus: &Odd<Self>) -> CtOption<Self>

👎Deprecated since 0.7.0:
please use invert_odd_mod instead

Computes the multiplicative inverse of self mod modulus, where modulus is odd.

pub const fn invert_odd_mod(&self, modulus: &Odd<Self>) -> CtOption<Self>

Computes the multiplicative inverse of self mod modulus, where modulus is odd.

pub const fn invert_odd_mod_vartime( &self, modulus: &Odd<Self>, ) -> CtOption<Self>

Computes the multiplicative inverse of self mod modulus, where modulus is odd.

This method is variable-time with respect to self.

pub const fn inv_mod(&self, modulus: &Self) -> CtOption<Self>

👎Deprecated since 0.7.0:
please use invert_mod instead

Computes the multiplicative inverse of self mod modulus.

Returns some if an inverse exists, otherwise none.

pub const fn invert_mod(&self, modulus: &NonZero<Self>) -> CtOption<Self>

Computes the multiplicative inverse of self mod modulus.

Returns some if an inverse exists, otherwise none.

impl<const LIMBS: usize> Uint<LIMBS>

pub const fn lcm<const WIDE_LIMBS: usize>(&self, rhs: &Self) -> Uint<WIDE_LIMBS>
where Self: Concat<LIMBS, Output = Uint<WIDE_LIMBS>>,

Compute the least common multiple of self and rhs.

pub const fn lcm_vartime<const WIDE_LIMBS: usize>( &self, rhs: &Self, ) -> Uint<WIDE_LIMBS>
where Self: Concat<LIMBS, Output = Uint<WIDE_LIMBS>>,

Compute the least common multiple of self and rhs.

This method is variable time with respect to self and rhs.

impl<const LIMBS: usize> Uint<LIMBS>

pub const fn jacobi_symbol<const RHS_LIMBS: usize>( &self, rhs: &Odd<Uint<RHS_LIMBS>>, ) -> JacobiSymbol

Compute the Jacobi symbol (self|rhs).

For prime rhs, this corresponds to the Legendre symbol and indicates whether self is quadratic residue modulo rhs.

pub const fn jacobi_symbol_vartime<const RHS_LIMBS: usize>( &self, rhs: &Odd<Uint<RHS_LIMBS>>, ) -> JacobiSymbol

Compute the Jacobi symbol (self|rhs).

For prime rhs, this corresponds to the Legendre symbol and indicates whether self is quadratic residue modulo rhs.

This method executes in variable-time for both inputs.

impl<const LIMBS: usize> Uint<LIMBS>

pub const fn concatenating_mul<const RHS_LIMBS: usize, const WIDE_LIMBS: usize>( &self, rhs: &Uint<RHS_LIMBS>, ) -> Uint<WIDE_LIMBS>
where Self: Concat<RHS_LIMBS, Output = Uint<WIDE_LIMBS>>,

Multiply self by rhs, returning a concatenated “wide” result.

pub const fn split_mul<const RHS_LIMBS: usize>( &self, rhs: &Uint<RHS_LIMBS>, ) -> (Self, Uint<RHS_LIMBS>)

👎Deprecated since 0.7.0:
please use widening_mul instead

Compute “wide” multiplication as a 2-tuple containing the (lo, hi) components of the product, whose sizes correspond to the sizes of the operands.

pub const fn widening_mul<const RHS_LIMBS: usize>( &self, rhs: &Uint<RHS_LIMBS>, ) -> (Self, Uint<RHS_LIMBS>)

Compute “wide” multiplication as a 2-tuple containing the (lo, hi) components of the product, whose sizes correspond to the sizes of the operands.

pub const fn wrapping_mul<const RHS_LIMBS: usize>( &self, rhs: &Uint<RHS_LIMBS>, ) -> Self

Perform wrapping multiplication, discarding overflow.

pub const fn saturating_mul<const RHS_LIMBS: usize>( &self, rhs: &Uint<RHS_LIMBS>, ) -> Self

Perform saturating multiplication, returning MAX on overflow.

pub const fn checked_mul<const RHS_LIMBS: usize>( &self, rhs: &Uint<RHS_LIMBS>, ) -> CtOption<Uint<LIMBS>>

Perform wrapping multiplication, checking that the result fits in the original Uint size.

impl<const LIMBS: usize> Uint<LIMBS>

Squaring operations

pub const fn square_wide(&self) -> (Self, Self)

👎Deprecated since 0.7.0:
please use widening_square instead

Square self, returning a “wide” result in two parts as (lo, hi).

pub const fn widening_square(&self) -> (Self, Self)

Square self, returning a “wide” result in two parts as (lo, hi).

pub const fn concatenating_square<const WIDE_LIMBS: usize>( &self, ) -> Uint<WIDE_LIMBS>
where Self: Concat<LIMBS, Output = Uint<WIDE_LIMBS>>,

Square self, returning a concatenated “wide” result.

pub const fn checked_square(&self) -> CtOption<Uint<LIMBS>>

Square self, checking that the result fits in the original Uint size.

pub const fn wrapping_square(&self) -> Uint<LIMBS>

Perform wrapping square, discarding overflow.

pub const fn saturating_square(&self) -> Self

Perform saturating squaring, returning MAX on overflow.

impl<const LIMBS: usize, const WIDE_LIMBS: usize> Uint<LIMBS>
where Self: Concat<LIMBS, Output = Uint<WIDE_LIMBS>>,

pub const fn square(&self) -> Uint<WIDE_LIMBS>

👎Deprecated since 0.7.0:
please use concatenating_square instead

Square self, returning a concatenated “wide” result.

impl<const LIMBS: usize> Uint<LIMBS>

pub fn mul_mod( &self, rhs: &Uint<LIMBS>, p: &NonZero<Uint<LIMBS>>, ) -> Uint<LIMBS>

Computes self * rhs mod p.

pub fn mul_mod_vartime( &self, rhs: &Uint<LIMBS>, p: &NonZero<Uint<LIMBS>>, ) -> Uint<LIMBS>

Computes self * rhs mod p in variable time with respect to p.

pub const fn mul_mod_special(&self, rhs: &Self, c: Limb) -> Self

Computes self * rhs mod p for the special modulus p = MAX+1-c where c is small enough to fit in a single Limb.

For the modulus reduction, this function implements Algorithm 14.47 from the “Handbook of Applied Cryptography”, by A. Menezes, P. van Oorschot, and S. Vanstone, CRC Press, 1996.

pub const fn square_mod(&self, p: &NonZero<Uint<LIMBS>>) -> Self

Computes self * self mod p.

pub const fn square_mod_vartime(&self, p: &NonZero<Uint<LIMBS>>) -> Self

Computes self * self mod p in variable time with respect to p.

impl<const LIMBS: usize> Uint<LIMBS>

pub const fn widening_mul_signed<const RHS_LIMBS: usize>( &self, rhs: &Int<RHS_LIMBS>, ) -> (Uint<LIMBS>, Uint<RHS_LIMBS>, Choice)

Compute “wide” multiplication between an Uint and Int as 3-tuple (lo, hi, negate). The (lo, hi) components contain the magnitude of the product, with sizes corresponding to the sizes of the operands; negate indicates whether the result should be negated when converted from Uint to Int.

Note: even if negate is truthy, the magnitude might be zero!

pub const fn concatenating_mul_signed<const RHS_LIMBS: usize, const WIDE_LIMBS: usize>( &self, rhs: &Int<RHS_LIMBS>, ) -> Int<WIDE_LIMBS>
where Uint<LIMBS>: Concat<RHS_LIMBS, Output = Uint<WIDE_LIMBS>>,

Multiply self by Int rhs, returning a concatenated “wide” result.

pub fn checked_mul_signed<const RHS_LIMBS: usize>( &self, rhs: &Int<RHS_LIMBS>, ) -> CtOption<Int<LIMBS>>

Checked multiplication of self with Int rhs.

impl<const LIMBS: usize> Uint<LIMBS>

pub const fn wrapping_neg(&self) -> Self

Perform wrapping negation.

pub const fn carrying_neg(&self) -> (Self, Choice)

Perform negation: map self to (self ^ 1111...1111) + 0000...0001. Additionally return whether the carry flag is set on the addition.

Note: the carry is set if and only if self == Self::ZERO.

pub const fn wrapping_neg_if(&self, negate: Choice) -> Self

Perform wrapping negation, if negate is truthy. Otherwise, return self.

impl<const LIMBS: usize> Uint<LIMBS>

pub const fn neg_mod(&self, p: &NonZero<Self>) -> Self

Computes -a mod p. Assumes self is in [0, p).

pub const fn neg_mod_special(&self, c: Limb) -> Self

Computes -a mod p for the special modulus p = MAX+1-c where c is small enough to fit in a single Limb.

impl<const LIMBS: usize> Uint<LIMBS>

pub const fn checked_pow<const RHS_LIMBS: usize>( &self, exp: &Uint<RHS_LIMBS>, ) -> CtOption<Self>

Computes self^exp, returning a CtOption which is none in the case of overflow.

pub const fn checked_pow_bounded_exp<const RHS_LIMBS: usize>( &self, exp: &Uint<RHS_LIMBS>, exp_bits: u32, ) -> CtOption<Self>

Computes self^exp, returning a CtOption which is none in the case of overflow.

NOTE: exp_bits may be leaked in the time pattern.

Panics

if exp_bits exceeds the capacity of rhs

pub const fn checked_pow_vartime<const RHS_LIMBS: usize>( &self, exp: &Uint<RHS_LIMBS>, ) -> CtOption<Self>

Computes self^exp, returning a CtOption which is none in the case of overflow.

This method is variable time in the exponent exp only.

pub const fn saturating_pow<const RHS_LIMBS: usize>( &self, exp: &Uint<RHS_LIMBS>, ) -> Self

Computes self^exp, returning a Self::MAX in the case of overflow.

pub const fn saturating_pow_bounded_exp<const RHS_LIMBS: usize>( &self, exp: &Uint<RHS_LIMBS>, exp_bits: u32, ) -> Self

Computes self^exp, returning a Self::MAX in the case of overflow.

NOTE: exp_bits may be leaked in the time pattern.

Panics

if exp_bits exceeds the capacity of rhs

pub const fn saturating_pow_vartime<const RHS_LIMBS: usize>( &self, exp: &Uint<RHS_LIMBS>, ) -> Self

Computes self^exp, returning a Self::MAX in the case of overflow.

This method is variable time in the exponent exp.

pub const fn wrapping_pow<const RHS_LIMBS: usize>( &self, exp: &Uint<RHS_LIMBS>, ) -> Self

Computes self^exp, discarding overflow.

pub const fn wrapping_pow_bounded_exp<const RHS_LIMBS: usize>( &self, exp: &Uint<RHS_LIMBS>, exp_bits: u32, ) -> Self

Computes self^exp, discarding overflow.

NOTE: exp_bits may be leaked in the time pattern.

Panics

if exp_bits exceeds the capacity of rhs

pub const fn wrapping_pow_vartime<const RHS_LIMBS: usize>( &self, exp: &Uint<RHS_LIMBS>, ) -> Self

Computes self^exp, discarding overflow.

This method is variable time in the exponent exp only.

impl<const LIMBS: usize> Uint<LIMBS>

pub const fn resize<const T: usize>(&self) -> Uint<T>

Construct a Uint<T> from the unsigned integer value, truncating the upper bits if the value is too large to be represented.

pub const fn resize_checked<const T: usize>(&self) -> CtOption<Uint<T>>

Construct a Uint<T> from the unsigned integer value, returning a CtOption which is is_none if the upper limbs were non-zero.

impl<const LIMBS: usize> Uint<LIMBS>

pub const fn floor_root_vartime(&self, exp: NonZeroU32) -> Self

Computes floor(self^(1/exp)).

Callers can check if self is an exact power of exp by exponentiating the result.

This method is variable time in self and in the exponent.

pub fn checked_root_vartime(&self, exp: NonZeroU32) -> Option<Self>

Compute the root self^(1/exp) returning an Option which is_some only if the root is exact.

This method is variable time in self and in the exponent.

impl<const LIMBS: usize> Uint<LIMBS>

pub const fn shl(&self, shift: u32) -> Self

Computes self << shift.

Panics

if shift >= Self::BITS.

pub const fn shl_vartime(&self, shift: u32) -> Self

Computes self << shift in variable time.

NOTE: this operation is variable time with respect to shift ONLY.

When used with a fixed shift, this function is constant-time with respect to self.

Panics

if shift >= Self::BITS.

pub const fn overflowing_shl(&self, shift: u32) -> CtOption<Self>

Computes self << shift.

Returns None if shift >= Self::BITS.

pub const fn overflowing_shl_vartime(&self, shift: u32) -> Option<Self>

Computes self << shift.

Returns None if shift >= Self::BITS.

NOTE: this operation is variable time with respect to shift ONLY.

When used with a fixed shift, this function is constant-time with respect to self.

pub const fn overflowing_shl_vartime_wide( lower_upper: (Self, Self), shift: u32, ) -> Option<(Self, Self)>

Computes a left shift on a wide input as (lo, hi).

Returns None if shift >= Self::BITS.

NOTE: this operation is variable time with respect to shift ONLY.

When used with a fixed shift, this function is constant-time with respect to self.

pub const fn unbounded_shl(&self, shift: u32) -> Self

Computes self << shift in a panic-free manner, returning zero if the shift exceeds the precision.

pub const fn unbounded_shl_vartime(&self, shift: u32) -> Self

Computes self << shift in variable time in a panic-free manner, returning zero if the shift exceeds the precision.

NOTE: this operation is variable time with respect to shift ONLY.

When used with a fixed shift, this function is constant-time with respect to self.

pub const fn bounded_shl(&self, shift: u32, shift_upper_bound: u32) -> Self

Computes self << shift where shift < shift_upper_bound.

The runtime is determined by shift_upper_bound which may be larger or smaller than Self::BITS.

Panics

if the shift exceeds the upper bound.

pub const fn wrapping_shl(&self, shift: u32) -> Self

Computes self << shift in a panic-free manner, reducing shift modulo the type’s width.

pub const fn wrapping_shl_vartime(&self, shift: u32) -> Self

Computes self << shift in variable-time in a panic-free manner, reducing shift modulo the type’s width.

NOTE: this operation is variable time with respect to shift ONLY.

When used with a fixed shift, this function is constant-time with respect to self.

impl<const LIMBS: usize> Uint<LIMBS>

pub const fn shr(&self, shift: u32) -> Self

Computes self >> shift.

Panics

if shift >= Self::BITS.

pub const fn shr_vartime(&self, shift: u32) -> Self

Computes self >> shift in variable time.

Panics

if shift >= Self::BITS.

pub const fn overflowing_shr(&self, shift: u32) -> CtOption<Self>

Computes self >> shift.

Returns None if shift >= Self::BITS.

pub const fn overflowing_shr_vartime(&self, shift: u32) -> Option<Self>

Computes self >> shift.

Returns None if shift >= Self::BITS.

NOTE: this operation is variable time with respect to shift ONLY.

When used with a fixed shift, this function is constant-time with respect to self.

pub const fn overflowing_shr_vartime_wide( lower_upper: (Self, Self), shift: u32, ) -> Option<(Self, Self)>

Computes a right shift on a wide input as (lo, hi).

Returns None if shift >= Self::BITS.

NOTE: this operation is variable time with respect to shift ONLY.

When used with a fixed shift, this function is constant-time with respect to self.

pub const fn bounded_shr(&self, shift: u32, shift_upper_bound: u32) -> Self

Computes self >> shift where shift < shift_upper_bound.

The runtime is determined by shift_upper_bound which may be larger or smaller than Self::BITS.

Panics

if the shift exceeds the upper bound.

pub const fn unbounded_shr(&self, shift: u32) -> Self

Computes self >> shift in a panic-free manner, returning zero if the shift exceeds the precision.

pub const fn unbounded_shr_vartime(&self, shift: u32) -> Self

Computes self >> shift in variable time in a panic-free manner, returning zero if the shift exceeds the precision.

NOTE: this operation is variable time with respect to shift ONLY.

When used with a fixed shift, this function is constant-time with respect to self.

pub const fn wrapping_shr(&self, shift: u32) -> Self

Computes self >> shift in a panic-free manner, reducing shift modulo the type’s width.

pub const fn wrapping_shr_vartime(&self, shift: u32) -> Self

Computes self >> shift in variable-time in a panic-free manner, reducing shift modulo the type’s width.

NOTE: this operation is variable time with respect to shift ONLY.

When used with a fixed shift, this function is constant-time with respect to self.

impl<const LIMBS: usize> Uint<LIMBS>

pub const fn split<const HALF_LIMBS: usize>( &self, ) -> (Uint<HALF_LIMBS>, Uint<HALF_LIMBS>)
where Self: SplitEven<Output = Uint<HALF_LIMBS>>,

Split this number in half into low and high components.

pub const fn split_mixed<const LO_LIMBS: usize, const HI_LIMBS: usize>( &self, ) -> (Uint<LO_LIMBS>, Uint<HI_LIMBS>)
where Self: Split<LO_LIMBS, Output = Uint<HI_LIMBS>>,

Split this number into low and high components respectively.

pub const fn split_checked<const LO_LIMBS: usize, const HI_LIMBS: usize>( &self, ) -> Option<(Uint<LO_LIMBS>, Uint<HI_LIMBS>)>

Split this integer into low and high components. If LIMBS is not equal to the sum of LIMBS and HI_LIMBS, then None is returned.

pub const fn split_resize<const LO_LIMBS: usize, const HI_LIMBS: usize>( &self, ) -> (Uint<LO_LIMBS>, Uint<HI_LIMBS>)

Split this integer into low and high components. The resulting values may be truncated or extended with zeros depending upon whether LIMBS is greater or less than the sum of LO_LIMBS and HI_LIMBS.

impl<const LIMBS: usize> Uint<LIMBS>

pub const fn sqrt(&self) -> Self

👎Deprecated since 0.7.0:
please use floor_sqrt instead

Computes floor(√(self)) in constant time.

Callers can check if self is a square by squaring the result.

pub const fn floor_sqrt(&self) -> Self

Computes floor(√(self)) in constant time.

Callers can check if self is a square by squaring the result.

pub const fn sqrt_vartime(&self) -> Self

👎Deprecated since 0.7.0:
please use floor_sqrt_vartime instead

Computes floor(√(self)).

Callers can check if self is a square by squaring the result.

Variable time with respect to self.

pub const fn floor_sqrt_vartime(&self) -> Self

Computes floor(√(self)).

Callers can check if self is a square by squaring the result.

Variable time with respect to self.

pub const fn wrapping_sqrt(&self) -> Self

Wrapped sqrt is just floor(√(self)). There’s no way wrapping could ever happen. This function exists so that all operations are accounted for in the wrapping operations.

pub const fn wrapping_sqrt_vartime(&self) -> Self

Wrapped sqrt is just floor(√(self)). There’s no way wrapping could ever happen. This function exists so that all operations are accounted for in the wrapping operations.

Variable time with respect to self.

pub fn checked_sqrt(&self) -> CtOption<Self>

Perform checked sqrt, returning a CtOption which is_some only if the square root is exact.

pub fn checked_sqrt_vartime(&self) -> Option<Self>

Perform checked sqrt, returning an Option which is_some only if the square root is exact.

Variable time with respect to self.

impl<const LIMBS: usize> Uint<LIMBS>

pub const fn sbb(&self, rhs: &Self, borrow: Limb) -> (Self, Limb)

👎Deprecated since 0.7.0:
please use borrowing_sub instead

Computes self - (rhs + borrow), returning the result along with the new borrow.

pub const fn borrowing_sub(&self, rhs: &Self, borrow: Limb) -> (Self, Limb)

Computes self - (rhs + borrow), returning the result along with the new borrow.

pub const fn saturating_sub(&self, rhs: &Self) -> Self

Perform saturating subtraction, returning ZERO on underflow.

pub const fn wrapping_sub(&self, rhs: &Self) -> Self

Perform wrapping subtraction, discarding underflow and wrapping around the boundary of the type.

impl<const LIMBS: usize> Uint<LIMBS>

pub const fn sub_mod(&self, rhs: &Self, p: &NonZero<Self>) -> Self

Computes self - rhs mod p.

Assumes self - rhs as unbounded signed integer is in [-p, p).

pub const fn sub_mod_special(&self, rhs: &Self, c: Limb) -> Self

Computes self - rhs mod p for the special modulus p = MAX+1-c where c is small enough to fit in a single Limb.

Assumes self - rhs as unbounded signed integer is in [-p, p).

impl<const LIMBS: usize> Uint<LIMBS>

pub const ZERO: Self

The value 0.

pub const ONE: Self

The value 1.

pub const MAX: Self

Maximum value this Uint can express.

pub const BITS: u32

Total size of the represented integer in bits.

pub const BYTES: usize

Total size of the represented integer in bytes.

pub const LIMBS: usize = LIMBS

The number of limbs used on this platform.

pub const fn new(limbs: [Limb; LIMBS]) -> Self

Const-friendly Uint constructor.

pub const fn from_words(arr: [Word; LIMBS]) -> Self

Create a Uint from an array of Words (i.e. word-sized unsigned integers).

pub const fn to_words(self) -> [Word; LIMBS]

Create an array of Words (i.e. word-sized unsigned integers) from a Uint.

pub const fn as_words(&self) -> &[Word; LIMBS]

Borrow the inner limbs as an array of Words.

pub const fn as_mut_words(&mut self) -> &mut [Word; LIMBS]

Borrow the inner limbs as a mutable array of Words.

pub const fn as_words_mut(&mut self) -> &mut [Word]

👎Deprecated since 0.7.0:
please use as_mut_words instead

Borrow the inner limbs as a mutable slice of Words.

pub const fn as_limbs(&self) -> &[Limb; LIMBS]

Borrow the limbs of this Uint.

pub const fn as_mut_limbs(&mut self) -> &mut [Limb; LIMBS]

Borrow the limbs of this Uint mutably.

pub const fn as_limbs_mut(&mut self) -> &mut [Limb]

👎Deprecated since 0.7.0:
please use as_mut_limbs instead

Borrow the limbs of this Uint mutably.

pub const fn to_limbs(self) -> [Limb; LIMBS]

Convert this Uint into its inner limbs.

pub const fn as_uint_ref(&self) -> &UintRef

Borrow the limbs of this Uint as a UintRef.

pub const fn as_mut_uint_ref(&mut self) -> &mut UintRef

Mutably borrow the limbs of this Uint as a UintRef.

pub const fn as_nz_vartime(&self) -> Option<&NonZero<Self>>

Construct a NonZero reference, returning None in the event self is 0.

pub const fn to_nz(&self) -> CtOption<NonZero<Self>>

Convert to a NonZeroUint<LIMBS>.

Returns some if the original value is non-zero, and none otherwise.

pub const fn to_nz_vartime(&self) -> Option<NonZero<Self>>

Convert to a NonZeroUint<LIMBS>.

Returns Some if the original value is non-zero, and none otherwise.

pub const fn to_odd(&self) -> CtOption<Odd<Self>>

Convert to a OddUint<LIMBS>.

Returns some if the original value is odd, and none otherwise.

pub const fn as_int(&self) -> &Int<LIMBS>

Interpret this object as an Int instead.

Note: this is a casting operation. See Self::try_into_int for the checked equivalent.

pub const fn try_into_int(self) -> CtOption<Int<LIMBS>>

Convert this type into an Int; returns None if this value is greater than Int::MAX.

Note: this is the conversion operation. See Self::as_int for the unchecked equivalent.

pub const fn is_zero(&self) -> Choice

Is this Uint equal to Uint::ZERO?