Struct crypto_bigint::Uint

pub struct Uint<const LIMBS: usize> { /* private fields */ }

Stack-allocated big unsigned integer.

Generic over the given number of LIMBS

Encoding support

This type supports many different types of encodings, either via the Encoding trait or various const fn decoding and encoding functions that can be used with Uint constants.

Optional crate features for encoding (off-by-default):

generic-array: enables ArrayEncoding trait which can be used to Uint as GenericArray<u8, N> and a ArrayDecoding trait which can be used to GenericArray<u8, N> as Uint.

rlp: support for Recursive Length Prefix (RLP) encoding.

Implementations

impl<const LIMBS: usize> Uint<LIMBS>

pub const fn adc(&self, rhs: &Self, carry: Limb) -> (Self, Limb)

Computes a + b + carry, returning the result along with the new carry.

pub const fn saturating_add(&self, rhs: &Self) -> Self

Perform saturating addition, returning MAX on overflow.

pub const fn wrapping_add(&self, rhs: &Self) -> Self

Perform wrapping addition, discarding overflow.

impl<const LIMBS: usize> Uint<LIMBS>

pub const fn add_mod(&self, rhs: &Uint<LIMBS>, p: &Uint<LIMBS>) -> Uint<LIMBS>

Computes self + rhs mod p.

Assumes self + rhs as unbounded integer is < 2p.

pub const fn add_mod_special(&self, rhs: &Self, c: Limb) -> Self

Computes self + rhs mod p for the special modulus p = MAX+1-c where c is small enough to fit in a single Limb.

Assumes self + rhs as unbounded integer is < 2p.

impl<const LIMBS: usize> Uint<LIMBS>

pub const fn bitand(&self, rhs: &Self) -> Self

Computes bitwise a & b.

pub const fn wrapping_and(&self, rhs: &Self) -> Self

Perform wrapping bitwise AND.

Theres no way wrapping could ever happen. This function exists so that all operations are accounted for in the wrapping operations

pub fn checked_and(&self, rhs: &Self) -> CtOption<Self>

Perform checked bitwise AND, returning a CtOption which is_some always

impl<const LIMBS: usize> Uint<LIMBS>

pub const fn not(&self) -> Self

Computes bitwise !a.

impl<const LIMBS: usize> Uint<LIMBS>

pub const fn bitor(&self, rhs: &Self) -> Self

Computes bitwise a & b.

pub const fn wrapping_or(&self, rhs: &Self) -> Self

Perform wrapping bitwise OR.

Theres no way wrapping could ever happen. This function exists so that all operations are accounted for in the wrapping operations

pub fn checked_or(&self, rhs: &Self) -> CtOption<Self>

Perform checked bitwise OR, returning a CtOption which is_some always

impl<const LIMBS: usize> Uint<LIMBS>

pub const fn bitxor(&self, rhs: &Self) -> Self

Computes bitwise a ^ b.

pub const fn wrapping_xor(&self, rhs: &Self) -> Self

Perform wrapping bitwise `XOR``.

Theres no way wrapping could ever happen. This function exists so that all operations are accounted for in the wrapping operations

pub fn checked_xor(&self, rhs: &Self) -> CtOption<Self>

Perform checked bitwise XOR, returning a CtOption which is_some always

impl<const LIMBS: usize> Uint<LIMBS>

pub const fn bit_vartime(&self, index: usize) -> bool

Returns true if the bit at position index is set, false otherwise.

Remarks

This operation is variable time with respect to index only.

pub const fn bits_vartime(&self) -> usize

Calculate the number of bits needed to represent this number.

pub const fn leading_zeros(&self) -> usize

Calculate the number of leading zeros in the binary representation of this number.

pub const fn leading_zeros_vartime(&self) -> usize

Calculate the number of leading zeros in the binary representation of this number, variable time in self.

pub const fn trailing_zeros(&self) -> usize

Calculate the number of trailing zeros in the binary representation of this number.

pub const fn trailing_zeros_vartime(&self) -> usize

Calculate the number of trailing zeros in the binary representation of this number, variable time in self.

pub const fn trailing_ones(&self) -> usize

Calculate the number of trailing ones in the binary representation of this number.

pub const fn trailing_ones_vartime(&self) -> usize

Calculate the number of trailing ones in the binary representation of this number, variable time in self.

pub const fn bits(&self) -> usize

Calculate the number of bits needed to represent this number.

pub const fn bit(&self, index: usize) -> CtChoice

Get the value of the bit at position index, as a truthy or falsy CtChoice. Returns the falsy value for indices out of range.

impl<const LIMBS: usize> Uint<LIMBS>

pub const fn cmp_vartime(&self, rhs: &Self) -> Ordering

Returns the Ordering between self and rhs in variable time.

impl<const LIMBS: usize> Uint<LIMBS>

pub const fn ct_div_rem_limb_with_reciprocal( &self, reciprocal: &Reciprocal, ) -> (Self, Limb)

Computes self / rhs using a pre-made reciprocal, returns the quotient (q) and remainder (r).

pub fn div_rem_limb_with_reciprocal( &self, reciprocal: &CtOption<Reciprocal>, ) -> CtOption<(Self, Limb)>

Computes self / rhs using a pre-made reciprocal, returns the quotient (q) and remainder (r).

pub fn div_rem_limb(&self, rhs: NonZero<Limb>) -> (Self, Limb)

Computes self / rhs, returns the quotient (q) and remainder (r).

pub const fn const_rem(&self, rhs: &Self) -> (Self, CtChoice)

Computes self % rhs, returns the remainder and and the truthy value for is_some or the falsy value for is_none.

NOTE: Use only if you need to access const fn. Otherwise use Self::rem. This is variable only with respect to rhs.

When used with a fixed rhs, this function is constant-time with respect to self.

pub const fn const_rem_wide( lower_upper: (Self, Self), rhs: &Self, ) -> (Self, CtChoice)

Computes self % rhs, returns the remainder and and the truthy value for is_some or the falsy value for is_none.

This is variable only with respect to rhs.

When used with a fixed rhs, this function is constant-time with respect to self.

pub const fn rem2k(&self, k: usize) -> Self

Computes self % 2^k. Faster than reduce since its a power of 2. Limited to 2^16-1 since Uint doesnt support higher.

pub fn div_rem(&self, rhs: &NonZero<Self>) -> (Self, Self)

Computes self / rhs, returns the quotient, remainder.

pub fn rem(&self, rhs: &NonZero<Self>) -> Self

Computes self % rhs, returns the remainder.

pub const fn wrapping_div(&self, rhs: &Self) -> Self

Wrapped division is just normal division i.e. self / rhs Theres no way wrapping could ever happen. This function exists, so that all operations are accounted for in the wrapping operations.

Panics if rhs == 0.

pub fn checked_div(&self, rhs: &Self) -> CtOption<Self>

Perform checked division, returning a CtOption which is_some only if the rhs != 0

pub const fn wrapping_rem(&self, rhs: &Self) -> Self

Wrapped (modular) remainder calculation is just self % rhs. Theres no way wrapping could ever happen. This function exists, so that all operations are accounted for in the wrapping operations.

Panics if rhs == 0.

pub fn checked_rem(&self, rhs: &Self) -> CtOption<Self>

Perform checked reduction, returning a CtOption which is_some only if the rhs != 0

impl<const LIMBS: usize> Uint<LIMBS>

pub const fn from_be_slice(bytes: &[u8]) -> Self

Create a new Uint from the provided big endian bytes.

pub const fn from_be_hex(hex: &str) -> Self

Create a new Uint from the provided big endian hex string.

pub const fn from_le_slice(bytes: &[u8]) -> Self

Create a new Uint from the provided little endian bytes.

pub const fn from_le_hex(hex: &str) -> Self

Create a new Uint from the provided little endian hex string.

impl<const LIMBS: usize> Uint<LIMBS>

pub const fn from_u8(n: u8) -> Self

Create a Uint from a u8 (const-friendly)

pub const fn from_u16(n: u16) -> Self

Create a Uint from a u16 (const-friendly)

pub const fn from_u32(n: u32) -> Self

Create a Uint from a u32 (const-friendly)

pub const fn from_u64(n: u64) -> Self

Create a Uint from a u64 (const-friendly)

pub const fn from_u128(n: u128) -> Self

Create a Uint from a u128 (const-friendly)

pub const fn from_word(n: Word) -> Self

Create a Uint from a Word (const-friendly)

pub const fn from_wide_word(n: WideWord) -> Self

Create a Uint from a WideWord (const-friendly)

impl<const LIMBS: usize> Uint<LIMBS>

pub const fn inv_mod2k_vartime(&self, k: usize) -> Self

Computes 1/self mod 2^k. This method is constant-time w.r.t. self but not k.

Conditions: self < 2^k and self must be odd

pub const fn inv_mod2k(&self, k: usize) -> Self

Computes 1/self mod 2^k.

Conditions: self < 2^k and self must be odd

pub const fn inv_odd_mod_bounded( &self, modulus: &Self, bits: usize, modulus_bits: usize, ) -> (Self, CtChoice)

Computes the multiplicative inverse of self mod modulus, where modulus is odd. In other words self^-1 mod modulus. bits and modulus_bits are the bounds on the bit size of self and modulus, respectively (the inversion speed will be proportional to bits + modulus_bits). The second element of the tuple is the truthy value if an inverse exists, otherwise it is a falsy value.

Note: variable time in bits and modulus_bits.

The algorithm is the same as in GMP 6.2.1s mpn_sec_invert.

pub const fn inv_odd_mod(&self, modulus: &Self) -> (Self, CtChoice)

Computes the multiplicative inverse of self mod modulus, where modulus is odd. Returns (inverse, CtChoice::TRUE) if an inverse exists, otherwise (undefined, CtChoice::FALSE).

pub const fn inv_mod(&self, modulus: &Self) -> (Self, CtChoice)

Computes the multiplicative inverse of self mod modulus. Returns (inverse, CtChoice::TRUE) if an inverse exists, otherwise (undefined, CtChoice::FALSE).

impl<const LIMBS: usize> Uint<LIMBS>

pub fn mul<const HLIMBS: usize>( &self, rhs: &Uint<HLIMBS>, ) -> <Uint<HLIMBS> as ConcatMixed<Self>>::MixedOutput
where Uint<HLIMBS>: ConcatMixed<Self>,

Multiply self by rhs, returning a concatenated wide result.

pub const fn mul_wide<const HLIMBS: usize>( &self, rhs: &Uint<HLIMBS>, ) -> (Self, Uint<HLIMBS>)

Compute wide multiplication, with a product twice the size of the input.

Returns a tuple containing the (lo, hi) components of the product.

Ordering note

Releases of crypto-bigint prior to v0.3 used (hi, lo) ordering instead. This has been changed for better consistency with the rest of the APIs in this crate.

For more info see: https://github.com/RustCrypto/crypto-bigint/issues/4

pub const fn saturating_mul<const HLIMBS: usize>( &self, rhs: &Uint<HLIMBS>, ) -> Self

Perform saturating multiplication, returning MAX on overflow.

pub const fn wrapping_mul<const H: usize>(&self, rhs: &Uint<H>) -> Self

Perform wrapping multiplication, discarding overflow.

pub fn square(&self) -> <Self as Concat>::Output
where Self: Concat,

Square self, returning a concatenated wide result.

pub const fn square_wide(&self) -> (Self, Self)

Square self, returning a wide result in two parts as (lo, hi).

impl<const LIMBS: usize> Uint<LIMBS>

pub const fn mul_mod_special(&self, rhs: &Self, c: Limb) -> Self

Computes self * rhs mod p for the special modulus p = MAX+1-c where c is small enough to fit in a single Limb. For the modulus reduction, this function implements Algorithm 14.47 from the Handbook of Applied Cryptography, by A. Menezes, P. van Oorschot, and S. Vanstone, CRC Press, 1996.

impl<const LIMBS: usize> Uint<LIMBS>

pub const fn wrapping_neg(&self) -> Self

Perform wrapping negation.

impl<const LIMBS: usize> Uint<LIMBS>

pub const fn neg_mod(&self, p: &Self) -> Self

Computes -a mod p. Assumes self is in [0, p).

pub const fn neg_mod_special(&self, c: Limb) -> Self

Computes -a mod p for the special modulus p = MAX+1-c where c is small enough to fit in a single Limb.

impl<const LIMBS: usize> Uint<LIMBS>

pub const fn resize<const T: usize>(&self) -> Uint<T>

Construct a Uint<T> from the unsigned integer value, truncating the upper bits if the value is too large to be represented.

impl<const LIMBS: usize> Uint<LIMBS>

pub const fn shl_vartime(&self, n: usize) -> Self

Computes self << shift.

NOTE: this operation is variable time with respect to n ONLY.

When used with a fixed n, this function is constant-time with respect to self.

pub const fn shl_vartime_wide( lower_upper: (Self, Self), n: usize, ) -> (Self, Self)

Computes a left shift on a wide input as (lo, hi).

NOTE: this operation is variable time with respect to n ONLY.

When used with a fixed n, this function is constant-time with respect to self.

pub const fn shl(&self, shift: usize) -> Self

Computes self << n. Returns zero if n >= Self::BITS.

impl<const LIMBS: usize> Uint<LIMBS>

pub const fn shr_vartime(&self, shift: usize) -> Self

Computes self >> n.

NOTE: this operation is variable time with respect to n ONLY.

When used with a fixed n, this function is constant-time with respect to self.

pub const fn shr_vartime_wide( lower_upper: (Self, Self), n: usize, ) -> (Self, Self)

Computes a right shift on a wide input as (lo, hi).

NOTE: this operation is variable time with respect to n ONLY.

When used with a fixed n, this function is constant-time with respect to self.

pub const fn shr(&self, shift: usize) -> Self

Computes self << n. Returns zero if n >= Self::BITS.

impl<const LIMBS: usize> Uint<LIMBS>

pub const fn sqrt(&self) -> Self

Deprecated since 0.5.3: This functionality will be moved to sqrt_vartime in a future release.

See Self::sqrt_vartime.

pub const fn sqrt_vartime(&self) -> Self

Computes (self) Uses Brent & Zimmermann, Modern Computer Arithmetic, v0.5.9, Algorithm 1.13

Callers can check if self is a square by squaring the result

pub const fn wrapping_sqrt(&self) -> Self

Deprecated since 0.5.3: This functionality will be moved to wrapping_sqrt_vartime in a future release.

See Self::wrapping_sqrt_vartime.

pub const fn wrapping_sqrt_vartime(&self) -> Self

Wrapped sqrt is just normal (self) Theres no way wrapping could ever happen. This function exists, so that all operations are accounted for in the wrapping operations.

pub fn checked_sqrt(&self) -> CtOption<Self>

Deprecated since 0.5.3: This functionality will be moved to checked_sqrt_vartime in a future release.

See Self::checked_sqrt_vartime.

pub fn checked_sqrt_vartime(&self) -> CtOption<Self>

Perform checked sqrt, returning a CtOption which is_some only if the (self) == self

impl<const LIMBS: usize> Uint<LIMBS>

pub const fn sbb(&self, rhs: &Self, borrow: Limb) -> (Self, Limb)

Computes a - (b + borrow), returning the result along with the new borrow.

pub const fn saturating_sub(&self, rhs: &Self) -> Self

Perform saturating subtraction, returning ZERO on underflow.

pub const fn wrapping_sub(&self, rhs: &Self) -> Self

Perform wrapping subtraction, discarding underflow and wrapping around the boundary of the type.

impl<const LIMBS: usize> Uint<LIMBS>

pub const fn sub_mod(&self, rhs: &Uint<LIMBS>, p: &Uint<LIMBS>) -> Uint<LIMBS>

Computes self - rhs mod p.

Assumes self - rhs as unbounded signed integer is in [-p, p).

pub const fn sub_mod_special(&self, rhs: &Self, c: Limb) -> Self

Computes self - rhs mod p for the special modulus p = MAX+1-c where c is small enough to fit in a single Limb.

Assumes self - rhs as unbounded signed integer is in [-p, p).

impl<const LIMBS: usize> Uint<LIMBS>

pub const ZERO: Self = _

The value 0.

pub const ONE: Self = _

The value 1.

pub const MAX: Self = _

Maximum value this Uint can express.

pub const BITS: usize = _

Total size of the represented integer in bits.

pub const BYTES: usize = _

Total size of the represented integer in bytes.

pub const LIMBS: usize = LIMBS

The number of limbs used on this platform.

pub const fn new(limbs: [Limb; LIMBS]) -> Self

Const-friendly Uint constructor.

pub const fn from_words(arr: [Word; LIMBS]) -> Self

Create a Uint from an array of Words (i.e. word-sized unsigned integers).

pub const fn to_words(self) -> [Word; LIMBS]

Create an array of Words (i.e. word-sized unsigned integers) from a Uint.

pub const fn as_words(&self) -> &[Word; LIMBS]

Borrow the inner limbs as an array of Words.

pub fn as_words_mut(&mut self) -> &mut [Word; LIMBS]

Borrow the inner limbs as a mutable array of Words.

pub const fn as_limbs(&self) -> &[Limb; LIMBS]

Borrow the limbs of this Uint.

pub fn as_limbs_mut(&mut self) -> &mut [Limb; LIMBS]

Borrow the limbs of this Uint mutably.

pub const fn to_limbs(self) -> [Limb; LIMBS]

Convert this Uint into its inner limbs.

impl Uint<{ _ }>

pub const fn concat(&self, lo: &Uint<{ _ }>) -> U128

Concatenate the two values, with self as most significant and rhs as the least significant.

impl Uint<{nlimbs!($bits)}>

pub const fn split(&self) -> (Uint<{ _ }>, Uint<{ _ }>)

Split this number in half, returning its high and low components respectively.

impl Uint<{ _ }>

pub const fn concat(&self, lo: &Uint<{ _ }>) -> U256

Concatenate the two values, with self as most significant and rhs as the least significant.

impl Uint<{nlimbs!($bits)}>

pub const fn split(&self) -> (Uint<{ _ }>, Uint<{ _ }>)

Split this number in half, returning its high and low components respectively.

impl Uint<{ _ }>

pub const fn concat(&self, lo: &Uint<{ _ }>) -> U384

Concatenate the two values, with self as most significant and rhs as the least significant.

impl Uint<{nlimbs!($bits)}>

pub const fn split(&self) -> (Uint<{ _ }>, Uint<{ _ }>)

Split this number in half, returning its high and low components respectively.

impl Uint<{ _ }>

pub const fn concat(&self, lo: &Uint<{ _ }>) -> U512

Concatenate the two values, with self as most significant and rhs as the least significant.

impl Uint<{nlimbs!($bits)}>

pub const fn split(&self) -> (Uint<{ _ }>, Uint<{ _ }>)

Split this number in half, returning its high and low components respectively.

impl Uint<{ _ }>

pub const fn concat(&self, lo: &Uint<{ _ }>) -> U640

Concatenate the two values, with self as most significant and rhs as the least significant.

impl Uint<{nlimbs!($bits)}>

pub const fn split(&self) -> (Uint<{ _ }>, Uint<{ _ }>)

Split this number in half, returning its high and low components respectively.

impl Uint<{ _ }>

pub const fn concat(&self, lo: &Uint<{ _ }>) -> U768

Concatenate the two values, with self as most significant and rhs as the least significant.

impl Uint<{nlimbs!($bits)}>

pub const fn split(&self) -> (Uint<{ _ }>, Uint<{ _ }>)

Split this number in half, returning its high and low components respectively.

impl Uint<{ _ }>

pub const fn concat(&self, lo: &Uint<{ _ }>) -> U896

Concatenate the two values, with self as most significant and rhs as the least significant.

impl Uint<{nlimbs!($bits)}>

pub const fn split(&self) -> (Uint<{ _ }>, Uint<{ _ }>)

Split this number in half, returning its high and low components respectively.

impl Uint<{ _ }>

pub const fn concat(&self, lo: &Uint<{ _ }>) -> U1024

Concatenate the two values, with self as most significant and rhs as the least significant.

impl Uint<{nlimbs!($bits)}>

pub const fn split(&self) -> (Uint<{ _ }>, Uint<{ _ }>)

Split this number in half, returning its high and low components respectively.

impl Uint<{ _ }>

pub const fn concat(&self, lo: &Uint<{ _ }>) -> U1280

Concatenate the two values, with self as most significant and rhs as the least significant.

impl Uint<{nlimbs!($bits)}>

pub const fn split(&self) -> (Uint<{ _ }>, Uint<{ _ }>)

Split this number in half, returning its high and low components respectively.

impl Uint<{ _ }>

pub const fn concat(&self, lo: &Uint<{ _ }>) -> U1536

Concatenate the two values, with self as most significant and rhs as the least significant.

impl Uint<{nlimbs!($bits)}>

pub const fn split(&self) -> (Uint<{ _ }>, Uint<{ _ }>)

Split this number in half, returning its high and low components respectively.

impl Uint<{ _ }>

pub const fn concat(&self, lo: &Uint<{ _ }>) -> U1792

Concatenate the two values, with self as most significant and rhs as the least significant.

impl Uint<{nlimbs!($bits)}>

pub const fn split(&self) -> (Uint<{ _ }>, Uint<{ _ }>)

Split this number in half, returning its high and low components respectively.

impl Uint<{ _ }>

pub const fn concat(&self, lo: &Uint<{ _ }>) -> U2048

Concatenate the two values, with self as most significant and rhs as the least significant.

impl Uint<{nlimbs!($bits)}>

pub const fn split(&self) -> (Uint<{ _ }>, Uint<{ _ }>)

Split this number in half, returning its high and low components respectively.

impl Uint<{ _ }>

pub const fn concat(&self, lo: &Uint<{ _ }>) -> U3072

Concatenate the two values, with self as most significant and rhs as the least significant.

impl Uint<{nlimbs!($bits)}>

pub const fn split(&self) -> (Uint<{ _ }>, Uint<{ _ }>)

Split this number in half, returning its high and low components respectively.

impl Uint<{ _ }>

pub const fn concat(&self, lo: &Uint<{ _ }>) -> U3584

Concatenate the two values, with self as most significant and rhs as the least significant.

impl Uint<{nlimbs!($bits)}>

pub const fn split(&self) -> (Uint<{ _ }>, Uint<{ _ }>)

Split this number in half, returning its high and low components respectively.

impl Uint<{ _ }>

pub const fn concat(&self, lo: &Uint<{ _ }>) -> U4096

Concatenate the two values, with self as most significant and rhs as the least significant.

impl Uint<{nlimbs!($bits)}>

pub const fn split(&self) -> (Uint<{ _ }>, Uint<{ _ }>)

Split this number in half, returning its high and low components respectively.

impl Uint<{ _ }>

pub const fn concat(&self, lo: &Uint<{ _ }>) -> U4224

Concatenate the two values, with self as most significant and rhs as the least significant.

impl Uint<{nlimbs!($bits)}>

pub const fn split(&self) -> (Uint<{ _ }>, Uint<{ _ }>)

Split this number in half, returning its high and low components respectively.

impl Uint<{ _ }>

pub const fn concat(&self, lo: &Uint<{ _ }>) -> U4352

Concatenate the two values, with self as most significant and rhs as the least significant.

impl Uint<{nlimbs!($bits)}>

pub const fn split(&self) -> (Uint<{ _ }>, Uint<{ _ }>)

Split this number in half, returning its high and low components respectively.

impl Uint<{ _ }>

pub const fn concat(&self, lo: &Uint<{ _ }>) -> U6144

Concatenate the two values, with self as most significant and rhs as the least significant.

impl Uint<{nlimbs!($bits)}>

pub const fn split(&self) -> (Uint<{ _ }>, Uint<{ _ }>)

Split this number in half, returning its high and low components respectively.

impl Uint<{ _ }>

pub const fn concat(&self, lo: &Uint<{ _ }>) -> U8192

Concatenate the two values, with self as most significant and rhs as the least significant.

impl Uint<{nlimbs!($bits)}>

pub const fn split(&self) -> (Uint<{ _ }>, Uint<{ _ }>)

Split this number in half, returning its high and low components respectively.

impl Uint<{ _ }>

pub const fn concat(&self, lo: &Uint<{ _ }>) -> U16384

Concatenate the two values, with self as most significant and rhs as the least significant.

impl Uint<{nlimbs!($bits)}>

pub const fn split(&self) -> (Uint<{ _ }>, Uint<{ _ }>)

Split this number in half, returning its high and low components respectively.