crypto_bigint

Struct UintRef

pub struct UintRef { /* private fields */ }

Unsigned integer reference type.

This type contains a limb slice which can be borrowed from either a Uint or [BoxedUint] and thus provides an abstraction for writing shared implementations.

Implementations

impl UintRef

pub const fn add_assign_limb(&mut self, rhs: Limb) -> Limb

Perform an in-place carrying add of a limb, returning the carried limb value.

pub const fn carrying_add_assign(&mut self, rhs: &Self, carry: Limb) -> Limb

Perform an in-place carrying add of another UintRef, returning the carried limb value.

pub const fn carrying_add_assign_slice( &mut self, rhs: &[Limb], carry: Limb, ) -> Limb

Perform an in-place carrying add of another limb slice, returning the carried limb value.

Panics

If self and rhs have different lengths.

pub const fn conditional_add_assign( &mut self, rhs: &Self, carry: Limb, choice: Choice, ) -> Limb

Perform an in-place carrying add of another limb slice, returning the carried limb value.

pub const fn conditional_add_assign_slice( &mut self, rhs: &[Limb], carry: Limb, choice: Choice, ) -> Limb

Perform an in-place carrying add of another limb slice, returning the carried limb value.

Panics

If self and rhs have different lengths.

pub const fn carrying_add_assign_mul_limb( &mut self, rhs: &Self, rhs_mul: Limb, carry: Limb, ) -> Limb

Perform an in-place carrying add of another UintRef multiplied by a limb, returning the carried limb value.

Panics

If self is shorter than rhs.

impl UintRef

pub const fn bits_precision(&self) -> u32

Get the precision of this number in bits.

pub const fn bit(&self, index: u32) -> Choice

Get the value of the bit at position index, as a truthy or falsy Choice. Returns the falsy value for indices out of range.

pub const fn bit_vartime(&self, index: u32) -> bool

Returns true if the bit at position index is set, false for an unset bit or for indices out of range.

Remarks

This operation is variable time with respect to index only.

pub const fn bits(&self) -> u32

Calculate the number of bits needed to represent this number, i.e. the index of the highest set bit.

Use UintRef::bits_precision to get the total capacity of this integer.

pub const fn bits_vartime(&self) -> u32

Calculate the number of bits needed to represent this number in variable-time with respect to self.

pub const fn set_bit(&mut self, index: u32, bit_value: Choice)

Sets the bit at index to 0 or 1 depending on the value of bit_value.

pub const fn set_bit_vartime(&mut self, index: u32, bit_value: bool)

Sets the bit at index to 0 or 1 depending on the value of bit_value, in variable-time with respect to index.

pub const fn leading_zeros(&self) -> u32

Calculate the number of leading zeros in the binary representation of this number.

pub const fn trailing_zeros(&self) -> u32

Calculate the number of trailing zeros in the binary representation of this number.

pub const fn trailing_zeros_vartime(&self) -> u32

Calculate the number of trailing zeros in the binary representation of this number, in variable-time with respect to self.

pub const fn trailing_ones(&self) -> u32

Calculate the number of trailing ones in the binary representation of this number.

pub const fn trailing_ones_vartime(&self) -> u32

Calculate the number of trailing ones in the binary representation of this number, in variable-time with respect to self.

pub const fn restrict_bits(&mut self, len: u32)

Clear all bits at or above a given bit position.

impl UintRef

pub const fn is_odd(&self) -> Choice

Returns the truthy value if self is odd or the falsy value otherwise.

pub const fn is_nonzero(&self) -> Choice

Returns Choice::TRUE if self != 0 or Choice::FALSE otherwise.

pub const fn is_zero(&self) -> Choice

Are all of limbs equal to 0?

pub const fn cmp_vartime(&self, rhs: &Self) -> Ordering

Returns the Ordering between self and rhs in variable time.

impl UintRef

pub const fn overflowing_mul(&self, rhs: &UintRef, out: &mut UintRef) -> Choice

Compute the wrapping product of self and rhs, placing the result into out and returning a Choice indicating whether overflow occurred.

pub const fn overflowing_square(&self, out: &mut UintRef) -> Choice

Compute the wrapping squaring of self, placing the result into out and returning a Choice indicating whether overflow occurred.

pub const fn wrapping_mul(&self, rhs: &UintRef, out: &mut UintRef) -> Limb

Compute the wrapping product of self and rhs, placing the result into out and returning a carry Limb.

pub const fn wrapping_square(&self, out: &mut UintRef) -> Limb

Compute the wrapping squaring of self, placing the result into out and returning a carry Limb.

impl UintRef

pub const fn shl_assign(&mut self, shift: u32)

Left-shifts by shift bits.

Panics

if the shift exceeds the type’s width.

pub const fn overflowing_shl_assign(&mut self, shift: u32) -> Choice

Left-shifts by shift bits in constant-time.

Returns truthy Choice and leaves self unmodified if shift >= self.bits_precision(), otherwise returns a falsy Choice and shifts self in place.

pub const fn unbounded_shl_assign(&mut self, shift: u32)

Left-shifts by shift bits, producing zero if the shift exceeds the precision.

pub const fn bounded_shl_assign(&mut self, shift: u32, shift_upper_bound: u32)

Left-shifts by shift bits where shift < shift_upper_bound.

The runtime is determined by shift_upper_bound which may be larger or smaller than self.bits_precision().

Panics

if the shift exceeds the upper bound.

pub const fn unbounded_shl_assign_vartime(&mut self, shift: u32)

Left-shifts by shift bits in a panic-free manner, producing zero if the shift exceeds the precision.

NOTE: this operation is variable time with respect to shift ONLY.

When used with a fixed shift, this function is constant-time with respect to self.

pub const fn wrapping_shl_assign(&mut self, shift: u32)

Left-shifts by shift bits in a panic-free manner, reducing shift modulo the type’s width.

pub const fn wrapping_shl_assign_vartime(&mut self, shift: u32)

Left-shifts by shift bits in a panic-free manner, reducing shift modulo the type’s width.

NOTE: this operation is variable time with respect to shift ONLY.

When used with a fixed shift, this function is constant-time with respect to self.

pub const fn shl1_assign(&mut self) -> Limb

Left-shifts by a single bit in constant-time, returning Limb::ONE if the least significant bit was set, and Limb::ZERO otherwise.

pub const fn shl_assign_limb(&mut self, shift: u32) -> Limb

Left-shifts by shift bits where 0 < shift < Limb::BITS, returning the carry.

Panics

if shift >= Limb::BITS.

impl UintRef

pub const fn shr_assign(&mut self, shift: u32)

Right-shifts by shift bits.

Panics

if the shift exceeds the type’s width.

pub const fn overflowing_shr_assign(&mut self, shift: u32) -> Choice

Right-shifts by shift bits in constant-time.

Returns truthy Choice and leaves self unmodified if shift >= self.bits_precision(), otherwise returns a falsy Choice and shifts self in place.

pub const fn unbounded_shr_assign(&mut self, shift: u32)

Right-shifts by shift bits, producing zero if the shift exceeds the precision.

pub const fn bounded_shr_assign(&mut self, shift: u32, shift_upper_bound: u32)

Right-shifts by shift bits where shift < shift_upper_bound, producing zero if the shift exceeds the precision.

The runtime is determined by shift_upper_bound which may be smaller than self.bits_precision().

Panics

if the shift exceeds the upper bound.

pub const fn unbounded_shr_assign_vartime(&mut self, shift: u32)

Right-shifts by shift bits in a panic-free manner, producing zero if the shift exceeds the precision.

NOTE: this operation is variable time with respect to shift ONLY.

When used with a fixed shift, this function is constant-time with respect to self.

pub const fn wrapping_shr_assign(&mut self, shift: u32)

Right-shifts by shift bits in a panic-free manner, reducing shift modulo the type’s width.

pub const fn wrapping_shr_assign_vartime(&mut self, shift: u32)

Right-shifts by shift bits in a panic-free manner, reducing shift modulo the type’s width.

NOTE: this operation is variable time with respect to shift ONLY.

When used with a fixed shift, this function is constant-time with respect to self.

pub const fn shr1_assign(&mut self) -> Limb

Right-shifts by a single bit in constant-time, returning Limb::ONE << Limb::HI_BIT if the least significant bit was set, and Limb::ZERO otherwise.

pub const fn shr_assign_limb(&mut self, shift: u32) -> Limb

Right-shifts by shift bits where 0 < shift < Limb::BITS, returning the carry.

Panics

if shift >= Limb::BITS.

impl UintRef

pub const fn copy_from(&mut self, rhs: &UintRef)

Copy the contents from a UintRef.

Panics

If self.nlimbs() != rhs.nlimbs()

pub const fn conditional_copy_from(&mut self, rhs: &UintRef, copy: Choice)

Conditionally copy the contents from a UintRef.

Panics

If self.nlimbs() != rhs.nlimbs()

pub const fn copy_from_slice(&mut self, limbs: &[Limb])

Copy the contents from a limb slice.

Panics

If self.nlimbs() != limbs.len()

pub const fn conditional_copy_from_slice( &mut self, limbs: &[Limb], copy: Choice, )

Conditionally copy the contents from a limb slice.

Panics

If self.nlimbs() != rhs.nlimbs()

pub const fn fill(&mut self, limb: Limb)

Fill the limb slice with a repeated limb value.

pub const fn split_at(&self, mid: usize) -> (&Self, &Self)

Split the limb slice at a fixed position, producing head and tail slices.

pub const fn split_at_mut(&mut self, mid: usize) -> (&mut Self, &mut Self)

Split the mutable limb slice at index mid, producing head and tail slices.

pub const fn leading(&self, len: usize) -> &Self

Access a limb slice up to a number of elements len.

pub const fn leading_mut(&mut self, len: usize) -> &mut Self

Access a mutable limb slice up to a number of elements len.

pub const fn trailing(&self, start: usize) -> &Self

Access a limb slice starting from the index start.

pub const fn trailing_mut(&mut self, start: usize) -> &mut Self

Access a mutable limb slice starting from the index start.

pub const fn is_empty(&self) -> bool

Determine if the slice has zero limbs.

impl UintRef

pub const fn borrowing_sub_assign_limb( &mut self, rhs: Limb, borrow: Limb, ) -> Limb

Perform an in-place borrowing sub of a limb, returning the borrowed limb value.

pub const fn borrowing_sub_assign(&mut self, rhs: &Self, borrow: Limb) -> Limb

Perform an in-place borrowing subtraction of another UintRef, returning the carried limb value.

pub const fn borrowing_sub_assign_slice( &mut self, rhs: &[Limb], borrow: Limb, ) -> Limb

Perform an in-place borrowing subtraction of another limb slice, returning the borrowed limb value.

Panics

If self and rhs have different lengths.

impl UintRef

pub const fn new(limbs: &[Limb]) -> &Self

Create a UintRef reference type from a Limb slice.

pub const fn new_mut(limbs: &mut [Limb]) -> &mut Self

Create a mutable UintRef reference type from a Limb slice.

pub const fn new_flattened_mut<const N: usize>( slice: &mut [[Limb; N]], ) -> &mut Self

Create a new mutable UintRef reference type from a slice of Limb arrays.

pub const fn as_limbs(&self) -> &[Limb]

Borrow the inner &[Limb] slice.

pub const fn as_mut_limbs(&mut self) -> &mut [Limb]

Mutably borrow the inner &mut [Limb] slice.

pub const fn as_words(&self) -> &[Word]

Borrow the inner limbs as a slice of Words.

pub const fn as_mut_words(&mut self) -> &mut [Word]

Borrow the inner limbs as a mutable slice of Words.

pub fn iter(&self) -> impl DoubleEndedIterator<Item = &Limb>

Get an iterator over the inner limbs.

pub fn iter_mut(&mut self) -> impl DoubleEndedIterator<Item = &mut Limb>

Get a mutable iterator over the inner limbs.

pub const fn nlimbs(&self) -> usize

Access the number of limbs.

pub const fn conditional_set_zero(&mut self, choice: Choice)

Conditionally assign all of the limbs to zero.

pub const fn conditional_set_one(&mut self, choice: Choice)

Conditionally assign the value to one.

pub const fn conditional_set_max(&mut self, choice: Choice)

Conditionally assign all of the limbs to the maximum.

pub const fn to_uint_resize<const LIMBS: usize>(&self) -> Uint<LIMBS>

Extract up to LIMBS limbs into a new Uint.

pub const fn as_nz_vartime(&self) -> Option<&NonZero<Self>>

Construct a NonZero reference, returning None in the event self is 0.

pub const fn as_odd_vartime(&self) -> Option<&Odd<Self>>

Construct a Odd reference, returning None in the event self is even.