Trait crypto_bigint::ArrayEncoding

pub trait ArrayEncoding: Encoding {
    type ByteSize: ArrayLength<u8> + Add + Eq + Ord + Unsigned;
    fn from_be_byte_array(bytes: ByteArray<Self>) -> Self;
    fn from_le_byte_array(bytes: ByteArray<Self>) -> Self;
    fn to_be_byte_array(&self) -> ByteArray<Self>;
    fn to_le_byte_array(&self) -> ByteArray<Self>;
}

Support for encoding a big integer as a GenericArray.

Associated Types

type ByteSize: ArrayLength<u8> + Add + Eq + Ord + Unsigned

Size of a byte array which encodes a big integer.

Required methods

fn from_be_byte_array(bytes: ByteArray<Self>) -> Self

Deserialize from a big-endian byte array.

fn from_le_byte_array(bytes: ByteArray<Self>) -> Self

Deserialize from a little-endian byte array.

fn to_be_byte_array(&self) -> ByteArray<Self>

Serialize to a big-endian byte array.

fn to_le_byte_array(&self) -> ByteArray<Self>

Serialize to a little-endian byte array.