Trait crypto_bigint::Concat

pub trait Concat<Rhs = Self> {
    type Output;
    fn concat(&self, rhs: &Self) -> Self::Output;
}

Concatenate two numbers into a “wide” twice-width value, using the rhs value as the least significant value.

Associated Types

type Output

Concatenated output: twice the width of Self.

Required methods

fn concat(&self, rhs: &Self) -> Self::Output

Concate the two values, with self as most significant and rhs as the least significant.