crypto_bigint
Trait Split

pub trait Split<Rhs = Self> {
    type Output;

    // Required method
    fn split(&self) -> (Self::Output, Self::Output);
}

Split a number in half, returning the most significant half followed by the least significant.

Required Associated Types

type Output

Split output: high/low components of the value.

Required Methods

fn split(&self) -> (Self::Output, Self::Output)

Split this number in half, returning its high and low components respectively.