crypto_bigint

Type Alias NonZeroUint

pub type NonZeroUint<const LIMBS: usize> = NonZero<Uint<LIMBS>>;

Non-zero unsigned integer.

Aliased Type

pub struct NonZeroUint<const LIMBS: usize>(/* private fields */);

Implementations

impl<const LIMBS: usize> NonZeroUint<LIMBS>

pub const fn new_unwrap(n: Uint<LIMBS>) -> Self

Creates a new non-zero integer in a const context.

In future versions of Rust it should be possible to replace this with NonZero::new(…).unwrap()

Panics

if the value is zero.

pub const fn from_be_hex(hex: &str) -> Self

Create a new NonZero<Uint> from the provided big endian hex string.

Panics

if the hex is zero, malformed, or not zero-padded accordingly for the size.

pub const fn from_le_hex(hex: &str) -> Self

Create a new NonZero<Uint> from the provided little endian hex string.

Panics

if the hex is zero, malformed, or not zero-padded accordingly for the size.

pub const fn from_u8(n: NonZeroU8) -> Self

Create a NonZeroUint from a NonZeroU8 (const-friendly)

pub const fn from_u16(n: NonZeroU16) -> Self

Create a NonZeroUint from a NonZeroU16 (const-friendly)

pub const fn from_u32(n: NonZeroU32) -> Self

Create a NonZeroUint from a NonZeroU32 (const-friendly)

pub const fn from_u64(n: NonZeroU64) -> Self

Create a NonZeroUint from a NonZeroU64 (const-friendly)

pub const fn from_u128(n: NonZeroU128) -> Self

Create a NonZeroUint from a NonZeroU128 (const-friendly)

pub const fn as_uint_ref(&self) -> &NonZeroUintRef

Borrow this NonZero<Uint> as a &NonZeroUintRef.

impl<const LIMBS: usize> NonZeroUint<LIMBS>

pub const fn gcd_unsigned(&self, rhs: &Uint<LIMBS>) -> Self

Compute the greatest common divisor of self and rhs.

pub const fn gcd_unsigned_vartime(&self, rhs: &Uint<LIMBS>) -> Self

Compute the greatest common divisor of self and rhs.

Executes in variable time w.r.t. all input parameters.

pub const fn xgcd(&self, rhs: &Self) -> XgcdOutput<LIMBS, NonZeroUint<LIMBS>>

Execute the Extended GCD algorithm.

Given (self, rhs), computes (g, x, y) s.t. self * x + rhs * y = g = gcd(self, rhs).