crypto_bigint

Type Alias OddUint

pub type OddUint<const LIMBS: usize> = Odd<Uint<LIMBS>>;

Odd unsigned integer.

Aliased Type

pub struct OddUint<const LIMBS: usize>(/* private fields */);

Implementations

impl<const LIMBS: usize> OddUint<LIMBS>

pub const fn from_be_hex(hex: &str) -> Self

Create a new OddUint from the provided big endian hex string.

Panics

if the hex is malformed or not zero-padded accordingly for the size.

if the value is even.

pub const fn from_le_hex(hex: &str) -> Self

Create a new Odd<Uint<LIMBS>> from the provided little endian hex string.

Panics

if the hex is malformed or not zero-padded accordingly for the size.

if the value is even.

pub const fn as_uint_ref(&self) -> &OddUintRef

Borrow this OddUint as a &OddUintRef.

pub const fn resize<const T: usize>(&self) -> Odd<Uint<T>>

Construct an Odd<Uint<T>> from the unsigned integer value, truncating the upper bits if the value is too large to be represented.

impl<const LIMBS: usize> OddUint<LIMBS>

pub const fn gcd_unsigned(&self, rhs: &Uint<LIMBS>) -> Self

Compute the greatest common divisor of self and rhs.

pub const fn gcd_unsigned_vartime(&self, rhs: &Uint<LIMBS>) -> Self

Compute the greatest common divisor of self and rhs.

Executes in variable time w.r.t. all input parameters.

pub const fn xgcd(&self, rhs: &Self) -> XgcdOutput<LIMBS, OddUint<LIMBS>>

Execute the Extended GCD algorithm.

Given (self, rhs), computes (g, x, y) s.t. self * x + rhs * y = g = gcd(self, rhs).